### 2022-09-07 #4746 @editorialbot
[PRE REVIEW]: Software Design and User Interface of ESPnet-SE++: Speech Enhancement for Robust Speech Processing
**Submitting author:** @neillu23 (<a href="http://orcid.org/0000-0001-8400-4188">Yen-Ju Lu</a>)
**Repository:** https://github.com/espnet/espnet
**Branch with paper.md** (empty if default branch):
**Version:** v.202207
