### 2022-09-07 #4746 @kthyng
Hi @neillu23! Your paper is too long. Please read the [docs](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) for details about how to write your paper.
