### 2022-09-08 ff6e77f
Add python=3.10 to CI
