+++ .github/workflows/ci.yaml
@@ -31,6 +31,11 @@ jobs:
             pytorch-version: 1.12.1
             chainer-verssion: 6.0.0
             use-conda: false
+          - os: ubuntu-20.04
+            python-version: "3.10"
+            pytorch-version: 1.12.1
+            chainer-verssion: 6.0.0
+            use-conda: false
     steps:
       - uses: actions/checkout@master
       - uses: actions/cache@v1
