### 2022-09-09 46652c0
fix parallel_wavegan installation
