+++ tools/installers/install_muskit.sh
@@ -9,13 +9,20 @@ fi
 rm -rf muskit.done
 
 # Install ParallelWaveGAN
-rm -rf ParallelWaveGAN
-git clone https://github.com/kan-bayashi/ParallelWaveGAN.git
-cd ParallelWaveGAN
-pip install -e .
-cd ../
+if ! python -c "import parallel_wavegan.version" &> /dev/null; then
+    (
+        set -euo pipefail
+        rm -rf ParallelWaveGAN
+        git clone https://github.com/kan-bayashi/ParallelWaveGAN.git
+        cd ParallelWaveGAN
+        pip install -e .
+        cd ../
+    )
+else
+    echo "parallel_wavegan is already installed"
+fi
 
-# Install 
+# Install pytsmod
 if ! python -c "import pytsmod.version" &> /dev/null; then
     (
         set -euo pipefail
@@ -25,6 +32,7 @@ else
     echo "pytsmod is already installed"
 fi
 
+# Install miditoolkit
 if ! python -c "import miditoolkit.version" &> /dev/null; then
     (
         set -euo pipefail
@@ -34,6 +42,7 @@ else
     echo "miditoolkit is already installed"
 fi
 
+# Install music21
 if ! python -c "import music21.version" &> /dev/null; then
     (
         set -euo pipefail
