### 2022-09-09 67369c6
add ground truth as well as asr transcript code
