+++ egs2/slurp/slu1/run.sh
@@ -9,12 +9,16 @@ set -o pipefail
 train_set="train"
 valid_set="devel"
 test_sets="test devel"
+local_data_opts="--gt false"
+# Make gt true to run using ground truth text as transcript
 
 slu_config=conf/tuning/train_asr_bert_conformer_deliberation.yaml
 
 ./slu.sh \
     --lang en \
     --ngpu 1 \
+    --stage 1 \
+    --stop_stage 1\
     --use_transcript true \
     --use_lm false \
     --nbpe 5000 \
@@ -23,9 +27,10 @@ slu_config=conf/tuning/train_asr_bert_conformer_deliberation.yaml
     --max_wav_duration 30 \
     --feats_normalize utterance_mvn\
     --inference_nj 12 \
-    --pretrained_model ../../slurp_new/asr1/exp/asr_train_asr_conformer_raw_en_word/valid.acc.ave_10best.pth:encoder:encoder\
-    --inference_slu_model valid.acc.ave_10best.pth\
+    --nj 12\
+    --inference_slu_model 1epoch.pth\
     --slu_config "${slu_config}" \
     --train_set "${train_set}" \
     --valid_set "${valid_set}" \
-    --test_sets "${test_sets}" "$@"
+    --test_sets "${test_sets}" \
+    --local_data_opts "${local_data_opts}" "$@"
