+++ egs2/slurp/slu1/README.md
@@ -12,6 +12,7 @@
 ## Using Conformer based deliberation encoder with ground truth transcript
 - ASR config: [conf/tuning/train_asr_bert_conformer_deliberation.yaml](conf/tuning/train_asr_bert_conformer_deliberation.yaml)
 - token_type: word
+- local_data_opts: "--gt true"
 
 |dataset|Snt|Intent Classification (%)|
 |---|---|---|
@@ -21,6 +22,7 @@
 ## Using Conformer based deliberation encoder with ASR transcript
 - ASR config: [conf/tuning/train_asr_bert_conformer_deliberation.yaml](conf/tuning/train_asr_bert_conformer_deliberation.yaml)
 - token_type: word
+- local_data_opts: "--gt false"
 
 |dataset|Snt|Intent Classification (%)|
 |---|---|---|
