### 2022-09-10 6b39896
Add MAGICDATA entry to db.sh
