### 2022-09-11 5908c72
Update egs2/TEMPLATE/svs1/pyscripts/utils/prep_segments_from_xml.py
