### 2022-09-11 051096d
Update egs2/TEMPLATE/svs1/pyscripts/utils/prep_segments_from_xml.py
