### 2022-09-11 04d316a
replace midi with xml in stage 2 & 3
