### 2022-09-11 cf710bc
Add MAGICDATA data preparation scripts
