### 2022-09-11 9b7724e
Remove unicode whitespaces in transcription
