### 2022-09-11 fb96781
Add training scripts and configs
