+++ _sources/_gen/espnet2_bin.rst.txt
@@ -1041,8 +1041,7 @@ asr_transducer_train.py
                                    [--scheduler {reducelronplateau,lambdalr,steplr,multisteplr,exponentiallr,cosineannealinglr,noamlr,warmupsteplr,warmuplr,cycliclr,onecyclelr,cosineannealingwarmrestarts,None}]
                                    [--scheduler_conf SCHEDULER_CONF]
                                    [--token_list TOKEN_LIST]
-                                   [--input_size INPUT_SIZE]
-                                   [--init {chainer,chainer_espnet1,xavier_uniform,xavier_normal,kaiming_uniform,kaiming_normal,None}]
+                                   [--input_size INPUT_SIZE] [--init INIT]
                                    [--model_conf MODEL_CONF]
                                    [--encoder_conf ENCODER_CONF]
                                    [--joint_network_conf JOINT_NETWORK_CONF]
@@ -1072,22 +1071,22 @@ asr_transducer_train.py
     optional arguments:
       --config CONFIG       Give config file in yaml format (default: None)
       --non_linguistic_symbols NON_LINGUISTIC_SYMBOLS
-                            non_linguistic_symbols file path (default: None)
+                            The 'non_linguistic_symbols' file path. (default: None)
       --cleaner {None,tacotron,jaconv,vietnamese}
-                            Apply text cleaning (default: None)
+                            Text cleaner to use. (default: None)
       --g2p {None,g2p_en,g2p_en_no_space,pyopenjtalk,pyopenjtalk_kana,pyopenjtalk_accent,pyopenjtalk_accent_with_pause,pyopenjtalk_prosody,pypinyin_g2p,pypinyin_g2p_phone,espeak_ng_arabic,espeak_ng_german,espeak_ng_french,espeak_ng_spanish,espeak_ng_russian,espeak_ng_greek,espeak_ng_finnish,espeak_ng_hungarian,espeak_ng_dutch,espeak_ng_english_us_vits,espeak_ng_hindi,g2pk,g2pk_no_space,korean_jaso,korean_jaso_no_space,g2p_is}
-                            Specify g2p method if --token_type=phn (default: None)
+                            g2p method to use if --token_type=phn. (default: None)
       --speech_volume_normalize SPEECH_VOLUME_NORMALIZE
-                            Scale the maximum amplitude to the given value. (default: None)
-      --rir_scp RIR_SCP     The file path of rir scp file. (default: None)
+                            Normalization value for maximum amplitude scaling. (default: None)
+      --rir_scp RIR_SCP     The RIR SCP file path. (default: None)
       --rir_apply_prob RIR_APPLY_PROB
-                            The probability for applying RIR convolution. (default: 1.0)
+                            The probability of the applied RIR convolution. (default: 1.0)
       --noise_scp NOISE_SCP
-                            The file path of noise scp file. (default: None)
+                            The path of noise SCP file. (default: None)
       --noise_apply_prob NOISE_APPLY_PROB
-                            The probability applying Noise adding. (default: 1.0)
+                            The probability of the applied noise addition. (default: 1.0)
       --noise_db_range NOISE_DB_RANGE
-                            The range of noise decibel level. (default: 13_15)
+                            The range of the noise decibel level. (default: 13_15)
     
     Common configuration:
       --print_config        Print the config file and exit (default: False)
@@ -1399,28 +1398,27 @@ asr_transducer_train.py
       --scheduler_conf SCHEDULER_CONF
                             The keyword arguments for lr scheduler (default: {})
     
-      Task related
+      Task related.
     
       --token_list TOKEN_LIST
-                            A text mapping int-id to token (default: None)
+                            Integer-string mapper for tokens. (default: None)
[…]
