### 2022-09-12 e443eec
Reduce batch_bins because of cuda OOM error
