+++ egs2/fsc_challenge/slu1/README.md
@@ -1,0 +1,43 @@
+<!-- Generated by ./scripts/utils/show_asr_result.sh -->
+# RESULTS
+
+## Environments
+- date: `Sun Oct  3 22:25:25 EDT 2021`
+- python version: `3.8.11 (default, Aug  3 2021, 15:09:35)  [GCC 7.5.0]`
+- espnet version: `espnet 0.10.3a3`
+- pytorch version: `pytorch 1.9.0+cu102`
+- Git hash: `a1a55e1eef2a74d2b8580d8071ce5229e7fa654c`
+  - Commit date: `Mon Nov 8 23:56:06 2021 -0500`
+
+## Using Transformer based deliberation encoder with ground truth transcript
+- ASR config: [conf/tuning/train_asr_hubert_transformer_adam_specaug_deliberation_transformer_gt.yaml](conf/tuning/train_asr_hubert_transformer_adam_specaug_deliberation_transformer_gt.yaml)
+- token_type: word
+- local_data_opts: "--gt true"
+
+|dataset|Snt|Intent Classification (%)|
+|---|---|---|
+|inference_asr_model_valid.acc.ave_5best/spk_test|3366|100.0|
+|inference_asr_model_valid.acc.ave_5best/utt_test|3970|87.6|
+|inference_asr_model_valid.acc.ave_5best/valid|2624|100.0|
+
+## Using Transformer based deliberation encoder with ASR transcript
+- ASR config: [conf/train_asr.yaml](conf/train_asr.yaml)
+- token_type: word
+- local_data_opts: "--gt false"
+
+|dataset|Snt|Intent Classification (%)|
+|---|---|---|
+|inference_asr_model_valid.acc.ave_5best/spk_test|3366|98.1|
+|inference_asr_model_valid.acc.ave_5best/utt_test|3970|82.3|
+|inference_asr_model_valid.acc.ave_5best/valid|2624|98.6|
+
+## Using Conformer based deliberation encoder with ASR transcript
+- ASR config: [conf/tuning/train_asr_hubert_transformer_adam_specaug_deliberation.yaml](conf/tuning/train_asr_hubert_transformer_adam_specaug_deliberation.yaml)
+- token_type: word
+- local_data_opts: "--gt false"
+
+|dataset|Snt|Intent Classification (%)|
+|---|---|---|
+|inference_asr_model_valid.acc.ave_5best/spk_test|3366|97.5|
+|inference_asr_model_valid.acc.ave_5best/utt_test|3970|81.9|
+|inference_asr_model_valid.acc.ave_5best/valid|2624|98.2|
