### 2022-09-12 d7f8047
add test for torch.complex in test_mixit_solver.py
