### 2022-09-13 dfef899
Reduce grad accum and clip to avoid overfitting
