### 2022-09-13 1cf18f8
fixed copy_data_dir.sh
