+++ egs2/slurp/slu1/README.md
@@ -13,6 +13,8 @@
 - ASR config: [conf/tuning/train_asr_bert_conformer_deliberation.yaml](conf/tuning/train_asr_bert_conformer_deliberation.yaml)
 - token_type: word
 - local_data_opts: "--gt true"
+- Pretrained Model
+  - Hugging Face : https://huggingface.co/espnet/slurp_slu_2pass_gt
 
 |dataset|Snt|Intent Classification (%)|
 |---|---|---|
@@ -23,6 +25,8 @@
 - ASR config: [conf/tuning/train_asr_bert_conformer_deliberation.yaml](conf/tuning/train_asr_bert_conformer_deliberation.yaml)
 - token_type: word
 - local_data_opts: "--gt false"
+- Pretrained Model
+  - Hugging Face : https://huggingface.co/espnet/slurp_slu_2pass_gt
 
 |dataset|Snt|Intent Classification (%)|
 |---|---|---|
