+++ egs2/magicdata/asr1/run.sh
@@ -35,7 +35,4 @@ speed_perturb_factors="0.9 1.0 1.1"
     --valid_set "${valid_set}"                         \
     --test_sets "${test_sets}"                         \
     --speed_perturb_factors "${speed_perturb_factors}" \
-    --asr_speech_fold_length 512 \
-    --asr_text_fold_length 150 \
-    --lm_fold_length 150 \
     --lm_train_text "data/${train_set}/text" "$@"
