+++ egs2/magicdata/asr1/conf/tuning/train_asr_conformer.yaml
@@ -37,12 +37,12 @@ model_conf:
 
 # minibatch related
 batch_type: numel
-batch_bins: 10000000
+batch_bins: 20000000
 num_workers: 4
 
 # optimization related
-accum_grad: 1
-grad_clip: 3
+accum_grad: 4
+grad_clip: 5
 max_epoch: 50
 val_scheduler_criterion:
     - valid
