+++ espnet2/enh/layers/beamformer_th.py
@@ -2,25 +2,8 @@
 from typing import List, Union
 
 import torch
-from packaging.version import parse as V
-from torch_complex import functional as FC
-from torch_complex.tensor import ComplexTensor
-
-from espnet2.enh.layers.complex_utils import (
-    cat,
-    complex_norm,
-    einsum,
-    inverse,
-    is_complex,
-    is_torch_complex_tensor,
-    matmul,
-    reverse,
-    solve,
-    to_double,
-)
-
-is_torch_1_9_plus = V(torch.__version__) >= V("1.9.0")
-EPS = torch.finfo(torch.double).eps
+import torch_complex.functional as FC
+import torchaudio
 
 
 def prepare_beamformer_stats(
@@ -36,7 +19,7 @@ def prepare_beamformer_stats(
     """Prepare necessary statistics for constructing the specified beamformer.
 
     Args:
-        signal (torch.complex64/ComplexTensor): (..., F, C, T)
+        signal (torch.complex64): (..., F, C, T)
         masks_speech (List[torch.Tensor]): (..., F, C, T) masks for all speech sources
         mask_noise (torch.Tensor): (..., F, C, T) noise mask
         powers (List[torch.Tensor]): powers for all speech sources (..., F, T)
@@ -60,9 +43,9 @@ def prepare_beamformer_stats(
     assert beamformer_type in BEAMFORMER_TYPES, "%s is not supported yet"
 
     if isinstance(masks_speech, (list, tuple)):
-        masks_speech = [to_double(m) for m in masks_speech]
+        masks_speech = [m.cdouble() for m in masks_speech]
     else:
-        masks_speech = [to_double(masks_speech)]
+        masks_speech = [masks_speech.cdouble()]
     num_spk = len(masks_speech)
 
     if (
@@ -74,12 +57,16 @@ def prepare_beamformer_stats(
         if powers is None:
             power_input = signal.real**2 + signal.imag**2
             # Averaging along the channel axis: (..., C, T) -> (..., T)
-            powers = [(power_input * m).mean(dim=-2) for m in masks_speech]
+            powers = [(power_input * m.abs()).mean(dim=-2) for m in masks_speech]
         else:
             assert len(powers) == num_spk, (len(powers), num_spk)
         inverse_powers = [1 / torch.clamp(p, min=eps) for p in powers]
 
-    psd_speeches = [get_power_spectral_density_matrix(signal, m) for m in masks_speech]
+    psd_transform = torchaudio.transforms.PSD(multi_mask=True, normalize=True)
+    psd_speeches = [
+        psd_transform(signal.transpose(-2, -3), m.transpose(-2, -3))
+        for m in masks_speech
+    ]
     if (
         beamformer_type == "mvdr_souden"
         or beamformer_type == "sdw_mwf"
@@ -89,7 +76,9 @@ def prepare_beamformer_stats(
     ):
         # MVDR or other RTF-based formulas
         if mask_noise is not None:
-            psd_bg = get_power_spectral_density_matrix(signal, to_double(mask_noise))
+            psd_bg = psd_transform(
+                signal.transpose(-2, -3), mask_noise.cdouble().transpose(-2, -3)
+            )
         if num_spk == 1:
             assert mask_noise is not None
             psd_noise = psd_bg
@@ -121,10 +110,10 @@ def prepare_beamformer_stats(
         psd_n = psd_noise
         psd_noise = [sum(psd_noise_i) for psd_noise_i in psd_noise]
     elif beamformer_type in ("mpdr", "mpdr_souden", "lcmp", "mwf"):
[…]
