+++ _sources/_gen/espnet2_bin.rst.txt
@@ -2776,7 +2776,6 @@ enh_train.py
                         [--scheduler_conf SCHEDULER_CONF]
                         [--init {chainer,xavier_uniform,xavier_normal,kaiming_uniform,kaiming_normal,None}]
                         [--model_conf MODEL_CONF] [--criterions CRITERIONS]
-                        [--use_preprocessor USE_PREPROCESSOR]
                         [--speech_volume_normalize SPEECH_VOLUME_NORMALIZE]
                         [--rir_scp RIR_SCP] [--rir_apply_prob RIR_APPLY_PROB]
                         [--noise_scp NOISE_SCP]
@@ -2795,6 +2794,8 @@ enh_train.py
                         [--decoder {stft,conv,same}] [--decoder_conf DECODER_CONF]
                         [--mask_module {multi_mask}]
                         [--mask_module_conf MASK_MODULE_CONF]
+                        [--preprocessor {dynamic_mixing,enh,None}]
+                        [--preprocessor_conf PREPROCESSOR_CONF]
     
     base parser
     
@@ -3122,8 +3123,6 @@ enh_train.py
     
       Preprocess related
     
-      --use_preprocessor USE_PREPROCESSOR
-                            Whether to apply preprocessing to data or not (default: False)
       --speech_volume_normalize SPEECH_VOLUME_NORMALIZE
                             Scale the maximum amplitude to the given value or range. e.g. --speech_volume_normalize 1.0 scales it to 1.0.
                             --speech_volume_normalize 0.5_1.0 scales it to a random number in the range [0.5, 1.0) (default: None)
@@ -3168,6 +3167,10 @@ enh_train.py
                             The mask_module type (default: multi_mask)
       --mask_module_conf MASK_MODULE_CONF
                             The keyword arguments for mask_module (default: {})
+      --preprocessor {dynamic_mixing,enh,None}
+                            The preprocessor type (default: None)
+      --preprocessor_conf PREPROCESSOR_CONF
+                            The keyword arguments for preprocessor (default: {})
 .. _gan_tts_train.py:
 
 gan_tts_train.py
