### 2022-09-14 f638685
Store test results in README.md
