+++ egs2/magicdata/asr1/README.md
@@ -1,0 +1,17 @@
+# RESULTS
+## Environments
+- date: `Wed Sep 14 16:38:44 EDT 2022`
+- python version: `3.9.12 (main, Apr  5 2022, 06:56:58)  [GCC 7.5.0]`
+- espnet version: `espnet 202207`
+- pytorch version: `pytorch 1.8.1+cu102`
+- Git hash: `de23a67dccfd2af3fc95fd50020a03130f6a0dc1`
+  - Commit date: `Tue Sep 13 18:15:33 2022 -0400`
+
+## asr_train_asr_conformer_raw_zh_char_sp
+
+### CER
+
+|dataset|Snt|Wrd|Corr|Sub|Del|Ins|Err|S.Err|
+|---|---|---|---|---|---|---|---|---|
+|decode_asr_rnn_lm_lm_train_lm_transformer_zh_char_valid.loss.ave_asr_model_valid.acc.best/test|24279|243325|94.5|3.2|2.3|0.1|5.6|22.0|
+
