### 2022-09-14 2713f7c
Simplifiy training configurations
