+++ egs2/magicdata/asr1/run.sh
@@ -9,7 +9,7 @@ train_set=train_noeng
 valid_set=dev
 test_sets="test"
 
-asr_config=conf/train_asr_conformer.yaml 
+asr_config=conf/train_asr.yaml
 inference_config=conf/decode_asr_rnn.yaml
 
 lm_config=conf/train_lm_transformer.yaml
