+++ egs2/magicdata/asr1/conf/tuning/transducer/decode.yaml
@@ -1,5 +1,0 @@
-beam_size: 10
-lm_weight: 0.2
-transducer_conf:
-    search_type: default
-    score_norm: False
