+++ egs2/magicdata/asr1/conf/tuning/transducer/train_conformer-rnn_transducer.yaml
@@ -1,78 +1,0 @@
-# Trained with Tesla V100 (32GB) x 8 GPUs. It takes about 4 days.
-
-# network architecture
-# encoder related
-encoder: conformer
-encoder_conf:
-    output_size: 512    # dimension of attention
-    attention_heads: 8
-    linear_units: 2048  # the number of units of position-wise feed forward
-    num_blocks: 12      # the number of encoder blocks
-    dropout_rate: 0.1
-    positional_dropout_rate: 0.1
-    attention_dropout_rate: 0.0
-    input_layer: conv2d # encoder architecture type
-    normalize_before: true
-    pos_enc_layer_type: rel_pos
-    selfattention_layer_type: rel_selfattn
-    activation_type: swish
-    macaron_style: true
-    use_cnn_module: true
-    cnn_module_kernel: 15
-
-# decoder related
-decoder: transducer
-decoder_conf:
-    rnn_type: lstm
-    num_layers: 1
-    hidden_size: 512
-    dropout: 0.1
-    dropout_embed: 0.2
-
-joint_net_conf:
-    joint_space_size: 512
-
-# Auxiliary CTC
-model_conf:
-    ctc_weight: 0.0
-
-# minibatch related
-batch_type: numel
-batch_bins: 20000000
-num_workers: 4
-
-# optimization related
-accum_grad: 4
-grad_clip: 5
-max_epoch: 50
-val_scheduler_criterion:
-    - valid
-    - loss
-best_model_criterion:
--   - valid
-    - cer_transducer
-    - min
-keep_nbest_models: 10
-
-optim: adam
-optim_conf:
-   lr: 0.001
-scheduler: warmuplr
-scheduler_conf:
-   warmup_steps: 30000
-
-specaug: specaug
-specaug_conf:
-    apply_time_warp: true
-    time_warp_window: 5
-    time_warp_mode: bicubic
-    apply_freq_mask: true
-    freq_mask_width_range:
-    - 0
-    - 30
-    num_freq_mask: 2
-    apply_time_mask: true
-    time_mask_width_range:
-    - 0
-    - 40
-    num_time_mask: 2
