### 2022-09-15 aaad835
Lower learning rate and re-train
