+++ espnet2/asvspoof/decoder/linear_decoder.py
@@ -1,0 +1,28 @@
+import torch
+
+from typing import Optional
+from espnet2.asvspoof.decoder.abs_decoder import AbsDecoder
+
+
+class LinearDecoder(AbsDecoder):
+    """Linear decoder for speaker diarization"""
+
+    def __init__(
+        self,
+        encoder_output_size: int,
+    ):
+        super().__init__()
+        # TODO1 (checkpoint3): initialize a linear projection layer
+
+    def forward(self, input: torch.Tensor, ilens: Optional[torch.Tensor]):
+        """Forward.
+        Args:
+            input (torch.Tensor): hidden_space [Batch, T, F]
+            ilens (torch.Tensor): input lengths [Batch]
+        """
+        # TODO2 (checkpoint3): compute mean over time-domain (dimension 1)
+        
+        # TODO3 (checkpoint3): apply the projection layer
+
+        # TODO4 (checkpoint3): change the return value accordingly
+        return None
