+++ espnet2/bin/asvspoof_inference.py
@@ -1,0 +1,252 @@
+#!/usr/bin/env python3
+import argparse
+import logging
+import sys
+from distutils.version import LooseVersion
+from pathlib import Path
+from typing import Any, List, Optional, Sequence, Tuple, Union
+
+import numpy as np
+import torch
+import torch.quantization
+from typeguard import check_argument_types, check_return_type
+
+from espnet2.fileio.datadir_writer import DatadirWriter
+from espnet2.tasks.asvspoof import ASVSpoofTask
+from espnet2.torch_utils.device_funcs import to_device
+from espnet2.torch_utils.set_all_random_seed import set_all_random_seed
+from espnet2.utils import config_argparse
+from espnet2.utils.types import str2bool, str2triple_str, str_or_none
+from espnet.nets.pytorch_backend.transformer.subsampling import TooShortUttError
+from espnet.utils.cli_utils import get_commandline_args
+
+
+class SpeechAntiSpoof:
+    """SpeechAntiSpoof class
+    Examples:
+        >>> import soundfile
+        >>> speech_anti_spoof = SpeechAntiSpoof("asvspoof_config.yml", "asvspoof.pth")
+        >>> audio, rate = soundfile.read("speech.wav")
+        >>> speech_anti_spoof(audio)
+        prediction_result (int)
+    """
+
+    def __init__(
+        self,
+        asvspoof_train_config: Union[Path, str] = None,
+        asvspoof_model_file: Union[Path, str] = None,
+        device: str = "cpu",
+        batch_size: int = 1,
+        dtype: str = "float32",
+    ):
+        assert check_argument_types()
+
+        asvspoof_model, asvspoof_train_args = ASVSpoofTask.build_model_from_file(
+            asvspoof_train_config, asvspoof_model_file, device
+        )
+        asvspoof_model.to(dtype=getattr(torch, dtype)).eval()
+
+        self.asvspoof_model = asvspoof_model
+        self.asvspoof_train_args = asvspoof_train_args
+        self.device = device
+        self.dtype = dtype
+
+    @torch.no_grad()
+    def __call__(
+        self, speech: Union[torch.Tensor, np.ndarray]
+    ) -> Tuple[float]:
+        """Inference
+        Args:
+            data: Input speech data
+        Returns:
+            [prediction, scores]
+        """
+        assert check_argument_types()
+
+        # Input as audio signal
+        if isinstance(speech, np.ndarray):
+            speech = torch.tensor(speech)
+
+        # data: (Nsamples,) -> (1, Nsamples)
+        speech = speech.unsqueeze(0).to(getattr(torch, self.dtype))
+        # lengths: (1,)
+        lengths = speech.new_full([1], dtype=torch.long, fill_value=speech.size(1))
+        batch = {"speech": speech, "speech_lengths": lengths}
+        logging.info("speech length: " + str(speech.size(1)))
+
+        # To device
+        batch = to_device(batch, device=self.device)
+
+        # TODO1 (checkpoint 4): Forward Encoder
+
+        # TODO2 (checkpoint 4): Pass the encoder result to decoder
+
+        # TODO3 (checkpoint 4): return the prediction score
+        return None
+
+
+def inference(
+    output_dir: str,
+    batch_size: int,
+    dtype: str,
+    ngpu: int,
+    seed: int,
+    num_workers: int,
+    log_level: Union[int, str],
+    data_path_and_name_and_type: Sequence[Tuple[str, str, str]],
+    key_file: Optional[str],
+    asvspoof_train_config: Optional[str],
+    asvspoof_model_file: Optional[str],
+    allow_variable_data_keys: bool,
+):
+    assert check_argument_types()
+    if batch_size > 1:
+        raise NotImplementedError("batch decoding is not implemented")
[…]
