+++ espnet2/asvspoof/decoder/abs_decoder.py
@@ -1,0 +1,16 @@
+from abc import ABC
+from abc import abstractmethod
+from typing import Tuple
+
+import torch
+
+
+class AbsDecoder(torch.nn.Module, ABC):
+    @abstractmethod
+    def forward(
+        self,
+        input: torch.Tensor,
+        ilens: torch.Tensor,
+    ) -> Tuple[torch.Tensor, torch.Tensor]:
+        raise NotImplementedError
+
