+++ egs2/TEMPLATE/svs1/scripts/audio/format_xml_scp.sh
@@ -1,5 +1,4 @@
 #!/usr/bin/env bash
-
 set -euo pipefail
 SECONDS=0
 log() {
@@ -7,38 +6,33 @@ log() {
     echo -e "$(date '+%Y-%m-%dT%H:%M:%S') (${fname}:${BASH_LINENO[0]}:${FUNCNAME[1]}) $*"
 }
 help_message=$(cat << EOF
-Usage: $0 <in-wav.scp> <out-datadir> [<logdir> [<outdir>]]
+Usage: $0 <in-xml.scp> <out-datadir> [<logdir> [<outdir>]]
 e.g.
-$0 data/test/wav.scp data/test_format/
+$0 data/test/xml.scp data/test_format/
 
-Format 'wav.scp': In short words,
+Format 'xml.scp': In short words,
 changing "kaldi-datadir" to "modified-kaldi-datadir"
 
-The 'wav.scp' format in kaldi is very flexible,
-e.g. It can use unix-pipe as describing that wav file,
+The 'xml.scp' format in kaldi is very flexible,
+e.g. It can use unix-pipe as describing that xml file,
 but it sometime looks confusing and make scripts more complex.
-This tools creates actual wav files from 'wav.scp'
-and also segments wav files using 'segments'.
+This tools creates actual xml files from 'xml.scp'.
 
 Options
-  --fs <fs>
-  --segments <segments>
   --nj <nj>
+  --segments <segments>
   --cmd <cmd>
 EOF
 )
 
-out_wavfilename=wav.scp
 out_xmlfilename=xml.scp
 cmd=utils/run.pl
 nj=30
-fs=none
 segments=
 
 ref_channels=
 utt2ref_channels=
 
-audio_format=wav
 write_utt2num_samples=true
 
 log "$0 $*"
@@ -53,16 +47,7 @@ fi
 . ./path.sh  # Setup the environment
 
 scp_dir=$1
-if [ ! -f "${scp_dir}/wav.scp" ]; then
-    log "${help_message}"
-    echo "$0: Error: No such file: ${scp_dir}/wav.scp"
-    exit 1
-fi
-if [ ! -d "xml_dump" ]; then
-    log "${help_message}"
-    echo "$0: Error: No such folder: xml_dump"
-    exit 1
-fi
+# NOTE(Yuning): If orgin xmls are segmented here, scp_filw need to be fixed like wav.scp.
 dir=$2
 
 
@@ -82,7 +67,6 @@ fi
 
 mkdir -p ${logdir}
 
-rm -f "${dir}/${out_wavfilename}"
 rm -f "${dir}/${out_xmlfilename}"
 
 opts=
@@ -93,34 +77,16 @@ elif [ -n "${ref_channels}" ]; then
 fi
 
 
-
-
 if [ -n "${segments}" ]; then
     log "[info]: using ${segments}"
     nutt=$(<${segments} wc -l)
     nj=$((nj<nutt?nj:nutt))
-
-    split_segments=""
-    for n in $(seq ${nj}); do
-        split_segments="${split_segments} ${logdir}/segments.${n}"
-    done
-
-    utils/split_scp.pl "${segments}" ${split_segments}
-
-    ${cmd} "JOB=1:${nj}" "${logdir}/format_wav_scp.JOB.log" \
-        pyscripts/audio/format_wav_scp.py \
-            ${opts} \
-            --fs ${fs} \
-            --audio-format "${audio_format}" \
-            "--segment=${logdir}/segments.JOB" \
-            "${scp_dir}/wav.scp" "${outdir}/format_wav.JOB"
     
     ${cmd} "JOB=1:${nj}" "${logdir}/format_xml_scp.JOB.log" \
[…]
