### 2022-09-18 3033fdf
Update egs2/TEMPLATE/svs1/svs.sh
