+++ egs2/TEMPLATE/svs1/svs.sh
@@ -51,7 +51,7 @@ min_wav_duration=0.1 # Minimum duration in second.
 max_wav_duration=20  # Maximum duration in second.
 use_sid=false        # Whether to use speaker id as the inputs (Need utt2spk in data directory).
 use_lid=false        # Whether to use language id as the inputs (Need utt2lang in data directory).
-use_xvector=false    # Whether to use x-vector (Require Kaldi).
+use_xvector=false    # Whether to use x-vector 
 feats_extract=fbank        # On-the-fly feature extractor.
 feats_normalize=global_mvn # On-the-fly feature normalizer.
 # Only used for feats_type != raw
