+++ egs2/TEMPLATE/asr1/scripts/utils/show_asvspoof_result.sh
@@ -1,0 +1,65 @@
+#!/usr/bin/env bash
+mindepth=0
+maxdepth=1
+
+. utils/parse_options.sh
+
+if [ $# -gt 1 ]; then
+    echo "Usage: $0 --mindepth 0 --maxdepth 1 [exp]" 1>&2
+    echo ""
+    echo "Show the system environments and the evaluation results in Markdown format."
+    echo 'The default of <exp> is "exp/".'
+    exit 1
+fi
+
+[ -f ./path.sh ] && . ./path.sh
+set -euo pipefail
+if [ $# -eq 1 ]; then
+    exp=$1
+else
+    exp=exp
+fi
+
+
+cat << EOF
+<!-- Generated by $0 -->
+# RESULTS
+## Environments
+- date: \`$(LC_ALL=C date)\`
+EOF
+
+python3 << EOF
+import sys, espnet, torch
+pyversion = sys.version.replace('\n', ' ')
+
+print(f"""- python version: \`{pyversion}\`
+- espnet version: \`espnet {espnet.__version__}\`
+- pytorch version: \`pytorch {torch.__version__}\`""")
+EOF
+
+cat << EOF
+- Git hash: \`$(git rev-parse HEAD)\`
+  - Commit date: \`$(git log -1 --format='%cd')\`
+
+EOF
+
+while IFS= read -r expdir; do
+    
+    if ls "${expdir}"/*/*/score/eer &> /dev/null; then
+        echo "## $(basename ${expdir})"
+                	cat << EOF
+### ${type^^}
+
+|dataset|EER|
+|---|---|
+EOF
+        if ls "${expdir}"/*/*/score/eer &> /dev/null; then
+                    cat "${expdir}"/*/*/score/eer \
+                        | sed -e "s#${expdir}/\([^/]*/[^/]*\)/score/eer:#|\1#g" \
+                        | sed -e 's#^##g' | tr '|' ' ' | tr -s ' ' '|'
+                    echo
+        fi
+    fi
+
+
+done < <(find ${exp} -mindepth ${mindepth} -maxdepth ${maxdepth} -type d)
