+++ espnet2/bin/asvspoof_inference.py
@@ -79,7 +79,10 @@ class SpeechAntiSpoof:
 
         # TODO1 (checkpoint 4): Forward Encoder
 
-        # TODO2 (checkpoint 4): Pass the encoder result to decoder
+        if "oc_softmax_loss" in self.asvspoof_model.loss:
+            pass # TODO1 (exercise2): use loss score function to estimate score
+        else:
+            pass # TODO2 (checkpoint 4): Pass the encoder result to decoder
 
         # TODO3 (checkpoint 4): return the prediction score
         return None
