+++ egs2/qasr_tts/tts1/run.sh
@@ -24,7 +24,7 @@ train_config=conf/train.yaml
 inference_config=conf/decode.yaml
 
 # g2p=g2p_en # Include word separator
-g2p=g2p_en_no_space # Include no word separator
+g2p=none # Include no word separator
 
 bash ./tts.sh \
     --stage 1 \
