+++ egs2/qasr_tts/tts1/local/data_download.sh
@@ -17,9 +17,7 @@ cwd=$(pwd)
 
 if [ ! -d "${download_dir}/qasr_tts-1.0" ] && [ -f "qasr_tts-1.0.zip" ]; then
     mkdir -p "${download_dir}"
-    cd "${download_dir}"
-    unzip "${cwd}/qasr_tts-1.0.zip"
-    cd "${cwd}"
+    unzip "${cwd}/qasr_tts-1.0.zip" -d "${download_dir}"
     echo "successfully prepared data."
     
 else
