+++ egs2/qasr_tts/tts1/conf/tuning/decode_vits.yaml
@@ -1,10 +1,0 @@
-# This configuration is the decoding setting for VITS.
-
-##########################################################
-#                    DECODING SETTING                    #
-##########################################################
-noise_scale: 0.667         # noise scale parameter for the flow in VITS.
-noise_scale_dur: 0.8       # noise scale parameter for the stochastic duration predictor in VITS.
-speed_control_alpha: 1     # alpha to control the speed of generated speech.
-                           # 1 < alpha makes slower and 1 > alpha makes faster.
-use_teacher_forcing: false # whether to use teacher forcing.
