+++ egs2/qasr_tts/tts1/conf/tuning/train_vits.yaml
@@ -1,185 +1,0 @@
-# This configuration is for ESPnet2 to train VITS, which
-# is truely end-to-end text-to-waveform model. To run
-# this config, you need to specify "--tts_task gan_tts"
-# option for tts.sh at least and use 22050 hz audio as
-# the training data (mainly tested on LJspeech).
-# This configuration tested on 4 GPUs (V100) with 32GB GPU
-# memory. It takes around 2 weeks to finish the training
-# but 100k iters model should generate reasonable results.
-
-##########################################################
-#                  TTS MODEL SETTING                     #
-##########################################################
-tts: vits
-tts_conf:
-    # generator related
-    generator_type: vits_generator
-    generator_params:
-        hidden_channels: 192
-        spks: -1
-        global_channels: -1
-        segment_size: 32
-        text_encoder_attention_heads: 2
-        text_encoder_ffn_expand: 4
-        text_encoder_blocks: 6
-        text_encoder_positionwise_layer_type: "conv1d"
-        text_encoder_positionwise_conv_kernel_size: 3
-        text_encoder_positional_encoding_layer_type: "rel_pos"
-        text_encoder_self_attention_layer_type: "rel_selfattn"
-        text_encoder_activation_type: "swish"
-        text_encoder_normalize_before: true
-        text_encoder_dropout_rate: 0.1
-        text_encoder_positional_dropout_rate: 0.0
-        text_encoder_attention_dropout_rate: 0.1
-        use_macaron_style_in_text_encoder: true
-        # NOTE(kan-bayashi): Conformer conv requires BatchNorm1d which causes
-        #   errors when multiple GPUs in pytorch 1.7.1. Therefore, we disable
-        #   it as a default. We need to consider the alternative normalization
-        #   or different version pytorch may solve this issue.
-        use_conformer_conv_in_text_encoder: false
-        text_encoder_conformer_kernel_size: -1
-        decoder_kernel_size: 7
-        decoder_channels: 512
-        decoder_upsample_scales: [8, 8, 2, 2]
-        decoder_upsample_kernel_sizes: [16, 16, 4, 4]
-        decoder_resblock_kernel_sizes: [3, 7, 11]
-        decoder_resblock_dilations: [[1, 3, 5], [1, 3, 5], [1, 3, 5]]
-        use_weight_norm_in_decoder: true
-        posterior_encoder_kernel_size: 5
-        posterior_encoder_layers: 16
-        posterior_encoder_stacks: 1
-        posterior_encoder_base_dilation: 1
-        posterior_encoder_dropout_rate: 0.0
-        use_weight_norm_in_posterior_encoder: true
-        flow_flows: 4
-        flow_kernel_size: 5
-        flow_base_dilation: 1
-        flow_layers: 4
-        flow_dropout_rate: 0.0
-        use_weight_norm_in_flow: true
-        use_only_mean_in_flow: true
-        stochastic_duration_predictor_kernel_size: 3
-        stochastic_duration_predictor_dropout_rate: 0.5
-        stochastic_duration_predictor_flows: 4
-        stochastic_duration_predictor_dds_conv_layers: 3
-    # discriminator related
-    discriminator_type: hifigan_multi_scale_multi_period_discriminator
-    discriminator_params:
-        scales: 1
-        scale_downsample_pooling: "AvgPool1d"
-        scale_downsample_pooling_params:
-            kernel_size: 4
-            stride: 2
-            padding: 2
-        scale_discriminator_params:
-            in_channels: 1
-            out_channels: 1
-            kernel_sizes: [15, 41, 5, 3]
-            channels: 128
-            max_downsample_channels: 1024
-            max_groups: 16
-            bias: True
[…]
