+++ egs2/qasr_tts/tts1/conf/tuning/train_conformer_fastspeech2.yaml
@@ -84,7 +84,7 @@ scheduler_conf:        # keyword arguments for selected scheduler
 ##########################################################
 #                OTHER TRAINING SETTING                  #
 ##########################################################
-num_iters_per_epoch: 1000 # number of iterations per epoch
+num_iters_per_epoch: 200 #1000 # number of iterations per epoch
 max_epoch: 1000           # number of epochs
 grad_clip: 1.0            # gradient clipping norm
 grad_noise: false         # whether to use gradient noise injection
