### 2022-09-19 da14886
Lower learning rate to avoid training loss hill climbing after 7-epoch
