### 2022-09-19 2a188ec
Download and untar magicdata automatically
