+++ egs2/magicdata/asr1/local/data.sh
@@ -10,55 +10,109 @@ set -o pipefail
 . ./db.sh
 
 log() {
-    local fname=${BASH_SOURCE[1]##*/}
-    echo -e "$(date '+%Y-%m-%dT%H:%M:%S') (${fname}:${BASH_LINENO[0]}:${FUNCNAME[1]}) $*"
+  local fname=${BASH_SOURCE[1]##*/}
+  echo -e "$(date '+%Y-%m-%dT%H:%M:%S') (${fname}:${BASH_LINENO[0]}:${FUNCNAME[1]}) $*"
 }
+help_message=$(
+  cat <<EOF
+Usage: $0
+
+Options:
+    --remove_archive (bool): true or false
+      With remove_archive=True, the archives will be removed after being successfully downloaded and un-tarred.
+EOF
+)
 SECONDS=0
 
 stage=1
 stop_stage=100
 
+data_url=www.openslr.org/resources/68
+remove_archive=false
+download_opt=
+
 log "$0 $*"
 . utils/parse_options.sh
 
-if [ $# -ne 0 ]; then
-    log "Error: No positional arguments are required."
-    exit 2
+if [ $# -gt 1 ]; then
+  log "${help_message}"
+  exit 2
 fi
 
-# TODO: auto download and:
-#   1. mkdir wav
-#   2. put train, test, dev subset under wav/
-#   3. copy train.scp, test,scp, dev.scp into wav/train/wav.scp, wav/test/wav.scp, wav/dev/wav.scp
+if "$remove_archive"; then
+  download_opt="--remove-archive"
+fi
+
 if [ -z "${MAGICDATA}" ]; then
   log "Error: \$MAGICDATA is not set in db.sh."
-  log "You may request the MAGICDATA dataset from http://www.openslr.org/68/"
   exit 2
 fi
 
 if [ ${stage} -le 1 ] && [ ${stop_stage} -ge 1 ]; then
-    ### Task dependent. You have to make data the following preparation part by yourself.
-    ### But you can utilize Kaldi recipes in most cases
-    echo "stage 0: Data preparation"
+  log "Download data to ${MAGICDATA}"
+  if [ ! -d "${MAGICDATA}" ]; then
+    mkdir -p "${MAGICDATA}"
+  fi
+  # To absolute path
+  MAGICDATA=$(
+    cd ${MAGICDATA}
+    pwd
+  )
 
-    local/prepare_data.sh ${MAGICDATA}/wav/train data/local/train data/train || exit 1;
-    local/prepare_data.sh ${MAGICDATA}/wav/test data/local/test data/test || exit 1;
-    local/prepare_data.sh ${MAGICDATA}/wav/dev data/local/dev data/dev || exit 1;
+  # download, untar, and:
+  #   1. mkdir wav
+  #   2. put train, test, dev under wav/
+  #   3. copy train.scp, test.scp, dev.scp into wav/train/wav.scp, wav/test/wav.scp, wav/dev/wav.scp
 
-    # Normalize text to capital letters
-    for x in train dev test; do
-        mv data/${x}/text data/${x}/text.org
-        paste <(cut -f 1 data/${x}/text.org) <(cut -f 2 data/${x}/text.org | tr '[:lower:]' '[:upper:]') \
-            > data/${x}/text
-        rm data/${x}/text.org
-    done
+  local/download_and_untar.sh ${download_opt} "${MAGICDATA}" "${data_url}" train_set
+  local/download_and_untar.sh ${download_opt} "${MAGICDATA}" "${data_url}" dev_set
+  local/download_and_untar.sh ${download_opt} "${MAGICDATA}" "${data_url}" test_set
[…]
