+++ egs2/TEMPLATE/asr1/db.sh
@@ -51,7 +51,7 @@ SLURP=
 SLURP_S= # Output file path
 LIBRITRANS_S= # Output file path
 VOXCELEB=
-MAGICDATA=
+MAGICDATA=downloads
 MEDIASPEECH=downloads
 MINI_LIBRISPEECH=downloads
 MISP2021=
