### 2022-09-20 6a61ef8
Update espnet2/svs/espnet_model.py
Co-authored-by: Jiatong <728307998@qq.com>
