### 2022-09-20 7ee2f84
Update egs2/qasr_tts/tts1/local/data.sh
Co-authored-by: Jiatong <728307998@qq.com>
