### 2022-09-20 7ca298f
Update egs2/qasr_tts/tts1/run.sh
Co-authored-by: Jiatong <728307998@qq.com>
