+++ egs2/qasr_tts/tts1/run.sh
@@ -23,7 +23,6 @@ test_sets="dev eval1"
 train_config=conf/train.yaml
 inference_config=conf/decode.yaml
 
-# g2p=g2p_en # Include word separator
 g2p=none # Include no word separator
 
 bash ./tts.sh \
