+++ espnet2/svs/naive_rnn/naive_rnn.py
@@ -134,8 +134,6 @@ class NaiveRNN(AbsSVS):
         use_masking: bool = False,
         use_weighted_masking: bool = False,
         loss_type: str = "L1",
-        use_mixup_training: bool = False,
-        loss_mixup_wight: float = 0.1,
     ):
         """Initialize NaiveRNN module.
         Args: TODO
@@ -153,10 +151,6 @@ class NaiveRNN(AbsSVS):
         self.loss_type = loss_type
 
         self.midi_embed_integration_type = midi_embed_integration_type
-
-        # mixup - augmentation
-        self.use_mixup_training = use_mixup_training
-        self.loss_mixup_wight = loss_mixup_wight
 
         # use idx 0 as padding idx
         self.padding_idx = 0
@@ -351,87 +345,6 @@ class NaiveRNN(AbsSVS):
         label_emb = self.encoder_input_layer(label)  # FIX ME: label Float to Int
         midi_emb = self.midi_encoder_input_layer(midi)
 
-        if self.use_mixup_training and flag_IsValid == False:
-            # batch_size_mixup = batch_size // 2                  # !!! NOTE: origin - 2
-            # lst = [i for i in range(batch_size_mixup * 2)]
-            # random.shuffle(lst)
-
-            batch_size_mixup = 2
-            lst = random.sample(
-                [i for i in range(batch_size)], batch_size_mixup * 2
-            )  # mix-up per 2 samples
-
-            # mix-up augmentation
-            feats_mixup = torch.zeros(
-                (batch_size_mixup, feats.shape[1], feats.shape[2]),
-                dtype=feats.dtype,
-                layout=feats.layout,
-                device=feats.device,
-            )
-            feats_lengths_mixup = torch.zeros(
-                batch_size_mixup,
-                dtype=feats_lengths.dtype,
-                layout=feats_lengths.layout,
-                device=feats_lengths.device,
-            )
-
-            midi_embed_mixup = torch.zeros(
-                (batch_size_mixup, midi_emb.shape[1], midi_emb.shape[2]),
-                dtype=midi_emb.dtype,
-                layout=midi_emb.layout,
-                device=midi_emb.device,
-            )
-            midi_lengths_mixup = torch.zeros(
-                batch_size_mixup,
-                dtype=midi_lengths.dtype,
-                layout=midi_lengths.layout,
-                device=midi_lengths.device,
-            )
-
-            label_embed_mixup = torch.zeros(
-                (batch_size_mixup, label_emb.shape[1], label_emb.shape[2]),
-                dtype=label_emb.dtype,
-                layout=label_emb.layout,
-                device=label_emb.device,
-            )
-            label_lengths_mixup = torch.zeros(
-                batch_size_mixup,
-                dtype=label_lengths.dtype,
-                layout=label_lengths.layout,
-                device=label_lengths.device,
-            )
-
-            for i in range(batch_size_mixup):
-                index1 = lst[2 * i]
-                index2 = lst[2 * i + 1]
-
-                w1 = Beta_distribution.sample().to(
-                    feats.device
-                )  # !!! NOTE:  random.random()
-                w2 = 1 - w1
-
-                feats_mixup[i] = w1 * feats[index1] + w2 * feats[index2]
-                feats_lengths_mixup[i] = max(
-                    feats_lengths[index1], feats_lengths[index2]
-                )
-
-                midi_embed_mixup[i] = w1 * midi_emb[index1] + w2 * midi_emb[index2]
-                midi_lengths_mixup[i] = max(midi_lengths[index1], midi_lengths[index2])
-
[…]
