+++ egs2/dns_icassp22/enh1/local/download_dns4_dataset.sh
@@ -1,0 +1,218 @@
+#!/usr/bin/bash
+
+# Adapted from DNS-Challenge official repo file: https://github.com/microsoft/DNS-Challenge/blob/master/download-dns-challenge-4.sh
+# Original License: CC-BY 4.0 International: https://github.com/microsoft/DNS-Challenge/blob/master/LICENSE
+# Retrieved on Sep. 7th, 2022, by Shih-Lun Wu (summer7sean@gmail.com) 
+
+# ***** Datasets for ICASSP 2022 DNS Challenge 4 - Main (Real-Time) Track *****
+
+# NOTE: Before downloading, make sure you have enough space
+# on your local storage!
+
+# In all, you will need about 1TB to store the UNPACKED data.
+# Archived, the same data takes about 550GB total.
+
+# Please comment out the files you don't need before launching
+# the script.
+
+# NOTE: By default, the script *DOES NOT* DOWNLOAD ANY FILES!
+# Please scroll down and edit this script to pick the
+# downloading method that works best for you.
+
+# -------------------------------------------------------------
+# The directory structure of the unpacked data is:
+
+# datasets_fullband 892G
+# +-- dev_testset 1.7G
+# +-- impulse_responses 5.9G
+# +-- noise_fullband 58G
+# \-- clean_fullband 827G
+#     +-- emotional_speech 2.4G
+#     +-- french_speech 62G
+#     +-- german_speech 319G
+#     +-- italian_speech 42G
+#     +-- read_speech 299G
+#     +-- russian_speech 12G
+#     +-- spanish_speech 65G
+#     +-- vctk_wav48_silence_trimmed 27G
+#     \-- VocalSet_48kHz_mono 974M
+
+# NOTE: for storage space considerations, by default, only English read speech data will be downloaded (Shih-Lun)
+BLOB_NAMES=(
+    # clean_fullband/datasets_fullband.clean_fullband.VocalSet_48kHz_mono_000_NA_NA.tar.bz2
+
+    # clean_fullband/datasets_fullband.clean_fullband.emotional_speech_000_NA_NA.tar.bz2
+
+    # clean_fullband/datasets_fullband.clean_fullband.french_speech_000_NA_NA.tar.bz2
+    # clean_fullband/datasets_fullband.clean_fullband.french_speech_001_NA_NA.tar.bz2
+    # clean_fullband/datasets_fullband.clean_fullband.french_speech_002_NA_NA.tar.bz2
+    # clean_fullband/datasets_fullband.clean_fullband.french_speech_003_NA_NA.tar.bz2
+    # clean_fullband/datasets_fullband.clean_fullband.french_speech_004_NA_NA.tar.bz2
+    # clean_fullband/datasets_fullband.clean_fullband.french_speech_005_NA_NA.tar.bz2
+    # clean_fullband/datasets_fullband.clean_fullband.french_speech_006_NA_NA.tar.bz2
+    # clean_fullband/datasets_fullband.clean_fullband.french_speech_007_NA_NA.tar.bz2
+    # clean_fullband/datasets_fullband.clean_fullband.french_speech_008_NA_NA.tar.bz2
+
+    # clean_fullband/datasets_fullband.clean_fullband.german_speech_000_0.00_3.47.tar.bz2
+    # clean_fullband/datasets_fullband.clean_fullband.german_speech_001_3.47_3.64.tar.bz2
[…]
