### 2022-09-22 f599585
update dns_icassp22 to use 48khz sampling rate
