### 2022-09-22 ca8239a
update dns_icassp22 to use 48khz sampling rate
