### 2022-09-22 8ff3ff9
Fix the path of calculate_rft.py
Given the suggestion in https://github.com/espnet/espnet/issues/4658
