+++ espnet2/asr/ctc.py
@@ -12,7 +12,7 @@ class CTC(torch.nn.Module):
         odim: dimension of outputs
         encoder_output_size: number of encoder projection units
         dropout_rate: dropout rate (0.0 ~ 1.0)
-        ctc_type: builtin or warpctc
+        ctc_type: builtin or gtnctc
         reduce: reduce the CTC loss into a scalar
         ignore_nan_grad: Same as zero_infinity (keeping for backward compatiblity)
         zero_infinity:  Whether to zero infinite losses and the associated gradients.
@@ -41,12 +41,6 @@ class CTC(torch.nn.Module):
             self.ctc_loss = torch.nn.CTCLoss(
                 reduction="none", zero_infinity=zero_infinity
             )
-        elif self.ctc_type == "warpctc":
-            import warpctc_pytorch as warp_ctc
-
-            if zero_infinity:
-                logging.warning("zero_infinity option is not supported for warp_ctc")
-            self.ctc_loss = warp_ctc.CTCLoss(size_average=True, reduce=reduce)
 
         elif self.ctc_type == "gtnctc":
             from espnet.nets.pytorch_backend.gtn_ctc import GTNCTCLossFunction
@@ -54,7 +48,7 @@ class CTC(torch.nn.Module):
             self.ctc_loss = GTNCTCLossFunction.apply
         else:
             raise ValueError(
-                f'ctc_type must be "builtin" or "warpctc": {self.ctc_type}'
+                f'ctc_type must be "builtin" or "gtnctc": {self.ctc_type}'
             )
 
         self.reduce = reduce
@@ -70,21 +64,6 @@ class CTC(torch.nn.Module):
                 loss = loss.sum() / size
             else:
                 loss = loss / size
-            return loss
-
-        elif self.ctc_type == "warpctc":
-            # warpctc only supports float32
-            th_pred = th_pred.to(dtype=torch.float32)
-
-            th_target = th_target.cpu().int()
-            th_ilen = th_ilen.cpu().int()
-            th_olen = th_olen.cpu().int()
-            loss = self.ctc_loss(th_pred, th_target, th_ilen, th_olen)
-            if self.reduce:
-                # NOTE: sum() is needed to keep consistency since warpctc
-                # return as tensor w/ shape (1,)
-                # but builtin return as tensor w/o shape (scalar).
-                loss = loss.sum()
             return loss
 
         elif self.ctc_type == "gtnctc":
