### 2022-09-23 6e03aed
fix ctc.py
