+++ espnet2/asr/ctc.py
@@ -1,5 +1,3 @@
-import logging
-
 import torch
 import torch.nn.functional as F
 from typeguard import check_argument_types
