+++ espnet/nets/chainer_backend/ctc.py
@@ -1,3 +1,5 @@
+import logging
+
 import chainer
 import chainer.functions as F
 import chainer.links as L
