+++ README.md
@@ -1,6 +1,6 @@
 <div align="left"><img src="doc/image/espnet_logo1.png" width="550"/></div>
 
-# ESPnet: end-to-end speech processing toolkit
+# ESPnet: end-to-end Multimodal processing toolkit
 
 |system/pytorch ver.|1.4.0|1.5.1|1.6.0|1.7.1|1.8.1|1.9.1|1.10.2|1.11.0|
 | :---: | :---: | :---: | :---: | :---: | :---: | :---: | :---: | :---: |
