+++ egs2/qasr_tts/tts1/run.sh
@@ -22,7 +22,7 @@ test_sets="dev eval1"
 
 train_config=conf/train.yaml
 inference_config=conf/decode.yaml
-
+g2p=g2p_en # Include word separator
 g2p=none # Include no word separator
 
 bash ./tts.sh \
