+++ egs2/qasr_tts/tts1/conf/train.yaml
@@ -1,1 +1,1 @@
-tuning/train_tacotron2.yaml
+/home/massa.baali/github_repos/espnet/espnet/bin/tts_train.py
