+++ egs2/qasr_tts/tts1/conf/train.yaml
@@ -1,1 +1,1 @@
-/home/massa.baali/github_repos/espnet/espnet/bin/tts_train.py
+tuning/finetune_transformer.yaml
