+++ egs2/dns_icassp22/enh1/local/noisyspeech_synthesizer.py
@@ -4,27 +4,34 @@
 
 # Adapted from DNS-Challenge official repo file: https://github.com/microsoft/DNS-Challenge/blob/master/noisyspeech_synthesizer_singleprocess.py
 # Original License: CC-BY 4.0 International: https://github.com/microsoft/DNS-Challenge/blob/master/LICENSE
-# Retrieved on Sep. 7th, 2022, by Shih-Lun Wu (summer7sean@gmail.com) 
+# Retrieved on Sep. 7th, 2022, by Shih-Lun Wu (summer7sean@gmail.com)
 
-import os
-import sys
-import multiprocessing
-import glob
 import argparse
 import ast
 import configparser as CP
-from random import shuffle
+import glob
+import multiprocessing
+import os
 import random
+import sys
+from pathlib import Path
+from random import shuffle
 
 import librosa
 import numpy as np
-from scipy import signal
-from audiolib import audioread, audiowrite, segmental_snr_mixer, activitydetector, is_clipped, add_clipping
-import utils
-
 import pandas as pd
-from pathlib import Path
+from audiolib import (
+    activitydetector,
+    add_clipping,
+    audioread,
+    audiowrite,
+    is_clipped,
+    segmental_snr_mixer,
+)
+from scipy import signal
 from scipy.io import wavfile
+
+import utils
 
 MAXTRIES = 50
 MAXFILELEN = 100
@@ -32,22 +39,24 @@ MAXFILELEN = 100
 # np.random.seed(5)
 # random.seed(5)
 
+
 def add_pyreverb(clean_speech, rir):
-    
+
     reverb_speech = signal.fftconvolve(clean_speech, rir, mode="full")
-    
+
     # make reverb_speech same length as clean_speech
     reverb_speech = reverb_speech[0 : clean_speech.shape[0]]
 
     return reverb_speech
 
-def build_audio(is_clean, params, index, audio_samples_length=-1):
-    '''Construct an audio signal from source files'''
 
-    fs_output = params['fs']
-    silence_length = params['silence_length']
+def build_audio(is_clean, params, index, audio_samples_length=-1):
+    """Construct an audio signal from source files"""
+
+    fs_output = params["fs"]
+    silence_length = params["silence_length"]
     if audio_samples_length == -1:
-        audio_samples_length = int(params['audio_length']*params['fs'])
+        audio_samples_length = int(params["audio_length"] * params["fs"])
 
     output_audio = np.zeros(0)
     remaining_length = audio_samples_length
@@ -55,25 +64,26 @@ def build_audio(is_clean, params, index, audio_samples_length=-1):
     clipped_files = []
 
     if is_clean:
-        source_files = params['cleanfilenames']
+        source_files = params["cleanfilenames"]
         idx = index
     else:
-        if 'noisefilenames' in params.keys():
-            source_files = params['noisefilenames']
+        if "noisefilenames" in params.keys():
+            source_files = params["noisefilenames"]
             idx = index
         # if noise files are organized into individual subdirectories, pick a directory randomly
         else:
-            noisedirs = params['noisedirs']
+            noisedirs = params["noisedirs"]
             # pick a noise category randomly
             idx_n_dir = np.random.randint(0, np.size(noisedirs))
-            source_files = glob.glob(os.path.join(noisedirs[idx_n_dir], 
-                                                  params['audioformat']))
+            source_files = glob.glob(
+                os.path.join(noisedirs[idx_n_dir], params["audioformat"])
+            )
             shuffle(source_files)
             # pick a noise source file index randomly
             idx = np.random.randint(0, np.size(source_files))
 
     # initialize silence
-    silence = np.zeros(int(fs_output*silence_length))
+    silence = np.zeros(int(fs_output * silence_length))
 
     # iterate through multiple clips until we have a long enough signal
[…]
