+++ egs2/dns_icassp22/enh1/local/audiolib.py
@@ -5,39 +5,44 @@
 
 # Copied from DNS-Challenge official repo file: https://github.com/microsoft/DNS-Challenge/edit/master/audiolib.py
 # Original License: CC-BY 4.0 International: https://github.com/microsoft/DNS-Challenge/blob/master/LICENSE
-# Retrieved on Sep. 7th, 2022, by Shih-Lun Wu (summer7sean@gmail.com) 
+# Retrieved on Sep. 7th, 2022, by Shih-Lun Wu (summer7sean@gmail.com)
 
+import glob
 import os
+import random
+import subprocess
+import tempfile
+
+import librosa
 import numpy as np
 import soundfile as sf
-import subprocess
-import glob
-import librosa
-import random
-import tempfile
 
 EPS = np.finfo(float).eps
 np.random.seed(0)
 
+
 def is_clipped(audio, clipping_threshold=0.99):
     return any(abs(audio) > clipping_threshold)
 
+
 def normalize(audio, target_level=-25):
-    '''Normalize the signal to the target level'''
-    rms = (audio ** 2).mean() ** 0.5
-    scalar = 10 ** (target_level / 20) / (rms+EPS)
+    """Normalize the signal to the target level"""
+    rms = (audio**2).mean() ** 0.5
+    scalar = 10 ** (target_level / 20) / (rms + EPS)
     audio = audio * scalar
     return audio
+
 
 def normalize_segmental_rms(audio, rms, target_level=-25):
-    '''Normalize the signal to the target level
-    based on segmental RMS'''
-    scalar = 10 ** (target_level / 20) / (rms+EPS)
+    """Normalize the signal to the target level
+    based on segmental RMS"""
+    scalar = 10 ** (target_level / 20) / (rms + EPS)
     audio = audio * scalar
     return audio
 
+
 def audioread(path, norm=False, start=0, stop=None, target_level=-25):
-    '''Function to read audio'''
+    """Function to read audio"""
 
     path = os.path.abspath(path)
     if not os.path.exists(path):
@@ -45,37 +50,47 @@ def audioread(path, norm=False, start=0, stop=None, target_level=-25):
     try:
         audio, sample_rate = sf.read(path, start=start, stop=stop)
     except RuntimeError:  # fix for sph pcm-embedded shortened v2
-        print('WARNING: Audio type not supported')
+        print("WARNING: Audio type not supported")
         return (None, None)
 
     if len(audio.shape) == 1:  # mono
         if norm:
-            rms = (audio ** 2).mean() ** 0.5
-            scalar = 10 ** (target_level / 20) / (rms+EPS)
+            rms = (audio**2).mean() ** 0.5
+            scalar = 10 ** (target_level / 20) / (rms + EPS)
             audio = audio * scalar
     else:  # multi-channel
         audio = audio.T
-        audio = audio.sum(axis=0)/audio.shape[0]
+        audio = audio.sum(axis=0) / audio.shape[0]
         if norm:
             audio = normalize(audio, target_level)
 
     return audio, sample_rate
 
 
-def audiowrite(destpath, audio, sample_rate=16000, norm=False, target_level=-25, \
-                clipping_threshold=0.99, clip_test=False):
-    '''Function to write audio'''
+def audiowrite(
+    destpath,
+    audio,
+    sample_rate=16000,
+    norm=False,
+    target_level=-25,
+    clipping_threshold=0.99,
+    clip_test=False,
+):
+    """Function to write audio"""
 
     if clip_test:
         if is_clipped(audio, clipping_threshold=clipping_threshold):
-            raise ValueError("Clipping detected in audiowrite()! " + \
[…]
