### 2022-09-25 d3e6a09
Create visual_light_hubert.yaml
