+++ egs2/open_li110/asr1/run.sh
@@ -13,10 +13,120 @@ _ru_rw_sah_sl_sv-SE_ta_tr_tt_uk_vi_\
 zh-CN_zh-HK_zh-TW"
 lid=true # whether to use language id as additional label
 
-train_set=train_li52_lid
-train_dev=dev_li52_lid
-# use the middle resource test set to avoid too long evaluation time
-test_set=${mid_resource_test_set}
+# train_set=train_pl_voxpopuli
+train_set=train_li110_lid
+train_dev=dev_li110_lid
+test_set="test_ab_commonvoice test_fy_NL_commonvoice test_or_commonvoice \
+test_af_openslr32 test_ga_IE_commonvoice test_pa_IN_commonvoice \
+test_ar_commonvoice test_gl_commonvoice \
+test_as_commonvoice test_gl_openslr77 test_pl_commonvoice \
+test_az_commonvoice test_gn_commonvoice test_pl_mls \
+test_ba_commonvoice test_gu_openslr78 test_pl_voxpopuli \
+test_bas_commonvoice test_ha_commonvoice test_pt_commonvoice \
+test_be_commonvoice test_hi_commonvoice test_pt_mls \
+test_bg_commonvoice test_hr_voxpopuli test_pt_voxforge \
+test_bn_commonvoice test_hsb_commonvoice test_rm_sursilv_commonvoice \
+test_bn_openslr37 test_hu_commonvoice \
+test_bn_openslr53 test_hu_voxpopuli test_rm_vallader_commonvoice \
+test_br_commonvoice test_hy_AM_commonvoice \
+test_ca_commonvoice test_ia_commonvoice test_ro_commonvoice \
+test_ca_openslr69 test_id_commonvoice test_ro_voxpopuli \
+test_ckb_commonvoice test_ig_commonvoice test_ru_commonvoice \
+test_cnh_commonvoice test_it_commonvoice test_ru_voxforge \
+test_cs_commonvoice test_it_mls test_rw_commonvoice \
+test_cs_voxpopuli test_it_voxforge test_sah_commonvoice \
+test_cv_commonvoice test_it_voxpopuli test_sat_commonvoice \
+test_cy_commonvoice test_ja_commonvoice test_si_openslr52 \
+test_da_commonvoice test_jv_openslr35 test_sk_commonvoice \
+test_de_commonvoice test_jv_openslr41_female test_sk_voxpopuli \
+test_de_mls test_jv_openslr41_male test_sl_commonvoice \
+test_de_voxforge test_kab_commonvoice test_sl_voxpopuli \
+test_de_voxpopuli test_ka_commonvoice test_sr_commonvoice \
+test_dv_commonvoice test_kk_commonvoice test_st_openslr32 \
+test_el_commonvoice test_km_openslr42_male test_su_openslr36 \
+test_en_commonvoice test_kmr_commonvoice test_su_openslr44_female \
+test_en_mls test_su_openslr44_male \
+test_en_openslr70 test_kn_openslr79 test_sv_SE_commonvoice \
+test_en_voxforge test_ky_commonvoice \
+test_en_voxpopuli test_lg_commonvoice test_sw_commonvoice \
+test_eo_commonvoice test_lt_commonvoice test_ta_commonvoice \
+test_es_commonvoice test_lt_voxpopuli test_ta_openslr65 \
+test_es_mls test_lv_commonvoice test_te_openslr66 \
+test_es_openslr61 test_mdf_commonvoice test_th_commonvoice \
+test_es_openslr71 test_mhr_commonvoice test_tig_commonvoice \
[…]
