+++ espnet2/svs/naive_rnn/naive_rnn_dp.py
@@ -10,7 +10,13 @@ import torch.nn.functional as F
 import logging
 from typeguard import check_argument_types
 
-from espnet.nets.pytorch_backend.nets_utils import make_non_pad_mask
+from espnet.nets.pytorch_backend.e2e_tts_fastspeech import (
+    FeedForwardTransformerLoss as FastSpeechLoss,
+)
+from espnet.nets.pytorch_backend.fastspeech.duration_predictor import DurationPredictor
+from espnet.nets.pytorch_backend.fastspeech.length_regulator import LengthRegulator
+
+from espnet.nets.pytorch_backend.nets_utils import make_non_pad_mask, make_pad_mask
 from espnet.nets.pytorch_backend.tacotron2.encoder import Encoder as EncoderPrenet
 from espnet.nets.pytorch_backend.tacotron2.decoder import Postnet
 from espnet.nets.pytorch_backend.transformer.mask import subsequent_mask
@@ -21,76 +27,11 @@ from espnet2.svs.abs_svs import AbsSVS
 import random
 from torch.distributions import Beta
 
-Beta_distribution = Beta(torch.tensor([0.5]), torch.tensor([0.5]))
 
-
-class NaiveRNNLoss(torch.nn.Module):
-    """Loss function module for Tacotron2."""
-
-    def __init__(self, use_masking=True, use_weighted_masking=False):
-        """Initialize Tactoron2 loss module.
-
-        Args:
-            use_masking (bool): Whether to apply masking
-                for padded part in loss calculation.
-            use_weighted_masking (bool):
-                Whether to apply weighted masking in loss calculation.
-        """
-        super(NaiveRNNLoss, self).__init__()
-        assert (use_masking != use_weighted_masking) or not use_masking
-        self.use_masking = use_masking
-        self.use_weighted_masking = use_weighted_masking
-
-        # define criterions
-        reduction = "none" if self.use_weighted_masking else "mean"
-        self.l1_criterion = torch.nn.L1Loss(reduction=reduction)
-        self.mse_criterion = torch.nn.MSELoss(reduction=reduction)
-
-        # NOTE(kan-bayashi): register pre hook function for the compatibility
-        # self._register_load_state_dict_pre_hook(self._load_state_dict_pre_hook)
-
-    def forward(self, after_outs, before_outs, ys, olens):
-        """Calculate forward propagation.
-
-        Args:
-            after_outs (Tensor): Batch of outputs after postnets (B, Lmax, odim).
-            before_outs (Tensor): Batch of outputs before postnets (B, Lmax, odim).
-            ys (Tensor): Batch of padded target features (B, Lmax, odim).
-            olens (LongTensor): Batch of the lengths of each target (B,).
-        
-        Returns:
-            Tensor: L1 loss value.
-            Tensor: Mean square error loss value.
-        """
-        # make mask and apply it
-        if self.use_masking:
-            masks = make_non_pad_mask(olens).unsqueeze(-1).to(ys.device)
-            ys = ys.masked_select(masks)
-            after_outs = after_outs.masked_select(masks)
-            before_outs = before_outs.masked_select(masks)
-
-        # calculate loss
-        l1_loss = self.l1_criterion(after_outs, ys) + self.l1_criterion(before_outs, ys)
-        mse_loss = self.mse_criterion(after_outs, ys) + self.mse_criterion(
-            before_outs, ys
-        )
-
-        # make weighted mask and apply it
-        if self.use_weighted_masking:
-            masks = make_non_pad_mask(olens).unsqueeze(-1).to(ys.device)
-            weights = masks.float() / masks.sum(dim=1, keepdim=True).float()
[…]
