+++ espnet2/train/preprocessor.py
@@ -875,14 +875,13 @@ class SVSPreprocessor(AbsPreprocessor):
         space_symbol: str = "<space>",
         non_linguistic_symbols: Union[Path, str, Iterable[str]] = None,
         delimiter: str = None,
-        rir_scp: str = None,
-        rir_apply_prob: float = 1.0,
         singing_volume_normalize: float = None,
         singing_name: str = "singing",
         text_name: str = "text",
         label_name: str = "label",
         midi_name: str = "midi",
         fs: np.int32 = 0,
+        align: list = ["singing", "label_lab", "midi_lab", "tempo_lab", "beat_lab"],  # TODO(Tao): add to args
     ):
         super().__init__(train)
         self.train = train
@@ -892,7 +891,7 @@ class SVSPreprocessor(AbsPreprocessor):
         self.midi_name = midi_name
         self.fs = fs
         self.singing_volume_normalize = singing_volume_normalize
-        self.rir_apply_prob = rir_apply_prob
+        self.align = align
         if token_type is not None:
             if token_list is None:
                 raise ValueError("token_list is required if token_type is not None")
@@ -914,19 +913,6 @@ class SVSPreprocessor(AbsPreprocessor):
             self.text_cleaner = None
             self.tokenizer = None
             self.token_id_converter = None
-        if train and rir_scp is not None:
-            self.rirs = []
-            with open(rir_scp, "r", encoding="utf-8") as f:
-                for line in f:
-                    sps = line.strip().split(None, 1)
-                    if len(sps) == 1:
-                        self.rirs.append(sps[0])
-                    else:
-                        self.rirs.append(sps[1])
-        else:
-            self.rirs = None
-
-        self.noises = None
 
     def __call__(
         self,
@@ -935,7 +921,14 @@ class SVSPreprocessor(AbsPreprocessor):
     ) -> Dict[str, np.ndarray]:
         assert check_argument_types()
 
-        if self.label_name in data and self.tokenizer is not None:
+        if self.singing_name in data:
+            if self.singing_volume_normalize is not None:
+                singing = data[self.singing_name]
+                ma = np.max(np.abs(singing))
+                data[self.singing_name] = singing * self.singing_volume_normalize / ma
+
+        if self.midi_name in data and self.label_name in data and self.tokenizer is not None:
+            # Load label info
             lab_timeseq, text = data[self.label_name]
             lab_len = len(text)
             text = " ".join(text)
@@ -943,111 +936,116 @@ class SVSPreprocessor(AbsPreprocessor):
             tokens = self.tokenizer.text2tokens(text)
             text_ints = self.token_id_converter.tokens2ids(tokens)
             data.pop(self.label_name)
+            
+            # Load xml info
+            syllables, notemidis, notetimeseq, tempo = data[self.midi_name]
+            midis = []
+            xml_timeseq = []
+            phn_cnt = 0
+            vowel = ["a", "i", "u", "e", "o"]
+            for i in range(len(syllables)):
+                # NOTE: Some phonemes are tagged differently
+                if syllables[i] == "へ":
+                    phn = ["h", "e"]
+                #elif syllables[i] == "でぇ":
+                #    phn = ["dy", "e"]
+                else:
+                    phn = pyopenjtalk.g2p(syllables[i]).split(' ')
+                    #if syllables[i][0] == "ヴ" and len(syllables[i]) == 2:
+                    #    phn.remove("u")
+                phn_num = len(phn)
+                vowel_num = 0
+                for p in phn:
+                    if p in vowel:
+                        vowel_num += 1
[…]
