### 2022-09-27 846aed8
Add tests on MacOS and Windows
