### 2022-09-27 d5891dd
Add tests on MacOS and Windows
