### 2022-09-27 f2a3edb
update run.sh (remove a empty test set)
