### 2022-09-27 ee7ae3f
update db.sh
