### 2022-09-27 99fd1e7
Update espnet2/train/preprocessor.py
Co-authored-by: Jiatong <728307998@qq.com>
