### 2022-09-27 2b7ac55
Update espnet2/train/preprocessor.py
Co-authored-by: Jiatong <728307998@qq.com>
