+++ _sources/_gen/espnet2.torch_utils.rst.txt
@@ -4,83 +4,6 @@ espnet2.torch_utils package
 
 
 
-.. _espnet2.torch_utils.pytorch_version:
-
-espnet2.torch_utils.pytorch_version
------------------------------------
-
-.. automodule:: espnet2.torch_utils.pytorch_version
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-.. _espnet2.torch_utils.forward_adaptor:
-
-espnet2.torch_utils.forward_adaptor
------------------------------------
-
-.. automodule:: espnet2.torch_utils.forward_adaptor
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-.. _espnet2.torch_utils.set_all_random_seed:
-
-espnet2.torch_utils.set_all_random_seed
----------------------------------------
-
-.. automodule:: espnet2.torch_utils.set_all_random_seed
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-.. _espnet2.torch_utils.load_pretrained_model:
-
-espnet2.torch_utils.load_pretrained_model
------------------------------------------
-
-.. automodule:: espnet2.torch_utils.load_pretrained_model
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-.. _espnet2.torch_utils.model_summary:
-
-espnet2.torch_utils.model_summary
----------------------------------
-
-.. automodule:: espnet2.torch_utils.model_summary
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-.. _espnet2.torch_utils.recursive_op:
-
-espnet2.torch_utils.recursive_op
---------------------------------
-
-.. automodule:: espnet2.torch_utils.recursive_op
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-.. _espnet2.torch_utils.device_funcs:
-
-espnet2.torch_utils.device_funcs
---------------------------------
-
-.. automodule:: espnet2.torch_utils.device_funcs
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
 .. _espnet2.torch_utils.initialize:
 
 espnet2.torch_utils.initialize
@@ -114,12 +37,89 @@ espnet2.torch_utils.add_gradient_noise
     :show-inheritance:
 
 
+.. _espnet2.torch_utils.load_pretrained_model:
+
+espnet2.torch_utils.load_pretrained_model
+-----------------------------------------
+
+.. automodule:: espnet2.torch_utils.load_pretrained_model
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+.. _espnet2.torch_utils.model_summary:
+
+espnet2.torch_utils.model_summary
+---------------------------------
+
+.. automodule:: espnet2.torch_utils.model_summary
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
 .. _espnet2.torch_utils.get_layer_from_string:
 
 espnet2.torch_utils.get_layer_from_string
 -----------------------------------------
 
 .. automodule:: espnet2.torch_utils.get_layer_from_string
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+.. _espnet2.torch_utils.device_funcs:
+
+espnet2.torch_utils.device_funcs
+--------------------------------
+
+.. automodule:: espnet2.torch_utils.device_funcs
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+.. _espnet2.torch_utils.set_all_random_seed:
+
+espnet2.torch_utils.set_all_random_seed
+---------------------------------------
+
+.. automodule:: espnet2.torch_utils.set_all_random_seed
+    :members:
+    :undoc-members:
[…]
