+++ _sources/_gen/espnet2.main_funcs.rst.txt
@@ -4,12 +4,12 @@ espnet2.main_funcs package
 
 
 
-.. _espnet2.main_funcs.collect_stats:
+.. _espnet2.main_funcs.__init__:
 
-espnet2.main_funcs.collect_stats
---------------------------------
+espnet2.main_funcs.__init__
+---------------------------
 
-.. automodule:: espnet2.main_funcs.collect_stats
+.. automodule:: espnet2.main_funcs.__init__
     :members:
     :undoc-members:
     :show-inheritance:
@@ -26,17 +26,6 @@ espnet2.main_funcs.calculate_all_attentions
     :show-inheritance:
 
 
-.. _espnet2.main_funcs.pack_funcs:
-
-espnet2.main_funcs.pack_funcs
------------------------------
-
-.. automodule:: espnet2.main_funcs.pack_funcs
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
 .. _espnet2.main_funcs.average_nbest_models:
 
 espnet2.main_funcs.average_nbest_models
@@ -48,12 +37,23 @@ espnet2.main_funcs.average_nbest_models
     :show-inheritance:
 
 
-.. _espnet2.main_funcs.__init__:
+.. _espnet2.main_funcs.collect_stats:
 
-espnet2.main_funcs.__init__
----------------------------
+espnet2.main_funcs.collect_stats
+--------------------------------
 
-.. automodule:: espnet2.main_funcs.__init__
+.. automodule:: espnet2.main_funcs.collect_stats
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+.. _espnet2.main_funcs.pack_funcs:
+
+espnet2.main_funcs.pack_funcs
+-----------------------------
+
+.. automodule:: espnet2.main_funcs.pack_funcs
     :members:
     :undoc-members:
     :show-inheritance:
