+++ _sources/_gen/espnet2.samplers.rst.txt
@@ -4,6 +4,17 @@ espnet2.samplers package
 
 
 
+.. _espnet2.samplers.length_batch_sampler:
+
+espnet2.samplers.length_batch_sampler
+-------------------------------------
+
+.. automodule:: espnet2.samplers.length_batch_sampler
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
 .. _espnet2.samplers.build_batch_sampler:
 
 espnet2.samplers.build_batch_sampler
@@ -15,12 +26,23 @@ espnet2.samplers.build_batch_sampler
     :show-inheritance:
 
 
-.. _espnet2.samplers.folded_batch_sampler:
+.. _espnet2.samplers.__init__:
 
-espnet2.samplers.folded_batch_sampler
--------------------------------------
+espnet2.samplers.__init__
+-------------------------
 
-.. automodule:: espnet2.samplers.folded_batch_sampler
+.. automodule:: espnet2.samplers.__init__
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+.. _espnet2.samplers.unsorted_batch_sampler:
+
+espnet2.samplers.unsorted_batch_sampler
+---------------------------------------
+
+.. automodule:: espnet2.samplers.unsorted_batch_sampler
     :members:
     :undoc-members:
     :show-inheritance:
@@ -48,23 +70,12 @@ espnet2.samplers.abs_sampler
     :show-inheritance:
 
 
-.. _espnet2.samplers.unsorted_batch_sampler:
+.. _espnet2.samplers.folded_batch_sampler:
 
-espnet2.samplers.unsorted_batch_sampler
----------------------------------------
-
-.. automodule:: espnet2.samplers.unsorted_batch_sampler
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-.. _espnet2.samplers.length_batch_sampler:
-
-espnet2.samplers.length_batch_sampler
+espnet2.samplers.folded_batch_sampler
 -------------------------------------
 
-.. automodule:: espnet2.samplers.length_batch_sampler
+.. automodule:: espnet2.samplers.folded_batch_sampler
     :members:
     :undoc-members:
     :show-inheritance:
@@ -76,17 +87,6 @@ espnet2.samplers.sorted_batch_sampler
 -------------------------------------
 
 .. automodule:: espnet2.samplers.sorted_batch_sampler
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-.. _espnet2.samplers.__init__:
-
-espnet2.samplers.__init__
--------------------------
-
-.. automodule:: espnet2.samplers.__init__
     :members:
     :undoc-members:
     :show-inheritance:

[…]
