### 2022-09-28 d955caf
download dns util scripts instead of copying; change raw data dest path
