### 2022-09-28 84db2a6
download dns util scripts instead of copying; change raw data dest path
