+++ espnet2/svs/naive_rnn/naive_rnn_dp.py
@@ -3,29 +3,27 @@
 
 """Transformer-SVS related modules."""
 
+import logging
+import random
 from typing import Dict, Optional, Sequence, Tuple
+
 import torch
 import torch.nn.functional as F
-
-import logging
+from torch.distributions import Beta
 from typeguard import check_argument_types
 
+from espnet2.svs.abs_svs import AbsSVS
+from espnet2.torch_utils.device_funcs import force_gatherable
+from espnet2.torch_utils.initialize import initialize
 from espnet.nets.pytorch_backend.e2e_tts_fastspeech import (
     FeedForwardTransformerLoss as FastSpeechLoss,
 )
 from espnet.nets.pytorch_backend.fastspeech.duration_predictor import DurationPredictor
 from espnet.nets.pytorch_backend.fastspeech.length_regulator import LengthRegulator
-
 from espnet.nets.pytorch_backend.nets_utils import make_non_pad_mask, make_pad_mask
-from espnet.nets.pytorch_backend.tacotron2.encoder import Encoder as EncoderPrenet
 from espnet.nets.pytorch_backend.tacotron2.decoder import Postnet
+from espnet.nets.pytorch_backend.tacotron2.encoder import Encoder as EncoderPrenet
 from espnet.nets.pytorch_backend.transformer.mask import subsequent_mask
-from espnet2.torch_utils.device_funcs import force_gatherable
-from espnet2.torch_utils.initialize import initialize
-from espnet2.svs.abs_svs import AbsSVS
-
-import random
-from torch.distributions import Beta
 
 
 class NaiveRNNDP(AbsSVS):
