+++ egs2/slurp_entity/asr1_no_intent/local/evaluation/util.py
@@ -1,9 +1,10 @@
 import json
 import logging
 import os
-from typing import Any, Dict, Tuple
-
 import tabulate
+
+from typing import Dict, Any, Tuple
+
 from progress.bar import Bar
 
 logging.basicConfig(
