+++ egs2/slurp_entity/asr1_no_intent/local/evaluation/metrics/__init__.py
@@ -1,2 +1,3 @@
 from .distance import Distance
-from .metrics import ErrorMetric, compute_metrics
+from .metrics import ErrorMetric
+from .metrics import compute_metrics
