### 2022-09-30 d63a9e1
use git clone to fetch scripts from DNS repo
