+++ egs2/ofuton_p_utagoe_db/svs1/local/data.sh
@@ -54,7 +54,6 @@ if [ ${stage} -le 2 ] && [ ${stop_stage} -ge 2 ]; then
         local/prep_segments.py --silence pau --silence sil --silence br ${src_data} 10000 # in ms
         mv ${src_data}/segments.tmp ${src_data}/segments
         mv ${src_data}/label.tmp ${src_data}/label
-        mv ${src_data}/phoneme.tmp ${src_data}/phoneme # can be removed if not used
         local/prep_segments_from_xml.py ${src_data} 10000 # in ms
         mv ${src_data}/text.tmp ${src_data}/text
         mv ${src_data}/segments_from_xml.tmp ${src_data}/segments_from_xml
