+++ egs2/ofuton_p_utagoe_db/svs1/local/prep_segments_from_xml.py
@@ -3,6 +3,7 @@ import argparse
 import math
 import os
 import sys
+
 import music21 as m21
 
 from espnet2.fileio.xml_scp import XMLScpReader
