+++ espnet2/svs/espnet_model.py
@@ -4,23 +4,23 @@
 
 """Singing-voice-synthesis ESPnet model."""
 
+import logging
 from contextlib import contextmanager
 from distutils.version import LooseVersion
-import logging
 from typing import Dict, Optional, Tuple
 
 import torch
-
 from typeguard import check_argument_types
 
 from espnet2.layers.abs_normalize import AbsNormalize
 from espnet2.layers.inversible_interface import InversibleInterface
-from espnet2.train.abs_espnet_model import AbsESPnetModel
 from espnet2.svs.abs_svs import AbsSVS
+from espnet2.svs.feats_extract.score_feats_extract import (
+    FrameScoreFeats,
+    SyllableScoreFeats,
+)
+from espnet2.train.abs_espnet_model import AbsESPnetModel
 from espnet2.tts.feats_extract.abs_feats_extract import AbsFeatsExtract
-
-from espnet2.svs.feats_extract.score_feats_extract import FrameScoreFeats
-from espnet2.svs.feats_extract.score_feats_extract import SyllableScoreFeats
 from espnet.nets.pytorch_backend.nets_utils import pad_list
 
 if LooseVersion(torch.__version__) >= LooseVersion("1.6.0"):
