+++ espnet2/svs/espnet_model.py
@@ -1,5 +1,6 @@
 # Copyright 2020 Nagoya University (Tomoki Hayashi)
 # Copyright 2021 Carnegie Mellon University (Jiatong Shi)
+# Copyright 2022 Renmin University of China (Yuning Wu)
 #  Apache 2.0  (http://www.apache.org/licenses/LICENSE-2.0)
 
 """Singing-voice-synthesis ESPnet model."""
@@ -149,7 +150,8 @@ class ESPnetSVSModel(AbsESPnetModel):
             # score : 128 midi pitch
             # tempo : bpm
             # duration :
-            #   input-> phone-id seqence | output -> frame level(take mode from window) or syllable level
+            #   input-> phone-id seqence
+            #   output -> frame level(take mode from window) or syllable level
             ds = None
             if isinstance(self.score_feats_extract, FrameScoreFeats):
                 (
@@ -177,7 +179,7 @@ class ESPnetSVSModel(AbsESPnetModel):
 
                 # calculate durations, new text & text_length
                 # Syllable Level duration info needs phone
-                # NOTE(Shuai) Duplicate adjacent phones will appear in text files sometimes
+                # NOTE(Shuai) Duplicate adjacent phones appear in text files sometimes
                 # e.g. oniku_0000000000000000hato_0002
                 # 10.951 11.107 sh
                 # 11.107 11.336 i
@@ -259,7 +261,7 @@ class ESPnetSVSModel(AbsESPnetModel):
                     :, : midiFrame_lab_lengths.max()
                 ]  # for data-parallel
 
-                # Extract Syllable Level label, midi, tempo, beat information from Frame Level
+                # Extract Syllable Level label, midi, tempo, beat from Frame Level
                 (
                     label_lab_after,
                     label_lab_lengths_after,
@@ -307,7 +309,7 @@ class ESPnetSVSModel(AbsESPnetModel):
                     :, : midiFrame_xml_lengths.max()
                 ]  # for data-parallel
 
-                # Extract Syllable Level label, midi, tempo, beat information from Frame Level
+                # Extract Syllable Level label, midi, tempo, beat from Frame Level
                 (
                     label_xml_after,
                     label_xml_lengths_after,
@@ -328,7 +330,7 @@ class ESPnetSVSModel(AbsESPnetModel):
                     beat_lengths=beatFrame_xml_lengths,
                 )
 
-                # calculate durations, represent syllable encoder outputs to feats mapping
+                # calculate durations for feature mapping
                 # Syllable Level duration info needs phone & midi
                 ds = []
                 for i, _ in enumerate(labelFrame_lab_lengths):
@@ -721,7 +723,7 @@ class ESPnetSVSModel(AbsESPnetModel):
             and tempo_xml_lengths == beat_xml_lengths
         )
 
-        # unsqueeze of singing must be here, or it'll cause error in the return dim of STFT
+        # unsqueeze of singing needed otherwise causing error in STFT dimension
         # for data-parallel
         text = text.unsqueeze(0)
         label_lab = label_lab.unsqueeze(0)
@@ -738,7 +740,8 @@ class ESPnetSVSModel(AbsESPnetModel):
         # score : 128 midi pitch
         # tempo : bpm
         # duration :
-        #   input-> phone-id seqence | output -> frame level(取众数 from window) or syllable level
+        #   input-> phone-id seqence
+        #   output -> frame level or syllable level
         ds = None
         batch_size = text.size(0)
         assert batch_size == 1
@@ -784,7 +787,6 @@ class ESPnetSVSModel(AbsESPnetModel):
                 ds.append(counts)
             ds = pad_list(ds, pad_value=0).to(text.device)
[…]
