### 2022-09-30 c61baeb
Implement talromur2 recipe data preprocessing scripts
