+++ egs2/talromur2/tts1/conf/tuning/train_xvector_tacotron2.yaml
@@ -54,7 +54,7 @@ optim_conf:           # keyword arguments for selected optimizer
 #                OTHER TRAINING SETTING                  #
 ##########################################################
 num_iters_per_epoch: 500    # number of iterations per epoch
-max_epoch: 500              # number of epochs
+max_epoch: 200              # number of epochs
 grad_clip: 1.0              # gradient clipping norm
 grad_noise: false           # whether to use gradient noise injection
 accum_grad: 1               # gradient accumulation
