+++ egs2/talromur2/tts1/conf/tuning/train_gst_fastspeech.yaml
@@ -1,85 +1,0 @@
-# This configuration is for ESPnet2 to train GST-FastSpeech.
-# It requires only a single GPU with 12 GB memory and it
-# takes ~4 days to finish the training on Titan V.
-
-# As a default, we assume that the training is based on knowledge
-# distillation, i.e., using teacher model outputs as the target.
-# This assumes that we always use feat_type=fbank. If you use
-# teacher forcing, please be careful to change batch_bins
-# according to feat_type (in default, we use feats_type=raw).
-
-##########################################################
-#                  TTS MODEL SETTING                     #
-##########################################################
-tts: fastspeech            # model architecture
-tts_conf:                  # keyword arguments for the selected model
-    adim: 384              # attention dimension
-    aheads: 4              # number of attention heads
-    elayers: 6             # number of encoder layers
-    eunits: 1536           # number of encoder ff units
-    dlayers: 6             # number of decoder layers
-    dunits: 1536           # number of decoder ff units
-    positionwise_layer_type: conv1d   # type of position-wise layer
-    positionwise_conv_kernel_size: 3  # kernel size of position wise conv layer
-    duration_predictor_layers: 2      # number of layers of duration predictor
-    duration_predictor_chans: 384     # number of channels of duration predictor
-    duration_predictor_kernel_size: 3 # filter size of duration predictor
-    postnet_layers: 5                 # number of layers of postnset
-    postnet_filts: 5                  # filter size of conv layers in postnet
-    postnet_chans: 256                # number of channels of conv layers in postnet
-    use_masking: True                 # whether to apply masking for padded part in loss calculation
-    use_scaled_pos_enc: True          # whether to use scaled positional encoding
-    encoder_normalize_before: True    # whether to perform layer normalization before the input
-    decoder_normalize_before: True    # whether to perform layer normalization before the input
-    reduction_factor: 1               # reduction factor
-    init_type: xavier_uniform         # initialization type
-    init_enc_alpha: 1.0               # initial value of alpha of encoder scaled position encoding
-    init_dec_alpha: 1.0               # initial value of alpha of decoder scaled position encoding
-    transformer_enc_dropout_rate: 0.1                # dropout rate for transformer encoder layer
-    transformer_enc_positional_dropout_rate: 0.1     # dropout rate for transformer encoder positional encoding
-    transformer_enc_attn_dropout_rate: 0.1           # dropout rate for transformer encoder attention layer
-    transformer_dec_dropout_rate: 0.1                # dropout rate for transformer decoder layer
-    transformer_dec_positional_dropout_rate: 0.1     # dropout rate for transformer decoder positional encoding
-    transformer_dec_attn_dropout_rate: 0.1           # dropout rate for transformer decoder attention layer
-    use_gst: true                 # whether to use GST embedding
-    gst_heads: 8                  # number of heads in GST multi-head attention
-    gst_tokens: 128               # number of global style tokens
-
-##########################################################
-#            OPTIMIZER & SCHEDULER SETTING               #
-##########################################################
-optim: adam            # optimizer type
-optim_conf:            # keyword arguments for selected optimizer
-    lr: 1.0            # learning rate
-scheduler: noamlr      # scheduler type
-scheduler_conf:        # keyword arguments for selected scheduler
-    model_size: 384    # model size, a.k.a., attention dimenstion
-    warmup_steps: 4000 # the number of warmup steps
-
-##########################################################
-#                OTHER TRAINING SETTING                  #
-##########################################################
-num_iters_per_epoch: 1000   # number of iterations per epoch
-max_epoch: 300              # number of epochs
-grad_clip: 1.0              # gradient clipping norm
[…]
