+++ egs2/talromur2/tts1/README.md
@@ -1,960 +1,51 @@
-# ESPnet2 TTS Recipe TEMPLATE
+# Talrómur 2 recipe
 
-This is a template of TTS recipe for ESPnet2.
+This is a recipe for the [Talrómur 2 corpus](https://repository.clarin.is/repository/xmlui/handle/20.500.12537/167), which is an Icelandic speech corpus intended for multi-speaker TTS development, containing studio recordings of speech from 40 Icelandic speakers, 20 male and 20 female. 
 
-## Table of Contents
+## Corpus
+The corpus contains 56,225 studio-recorded single-sentence audio clips.Each speaker in the corpus contributes 929 and 1879 clips.
 
-* [ESPnet2 TTS Recipe TEMPLATE](#espnet2-tts-recipe-template)
-  * [Table of Contents](#table-of-contents)
-  * [Recipe flow](#recipe-flow)
-    * [1\. Data preparation](#1-data-preparation)
-    * [2\. Wav dump / Embedding preparation](#2-wav-dump--embedding-preparation)
-    * [3\. Removal of long / short data](#3-removal-of-long--short-data)
-    * [4\. Token list generation](#4-token-list-generation)
-    * [5\. TTS statistics collection](#5-tts-statistics-collection)
-    * [6\. TTS training](#6-tts-training)
-    * [7\. TTS decoding](#7-tts-decoding)
-    * [8\-9\. (Optional) Pack results for upload](#8-9-optional-pack-results-for-upload)
-  * [How to run](#how-to-run)
-    * [FastSpeech training](#fastspeech-training)
-    * [FastSpeech2 training](#fastspeech2-training)
-    * [Multi speaker model with X-vector training](#multi-speaker-model-with-x-vector-training)
-    * [Multi speaker model with speaker ID embedding training](#multi-speaker-model-with-speaker-id-embedding-training)
-    * [Multi language model with language ID embedding training](#multi-language-model-with-language-id-embedding-training)
-    * [VITS training](#vits-training)
-    * [Joint text2wav training](#joint-text2wav-training)
-    * [Evaluation](#evaluation)
-  * [Supported text frontend](#supported-text-frontend)
-  * [Supported text cleaner](#supported-text-cleaner)
-  * [Supported Models](#supported-models)
-    * [Single speaker model](#single-speaker-model)
-    * [Multi speaker model](#multi-speaker-model)
-  * [FAQ](#faq)
-    * [ESPnet1 model is compatible with ESPnet2?](#espnet1-model-is-compatible-with-espnet2)
-    * [How to change minibatch size in training?](#how-to-change-minibatch-size-in-training)
-    * [How to make a new recipe for my own dataset?](#how-to-make-a-new-recipe-for-my-own-dataset)
-    * [How to add a new g2p module?](#how-to-add-a-new-g2p-module)
-    * [How to add a new cleaner module?](#how-to-add-a-new-cleaner-module)
-    * [How to use trained model in python?](#how-to-use-trained-model-in-python)
-    * [How to get pretrained models?](#how-to-get-pretrained-models)
-    * [How to load the pretrained parameters?](#how-to-load-the-pretrained-parameters)
-    * [How to finetune the pretrained model?](#how-to-finetune-the-pretrained-model)
-    * [How to add a new model?](#how-to-add-a-new-model)
-    * [How to test my model with an arbitrary given text?](#how-to-test-my-model-with-an-arbitrary-given-text)
-    * [How to train vocoder?](#how-to-train-vocoder)
-    * [How to train vocoder with text2mel GTA outputs?](#how-to-train-vocoder-with-text2mel-gta-outputs)
[…]
