### 2022-09-30 9989814
Add Talromur 2 entry in  egs2/README.md
