### 2022-10-01 8337edc
Update xml_scp.py
