+++ espnet2/gan_svs/style_melgan/__init__.py
@@ -1,0 +1,2 @@
+from espnet2.gan_tts.style_melgan.style_melgan import StyleMelGANDiscriminator  # NOQA
+from espnet2.gan_tts.style_melgan.style_melgan import StyleMelGANGenerator  # NOQA
