+++ espnet2/gan_svs/jets/__init__.py
@@ -1,0 +1,1 @@
+from espnet2.gan_tts.jets.jets import JETS  # NOQA

[…]
