+++ espnet2/gan_svs/vits/vits.py
@@ -1,0 +1,653 @@
+# Copyright 2021 Tomoki Hayashi
+#  Apache 2.0  (http://www.apache.org/licenses/LICENSE-2.0)
+
+"""VITS module for GAN-SVS task."""
+
+from contextlib import contextmanager
+from distutils.version import LooseVersion
+from typing import Any, Dict, Optional
+
+import torch
+from typeguard import check_argument_types
+
+from espnet2.gan_svs.abs_gan_svs import AbsGANSVS
+from espnet2.gan_tts.hifigan import (
+    HiFiGANMultiPeriodDiscriminator,
+    HiFiGANMultiScaleDiscriminator,
+    HiFiGANMultiScaleMultiPeriodDiscriminator,
+    HiFiGANPeriodDiscriminator,
+    HiFiGANScaleDiscriminator,
+)
+from espnet2.gan_tts.hifigan.loss import (
+    DiscriminatorAdversarialLoss,
+    FeatureMatchLoss,
+    GeneratorAdversarialLoss,
+    MelSpectrogramLoss,
+)
+from espnet2.gan_tts.utils import get_segments
+from espnet2.gan_svs.vits.generator import VITSGenerator
+from espnet2.gan_tts.vits.loss import KLDivergenceLoss
+from espnet2.torch_utils.device_funcs import force_gatherable
+
+AVAILABLE_GENERATERS = {
+    "vits_generator": VITSGenerator,
+}
+AVAILABLE_DISCRIMINATORS = {
+    "hifigan_period_discriminator": HiFiGANPeriodDiscriminator,
+    "hifigan_scale_discriminator": HiFiGANScaleDiscriminator,
+    "hifigan_multi_period_discriminator": HiFiGANMultiPeriodDiscriminator,
+    "hifigan_multi_scale_discriminator": HiFiGANMultiScaleDiscriminator,
+    "hifigan_multi_scale_multi_period_discriminator": HiFiGANMultiScaleMultiPeriodDiscriminator,  # NOQA
+}
+
+if LooseVersion(torch.__version__) >= LooseVersion("1.6.0"):
+    from torch.cuda.amp import autocast
+else:
+    # Nothing to do if torch<1.6.0
+    @contextmanager
+    def autocast(enabled=True):  # NOQA
+        yield
+
+
+class VITS(AbsGANSVS):
+    """VITS module (generator + discriminator).
+
+    This is a module of VITS described in `Conditional Variational Autoencoder
+    with Adversarial Learning for End-to-End Text-to-Speech`_.
+
+    .. _`Conditional Variational Autoencoder with Adversarial Learning for End-to-End
+        Text-to-Speech`: https://arxiv.org/abs/2006.04558
+
+    """
+
+    def __init__(
+        self,
+        # generator related
+        idim: int,
+        midi_dim: int,
+        odim: int,
+        midi_embed_integration_type: str = "add",
+        sampling_rate: int = 22050,
+        generator_type: str = "vits_generator",
+        generator_params: Dict[str, Any] = {
+            "hidden_channels": 192,
+            "spks": None,
+            "langs": None,
+            "spk_embed_dim": None,
+            "global_channels": -1,
+            "segment_size": 32,
+            "text_encoder_attention_heads": 2,
+            "text_encoder_ffn_expand": 4,
+            "text_encoder_blocks": 6,
+            "text_encoder_positionwise_layer_type": "conv1d",
+            "text_encoder_positionwise_conv_kernel_size": 1,
+            "text_encoder_positional_encoding_layer_type": "rel_pos",
+            "text_encoder_self_attention_layer_type": "rel_selfattn",
+            "text_encoder_activation_type": "swish",
+            "text_encoder_normalize_before": True,
[…]
