+++ espnet2/tasks/gan_svs.py
@@ -1,0 +1,456 @@
+# Copyright 2021 Tomoki Hayashi
+#  Apache 2.0  (http://www.apache.org/licenses/LICENSE-2.0)
+
+"""GAN-based Singing-voice-synthesis task."""
+
+import argparse
+import logging
+from typing import Callable, Collection, Dict, List, Optional, Tuple
+
+import numpy as np
+import torch
+from typeguard import check_argument_types, check_return_type
+
+from espnet2.gan_svs.abs_gan_svs import AbsGANSVS
+from espnet2.gan_svs.espnet_model import ESPnetGANSVSModel
+
+# from espnet2.gan_tts.jets import JETS
+# from espnet2.gan_tts.joint import JointText2Wav
+from espnet2.gan_svs.vits import VITS
+from espnet2.layers.abs_normalize import AbsNormalize
+from espnet2.layers.global_mvn import GlobalMVN
+from espnet2.layers.utterance_mvn import UtteranceMVN
+from espnet2.tasks.abs_task import AbsTask, optim_classes
+from espnet2.text.phoneme_tokenizer import g2p_choices
+from espnet2.train.class_choices import ClassChoices
+from espnet2.train.collate_fn import CommonCollateFn
+from espnet2.train.gan_trainer import GANTrainer
+from espnet2.train.preprocessor import SVSPreprocessor
+from espnet2.tts.feats_extract.abs_feats_extract import AbsFeatsExtract
+from espnet2.tts.feats_extract.dio import Dio
+from espnet2.svs.feats_extract.score_feats_extract import FrameScoreFeats
+from espnet2.svs.feats_extract.score_feats_extract import SyllableScoreFeats
+from espnet2.tts.feats_extract.energy import Energy
+from espnet2.tts.feats_extract.linear_spectrogram import LinearSpectrogram
+from espnet2.tts.feats_extract.log_mel_fbank import LogMelFbank
+from espnet2.tts.feats_extract.log_spectrogram import LogSpectrogram
+from espnet2.utils.get_default_kwargs import get_default_kwargs
+from espnet2.utils.nested_dict_action import NestedDictAction
+from espnet2.utils.types import int_or_none, str2bool, str_or_none
+
+feats_extractor_choices = ClassChoices(
+    "feats_extract",
+    classes=dict(
+        fbank=LogMelFbank,
+        log_spectrogram=LogSpectrogram,
+        linear_spectrogram=LinearSpectrogram,
+    ),
+    type_check=AbsFeatsExtract,
+    default="linear_spectrogram",
+)
+
+score_feats_extractor_choices = ClassChoices(
+    "score_feats_extract",
+    classes=dict(
+        frame_score_feats=FrameScoreFeats, syllable_score_feats=SyllableScoreFeats
+    ),
+    type_check=AbsFeatsExtract,
+    default="frame_score_feats",
+)
+
+pitch_extractor_choices = ClassChoices(
+    "pitch_extract",
+    classes=dict(dio=Dio),
+    type_check=AbsFeatsExtract,
+    default=None,
+    optional=True,
+)
+energy_extractor_choices = ClassChoices(
+    "energy_extract",
+    classes=dict(energy=Energy),
+    type_check=AbsFeatsExtract,
+    default=None,
+    optional=True,
+)
+normalize_choices = ClassChoices(
+    "normalize",
+    classes=dict(
+        global_mvn=GlobalMVN,
+        utterance_mvn=UtteranceMVN,
+    ),
+    type_check=AbsNormalize,
+    default=None,
+    optional=True,
+)
+pitch_normalize_choices = ClassChoices(
+    "pitch_normalize",
+    classes=dict(
[…]
