+++ egs2/TEMPLATE/asr1/db.sh
@@ -157,6 +157,7 @@ HARPERVALLEY=downloads
 TALROMUR=downloads
 DCASE=
 TEDX_SPANISH=downloads
+OFUTON=/data1/wyn/Mus_data/OFUTON_P_UTAGOE_DB/
 
 # For only CMU TIR environment
 if [[ "$(hostname)" == tir* ]]; then
