+++ egs2/ofuton_p_utagoe_db/svs1/db.sh
@@ -1,1 +1,1 @@
-../../TEMPLATE/svs1/db.sh
+../../TEMPLATE/asr1/db.sh
