+++ espnet/version.txt
@@ -1,1 +1,1 @@
-202207
+202209
