+++ _sources/notebook/onnx_conversion_demo.ipynb.txt
@@ -1,0 +1,429 @@
+{
+  "nbformat": 4,
+  "nbformat_minor": 0,
+  "metadata": {
+    "colab": {
+      "name": "espnet_onnx.ipynb",
+      "provenance": [],
+      "toc_visible": true
+    },
+    "kernelspec": {
+      "name": "python3",
+      "display_name": "Python 3"
+    },
+    "language_info": {
+      "name": "python"
+    }
+  },
+  "cells": [
+    {
+      "cell_type": "markdown",
+      "source": [
+        "# espnet_onnx demonstration\n",
+        "\n",
+        "This notebook provides a demonstration of how to export your trained model into onnx format.\n",
+        "Currently only ASR is supported.\n",
+        "\n",
+        "see also:\n",
+        "- ESPnet: https://github.com/espnet/espnet\n",
+        "- espnet_onnx: https://github.com/Masao-Someki/espnet_onnx\n",
+        "\n",
+        "Author: [Masao Someki](https://github.com/Masao-Someki)\n",
+        "\n",
+        "\n",
+        "## Table of Contents\n",
+        "\n",
+        "- Install Dependency\n",
+        "- Export your model\n",
+        "- Inference with onnx\n",
+        "- Using streaming model"
+      ],
+      "metadata": {
+        "id": "CBgUqWroOulc"
+      }
+    },
+    {
+      "cell_type": "markdown",
+      "source": [
+        "# Install Dependency"
+      ],
+      "metadata": {
+        "id": "lWtGQwmqTzn0"
+      }
+    },
+    {
+      "cell_type": "markdown",
+      "source": [
+        "To run this demo, you need to install the following packages.\n",
+        "- espnet_onnx\n",
+        "- torch >= 1.11.0 (already installed in Colab)\n",
+        "- espnet\n",
+        "- espnet_model_zoo\n",
+        "- onnx\n",
+        "\n",
+        "`torch`, `espnet`, `espnet_model_zoo`, `onnx` is required to run the exportation demo."
+      ],
+      "metadata": {
+        "id": "D7CYtP6-T2TD"
+      }
+    },
+    {
+      "cell_type": "code",
+      "source": [
+        "!pip install -U espnet_onnx espnet espnet_model_zoo onnx\n",
+        "\n",
+        "# in this demo, we need to update scipy to avoid an error\n",
+        "!pip install -U scipy"
+      ],
+      "metadata": {
+        "id": "4FEU9BUoUAqi"
+      },
+      "execution_count": null,
+      "outputs": []
+    },
+    {
+      "cell_type": "markdown",
+      "source": [
+        "# Export your model"
+      ],
[…]
