+++ _sources/notebook/espnet2_new_task_tutorial_CMU_11751_18781_Fall2022.ipynb.txt
@@ -1,0 +1,557 @@
+{
+  "cells": [
+    {
+      "cell_type": "markdown",
+      "metadata": {
+        "id": "kNbxe3gyfbPi"
+      },
+      "source": [
+        "# CMU 11751/18781 Fall 2022: ESPnet Tutorial2 (New task)\n",
+        "\n",
+        "[ESPnet](https://github.com/espnet/espnet) is a widely-used end-to-end speech processing toolkit. It has supported various speech processing tasks. ESPnet uses PyTorch as a main deep learning engine, and also follows Kaldi style data processing, feature extraction/format, and recipes to provide a complete setup for speech recognition and other speech processing experiments.\n",
+        "\n",
+        "Main references:\n",
+        "- [ESPnet repository](https://github.com/espnet/espnet)\n",
+        "- [ESPnet documentation](https://espnet.github.io/espnet/)\n",
+        "- [ESPnet tutorial in Speech Recognition and Understanding (Fall 2021)](https://colab.research.google.com/github/espnet/notebook/blob/master/espnet2_tutorial_2021_CMU_11751_18781.ipynb)\n",
+        "- [Recitation in Multilingual NLP (Spring 2022)](https://colab.research.google.com/drive/1tY6PxF_M5Nx5n488x0DrpujJOyqW-ATi?usp=sharing)\n",
+        "- [ESPnet tutorial1 in Speech Recognition and Understanding (Fall 2022)](https://colab.research.google.com/drive/1d3yMY3xv_hpr2mznskZpF2Cj-bEooDv-?usp=sharing)\n",
+        "\n",
+        "\n",
+        "Author: Jiatong Shi (jiatongs@andrew.cmu.edu)\n",
+        "\n",
+        "We would like to thank You (Neil) Zhang for kindly helping the hands-on tutorial and sharing his knowledge on the task. "
+      ]
+    },
+    {
+      "cell_type": "markdown",
+      "metadata": {
+        "id": "pwxgDdlMUE60"
+      },
+      "source": [
+        "## ❗Important Notes❗\n",
+        "- We are using Colab to show the demo. However, Colab has some constraints on the total GPU runtime. If you use too much GPU time, you may not be able to use GPU for some time.\n",
+        "- There are multiple in-class checkpoints ✅ throughout this tutorial. There will also be some after-class excersices 📗 after the tutorial. **Your participation points are based on these tasks.** Please try your best to follow all the steps! If you encounter issues, please notify the TAs as soon as possible so that we can make an adjustment for you.\n",
+        "- Please submit PDF files of your completed notebooks to Gradescope. You can print the notebook using `File -> Print` in the menu bar.\n",
+        "- This tutorial covers some advanced usage of ESPnet, which is the extension of the first tutorial."
+      ]
+    },
+    {
+      "cell_type": "markdown",
+      "metadata": {
+        "id": "gGg1N9jufpf2"
+      },
+      "source": [
+        "## Useful links\n",
+        "\n",
+        "- Installation https://espnet.github.io/espnet/installation.html\n",
+        "- Usage https://espnet.github.io/espnet/espnet2_tutorial.html\n",
+        "- Reference of task class in ESPnet https://espnet.github.io/espnet/espnet2_task.html\n"
+      ]
+    },
+    {
+      "cell_type": "markdown",
+      "metadata": {
+        "id": "wl9JFMNJ5iYu"
+      },
+      "source": [
+        "## Objectives\n",
+        "After this tutorial, you are expected to know:\n",
+        "- How to add new task in ESPnet2\n",
+        "- How to add new models in ESPnet2\n",
+        "- How to create a new recipe (and template) of a new task from scratch\n"
+      ]
+    },
+    {
[…]
