### 2022-10-04 4897058
support ST without texts in source language
