### 2022-10-04 3c95846
support ST without texts in source language
