### 2022-10-04 7cb7eef
vits singing draft 2
Added midi, tempo, and beat embedding
