+++ espnet2/gan_svs/vits/vits.py
@@ -64,12 +64,12 @@ class VITS(AbsGANSVS):
         self,
         # generator related
         idim: int,
-        midi_dim: int,
         odim: int,
-        midi_embed_integration_type: str = "add",
         sampling_rate: int = 22050,
         generator_type: str = "vits_generator",
         generator_params: Dict[str, Any] = {
+            "midi_dim": 129,
+            "midi_embed_integration_type": "add",
             "hidden_channels": 192,
             "spks": None,
             "langs": None,
@@ -338,6 +338,22 @@ class VITS(AbsGANSVS):
                 feats_lengths=feats_lengths,
                 singing=singing,
                 singing_lengths=singing_lengths,
+                label_lab=label_lab,
+                label_lab_lengths=label_lab_lengths,
+                label_xml=label_xml,
+                label_xml_lengths=label_xml_lengths,
+                midi_lab=midi_lab,
+                midi_lab_lengths=midi_lab_lengths,
+                midi_xml=midi_xml,
+                midi_xml_lengths=midi_xml_lengths,
+                tempo_lab=tempo_lab,
+                tempo_lab_lengths=tempo_lab_lengths,
+                tempo_xml=tempo_xml,
+                tempo_xml_lengths=tempo_xml_lengths,
+                beat_lab=beat_lab,
+                beat_lab_lengths=beat_lab_lengths,
+                beat_xml=beat_xml,
+                beat_xml_lengths=beat_xml_lengths,
                 sids=sids,
                 spembs=spembs,
                 lids=lids,
@@ -350,6 +366,22 @@ class VITS(AbsGANSVS):
                 feats_lengths=feats_lengths,
                 singing=singing,
                 singing_lengths=singing_lengths,
+                label_lab=label_lab,
+                label_lab_lengths=label_lab_lengths,
+                label_xml=label_xml,
+                label_xml_lengths=label_xml_lengths,
+                midi_lab=midi_lab,
+                midi_lab_lengths=midi_lab_lengths,
+                midi_xml=midi_xml,
+                midi_xml_lengths=midi_xml_lengths,
+                tempo_lab=tempo_lab,
+                tempo_lab_lengths=tempo_lab_lengths,
+                tempo_xml=tempo_xml,
+                tempo_xml_lengths=tempo_xml_lengths,
+                beat_lab=beat_lab,
+                beat_lab_lengths=beat_lab_lengths,
+                beat_xml=beat_xml,
+                beat_xml_lengths=beat_xml_lengths,
                 sids=sids,
                 spembs=spembs,
                 lids=lids,
@@ -363,6 +395,22 @@ class VITS(AbsGANSVS):
         feats_lengths: torch.Tensor,
         singing: torch.Tensor,
         singing_lengths: torch.Tensor,
+        label_lab: Optional[torch.Tensor] = None,
+        label_lab_lengths: Optional[torch.Tensor] = None,
+        label_xml: Optional[torch.Tensor] = None,
+        label_xml_lengths: Optional[torch.Tensor] = None,
+        midi_lab: Optional[torch.Tensor] = None,
+        midi_lab_lengths: Optional[torch.Tensor] = None,
+        midi_xml: Optional[torch.Tensor] = None,
+        midi_xml_lengths: Optional[torch.Tensor] = None,
+        tempo_lab: Optional[torch.Tensor] = None,
+        tempo_lab_lengths: Optional[torch.Tensor] = None,
+        tempo_xml: Optional[torch.Tensor] = None,
+        tempo_xml_lengths: Optional[torch.Tensor] = None,
+        beat_lab: Optional[torch.Tensor] = None,
+        beat_lab_lengths: Optional[torch.Tensor] = None,
[…]
