### 2022-10-04 1dbee0d
vits singing draft 2
Added midi, tempo, and beat embedding
