### 2022-10-04 934a923
add setup.py and sync pbs/slurm etc.
