### 2022-10-04 249c8e8
updata setup.sh
