### 2022-10-04 0649fee
apply black to new script
