### 2022-10-04 7508a55
Update egs2/TEMPLATE/st1/st.sh
Co-authored-by: Jiatong <728307998@qq.com>
