### 2022-10-05 05e7184
Update st.sh
