+++ _sources/_gen/espnet2_bin.rst.txt
@@ -679,6 +679,20 @@ asr_train.py
                                utterance_id_b b.wav
                                ...
                             
+                            "midi":
+                            MIDI format types which supported by sndfile mid, midi, etc.
+                            
+                               utterance_id_a a.mid
+                               utterance_id_b b.mid
+                               ...
+                            
+                            "duration":
+                            Return text as is. The text must be converted to ndarray by 'preprocess'.
+                            
+                               utterance_id_A start_1 end_1 phone_1 start_2 end_2 phone_2 ...
+                               utterance_id_B start_1 end_1 phone_1 start_2 end_2 phone_2 ...
+                               ...
+                            
                             "kaldi_ark":
                             Kaldi-ark file type.
                             
@@ -1303,6 +1317,20 @@ asr_transducer_train.py
                                utterance_id_b b.wav
                                ...
                             
+                            "midi":
+                            MIDI format types which supported by sndfile mid, midi, etc.
+                            
+                               utterance_id_a a.mid
+                               utterance_id_b b.mid
+                               ...
+                            
+                            "duration":
+                            Return text as is. The text must be converted to ndarray by 'preprocess'.
+                            
+                               utterance_id_A start_1 end_1 phone_1 start_2 end_2 phone_2 ...
+                               utterance_id_B start_1 end_1 phone_1 start_2 end_2 phone_2 ...
+                               ...
+                            
                             "kaldi_ark":
                             Kaldi-ark file type.
                             
@@ -1830,6 +1858,20 @@ diar_train.py
                             
                                utterance_id_a a.wav
                                utterance_id_b b.wav
+                               ...
+                            
+                            "midi":
+                            MIDI format types which supported by sndfile mid, midi, etc.
+                            
+                               utterance_id_a a.mid
+                               utterance_id_b b.mid
+                               ...
+                            
+                            "duration":
+                            Return text as is. The text must be converted to ndarray by 'preprocess'.
+                            
+                               utterance_id_A start_1 end_1 phone_1 start_2 end_2 phone_2 ...
+                               utterance_id_B start_1 end_1 phone_1 start_2 end_2 phone_2 ...
                                ...
                             
                             "kaldi_ark":
@@ -2426,6 +2468,20 @@ enh_s2t_train.py
                                utterance_id_b b.wav
                                ...
                             
+                            "midi":
+                            MIDI format types which supported by sndfile mid, midi, etc.
+                            
+                               utterance_id_a a.mid
+                               utterance_id_b b.mid
+                               ...
+                            
+                            "duration":
+                            Return text as is. The text must be converted to ndarray by 'preprocess'.
+                            
+                               utterance_id_A start_1 end_1 phone_1 start_2 end_2 phone_2 ...
+                               utterance_id_B start_1 end_1 phone_1 start_2 end_2 phone_2 ...
+                               ...
+                            
                             "kaldi_ark":
                             Kaldi-ark file type.
                             
@@ -3017,6 +3073,20 @@ enh_train.py
                                utterance_id_b b.wav
                                ...
                             
+                            "midi":
+                            MIDI format types which supported by sndfile mid, midi, etc.
+                            
+                               utterance_id_a a.mid
+                               utterance_id_b b.mid
+                               ...
+                            
+                            "duration":
+                            Return text as is. The text must be converted to ndarray by 'preprocess'.
+                            
+                               utterance_id_A start_1 end_1 phone_1 start_2 end_2 phone_2 ...
+                               utterance_id_B start_1 end_1 phone_1 start_2 end_2 phone_2 ...
+                               ...
+                            
                             "kaldi_ark":
[…]
