+++ egs2/ofuton_p_utagoe_db/svs1/run.sh
@@ -13,8 +13,8 @@ n_fft=2048
 n_shift=300
 win_length=1200
 
-score_feats_extract=frame_score_feats   # frame_score_feats | syllable_score_feats
-expdir=exp/vits
+score_feats_extract=syllable_score_feats   # frame_score_feats | syllable_score_feats
+expdir=exp
 
 opts=
 if [ "${fs}" -eq 48000 ]; then
