### 2022-10-05 c23e2b8
add gan_svs inference
note that dict is used instead of tuple, according to the latest vits model format, which may affect other svs models.
