+++ espnet2/bin/svs_inference.py
@@ -16,9 +16,11 @@ import torch
 from typeguard import check_argument_types
 
 from espnet2.fileio.npy_scp import NpyScpWriter
+from espnet2.gan_svs.vits import VITS
 from espnet2.tasks.svs import SVSTask
 from espnet2.torch_utils.device_funcs import to_device
 from espnet2.torch_utils.set_all_random_seed import set_all_random_seed
+from espnet2.tts.utils import DurationCalculator
 from espnet2.utils import config_argparse
 from espnet2.utils.types import str2bool, str2triple_str, str_or_none
 from espnet.utils.cli_utils import get_commandline_args
@@ -39,6 +41,9 @@ class SingingGenerate:
         train_config: Optional[Union[Path, str]],
         model_file: Optional[Union[Path, str]] = None,
         use_teacher_forcing: bool = False,
+        speed_control_alpha: float = 1.0,
+        noise_scale: float = 0.667,
+        noise_scale_dur: float = 0.8,
         vocoder_config: Union[Path, str] = None,
         vocoder_checkpoint: Union[Path, str] = None,
         dtype: str = "float32",
@@ -47,6 +52,7 @@ class SingingGenerate:
     ):
         assert check_argument_types()
 
+        # setup model
         model, train_args = SVSTask.build_model_from_file(
             train_config, model_file, device
         )
@@ -74,11 +80,20 @@ class SingingGenerate:
         logging.info(f"Extractor:\n{self.feats_extract}")
         logging.info(f"Normalizer:\n{self.normalize}")
         logging.info(f"SVS:\n{self.svs}")
+        if self.vocoder is not None:
+            logging.info(f"Vocoder:\n{self.vocoder}")
 
-        decode_config = {}
-        decode_config.update({"use_teacher_forcing": use_teacher_forcing})
-
-        self.decode_config = decode_config
+        # setup decoding config
+        decode_conf = {}
+        decode_conf.update({"use_teacher_forcing": use_teacher_forcing})
+        if isinstance(self.svs, (VITS)):
+            decode_conf.update(alpha=speed_control_alpha)
+        if isinstance(self.svs, VITS):
+            decode_conf.update(
+                noise_scale=noise_scale,
+                noise_scale_dur=noise_scale_dur,
+            )
+        self.decode_conf = decode_conf
 
     @torch.no_grad()
     def __call__(
@@ -110,9 +125,7 @@ class SingingGenerate:
         if self.use_spembs and spembs is None:
             raise RuntimeError("Missing required argument: 'spembs'")
 
-        batch = dict(
-            text=text,
-        )
+        batch = dict(text=text)
         if label_lab is not None:
             batch.update(label_lab=label_lab)
         if label_xml is not None:
@@ -141,31 +154,31 @@ class SingingGenerate:
             batch.update(lids=lids)
         batch = to_device(batch, self.device)
 
-        cfg = self.decode_config
+        cfg = self.decode_conf
         if decode_conf is not None:
             cfg = self.decode_conf.copy()
             cfg.update(decode_conf)
+        output_dict = self.model.inference(**batch, **cfg)
 
-        batch = to_device(batch, self.device)
-        outs, outs_denorm, probs, att_ws = self.model.inference(**batch, **cfg)
-
-        if att_ws is not None:
-            duration, focus_rate = self.duration_calculator(att_ws)
+        if output_dict.get("att_ws") is not None:
+            output_dict.update(duration=None, focus_rate=None)
+            # duration, focus_rate = self.duration_calculator(att_ws)
         else:
-            duration, focus_rate = None, None
+            output_dict.update(duration=None, focus_rate=None)
 
[…]
