### 2022-10-05 #4746 @Kevin-Mattheus-Moerman
@neillu23 did you see this comment :point_down: ?
> Hi @neillu23! Your paper is too long. Please read the [docs](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) for details about how to write your paper.

Can you work on changing the paper according to our guidelines? You can call `@editorialbot generate pdf` here to update the paper.
