### 2022-10-05 c02206d
Change duplicate code to symbolic links
