+++ espnet2/gan_svs/wavenet/__init__.py
@@ -1,1 +1,0 @@
-from espnet2.gan_tts.wavenet.wavenet import WaveNet  # NOQA
