+++ espnet2/gan_svs/jets/jets.py
@@ -1,651 +1,0 @@
-# Copyright 2022 Dan Lim
-#  Apache 2.0  (http://www.apache.org/licenses/LICENSE-2.0)
-
-"""JETS module for GAN-TTS task."""
-
-from typing import Any, Dict, Optional
-
-import torch
-from typeguard import check_argument_types
-
-from espnet2.gan_tts.abs_gan_tts import AbsGANTTS
-from espnet2.gan_tts.hifigan import (
-    HiFiGANMultiPeriodDiscriminator,
-    HiFiGANMultiScaleDiscriminator,
-    HiFiGANMultiScaleMultiPeriodDiscriminator,
-    HiFiGANPeriodDiscriminator,
-    HiFiGANScaleDiscriminator,
-)
-from espnet2.gan_tts.hifigan.loss import (
-    DiscriminatorAdversarialLoss,
-    FeatureMatchLoss,
-    GeneratorAdversarialLoss,
-    MelSpectrogramLoss,
-)
-from espnet2.gan_tts.jets.generator import JETSGenerator
-from espnet2.gan_tts.jets.loss import ForwardSumLoss, VarianceLoss
-from espnet2.gan_tts.utils import get_segments
-from espnet2.torch_utils.device_funcs import force_gatherable
-
-AVAILABLE_GENERATERS = {
-    "jets_generator": JETSGenerator,
-}
-AVAILABLE_DISCRIMINATORS = {
-    "hifigan_period_discriminator": HiFiGANPeriodDiscriminator,
-    "hifigan_scale_discriminator": HiFiGANScaleDiscriminator,
-    "hifigan_multi_period_discriminator": HiFiGANMultiPeriodDiscriminator,
-    "hifigan_multi_scale_discriminator": HiFiGANMultiScaleDiscriminator,
-    "hifigan_multi_scale_multi_period_discriminator": HiFiGANMultiScaleMultiPeriodDiscriminator,  # NOQA
-}
-
-
-class JETS(AbsGANTTS):
-    """JETS module (generator + discriminator).
-
-    This is a module of JETS described in `JETS: Jointly Training FastSpeech2
-    and HiFi-GAN for End to End Text to Speech'_.
-
-    .. _`JETS: Jointly Training FastSpeech2 and HiFi-GAN for End to End Text to Speech`
-        : https://arxiv.org/abs/2203.16852
-
-    """
-
-    def __init__(
-        self,
-        # generator related
-        idim: int,
-        odim: int,
-        sampling_rate: int = 22050,
-        generator_type: str = "jets_generator",
-        generator_params: Dict[str, Any] = {
-            "adim": 256,
-            "aheads": 2,
-            "elayers": 4,
-            "eunits": 1024,
-            "dlayers": 4,
-            "dunits": 1024,
-            "positionwise_layer_type": "conv1d",
-            "positionwise_conv_kernel_size": 1,
-            "use_scaled_pos_enc": True,
-            "use_batch_norm": True,
-            "encoder_normalize_before": True,
-            "decoder_normalize_before": True,
-            "encoder_concat_after": False,
-            "decoder_concat_after": False,
-            "reduction_factor": 1,
-            "encoder_type": "transformer",
-            "decoder_type": "transformer",
-            "transformer_enc_dropout_rate": 0.1,
-            "transformer_enc_positional_dropout_rate": 0.1,
-            "transformer_enc_attn_dropout_rate": 0.1,
-            "transformer_dec_dropout_rate": 0.1,
-            "transformer_dec_positional_dropout_rate": 0.1,
-            "transformer_dec_attn_dropout_rate": 0.1,
-            "conformer_rel_pos_type": "latest",
-            "conformer_pos_enc_layer_type": "rel_pos",
[…]
