+++ espnet2/gan_svs/joint/__init__.py
@@ -1,1 +1,0 @@
-from espnet2.gan_tts.joint.joint_text2wav import JointText2Wav  # NOQA
