+++ espnet2/gan_svs/hifigan/hifigan.py
@@ -1,745 +1,0 @@
-# Copyright 2021 Tomoki Hayashi
-#  Apache 2.0  (http://www.apache.org/licenses/LICENSE-2.0)
-
-"""HiFi-GAN Modules.
-
-This code is modified from https://github.com/kan-bayashi/ParallelWaveGAN.
-
-"""
-
-import copy
-import logging
-from typing import Any, Dict, List, Optional
-
-import numpy as np
-import torch
-import torch.nn.functional as F
-
-from espnet2.gan_tts.hifigan.residual_block import ResidualBlock
-
-
-class HiFiGANGenerator(torch.nn.Module):
-    """HiFiGAN generator module."""
-
-    def __init__(
-        self,
-        in_channels: int = 80,
-        out_channels: int = 1,
-        channels: int = 512,
-        global_channels: int = -1,
-        kernel_size: int = 7,
-        upsample_scales: List[int] = [8, 8, 2, 2],
-        upsample_kernel_sizes: List[int] = [16, 16, 4, 4],
-        resblock_kernel_sizes: List[int] = [3, 7, 11],
-        resblock_dilations: List[List[int]] = [[1, 3, 5], [1, 3, 5], [1, 3, 5]],
-        use_additional_convs: bool = True,
-        bias: bool = True,
-        nonlinear_activation: str = "LeakyReLU",
-        nonlinear_activation_params: Dict[str, Any] = {"negative_slope": 0.1},
-        use_weight_norm: bool = True,
-    ):
-        """Initialize HiFiGANGenerator module.
-
-        Args:
-            in_channels (int): Number of input channels.
-            out_channels (int): Number of output channels.
-            channels (int): Number of hidden representation channels.
-            global_channels (int): Number of global conditioning channels.
-            kernel_size (int): Kernel size of initial and final conv layer.
-            upsample_scales (List[int]): List of upsampling scales.
-            upsample_kernel_sizes (List[int]): List of kernel sizes for upsample layers.
-            resblock_kernel_sizes (List[int]): List of kernel sizes for residual blocks.
-            resblock_dilations (List[List[int]]): List of list of dilations for residual
-                blocks.
-            use_additional_convs (bool): Whether to use additional conv layers in
-                residual blocks.
-            bias (bool): Whether to add bias parameter in convolution layers.
-            nonlinear_activation (str): Activation function module name.
-            nonlinear_activation_params (Dict[str, Any]): Hyperparameters for activation
-                function.
-            use_weight_norm (bool): Whether to use weight norm. If set to true, it will
-                be applied to all of the conv layers.
-
-        """
-        super().__init__()
-
-        # check hyperparameters are valid
-        assert kernel_size % 2 == 1, "Kernel size must be odd number."
-        assert len(upsample_scales) == len(upsample_kernel_sizes)
-        assert len(resblock_dilations) == len(resblock_kernel_sizes)
-
-        # define modules
-        self.upsample_factor = int(np.prod(upsample_scales) * out_channels)
-        self.num_upsamples = len(upsample_kernel_sizes)
-        self.num_blocks = len(resblock_kernel_sizes)
-        self.input_conv = torch.nn.Conv1d(
-            in_channels,
-            channels,
-            kernel_size,
-            1,
-            padding=(kernel_size - 1) // 2,
-        )
-        self.upsamples = torch.nn.ModuleList()
-        self.blocks = torch.nn.ModuleList()
-        for i in range(len(upsample_kernel_sizes)):
-            assert upsample_kernel_sizes[i] == 2 * upsample_scales[i]
-            self.upsamples += [
-                torch.nn.Sequential(
-                    getattr(torch.nn, nonlinear_activation)(
-                        **nonlinear_activation_params
-                    ),
-                    torch.nn.ConvTranspose1d(
[…]
