+++ espnet2/gan_svs/melgan/__init__.py
@@ -1,3 +1,0 @@
-from espnet2.gan_tts.melgan.melgan import MelGANDiscriminator  # NOQA
-from espnet2.gan_tts.melgan.melgan import MelGANGenerator  # NOQA
-from espnet2.gan_tts.melgan.melgan import MelGANMultiScaleDiscriminator  # NOQA
