+++ espnet2/gan_svs/parallel_wavegan/__init__.py
@@ -1,6 +1,0 @@
-from espnet2.gan_tts.parallel_wavegan.parallel_wavegan import (
-    ParallelWaveGANDiscriminator,
-    ParallelWaveGANGenerator,
-)
-
-__all__ = ["ParallelWaveGANDiscriminator", "ParallelWaveGANGenerator"]
