### 2022-10-05 #4746 @neillu23
> @neillu23 did you see this comment 👇 ?
>
> > Hi @neillu23! Your paper is too long. Please read the [docs](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) for details about how to write your paper.
>
> Can you work on changing the paper according to our guidelines? You can call `@editorialbot generate pdf` here to update the paper.

Hi @Kevin-Mattheus-Moerman, thank you for your response! We are currently working on reducing the word count of the paper. We will further update the information here.
