### 2022-10-05 161ab4c
remove setup.sh
