+++ egs2/aphasiabank/asr1/local/split_train_test_val.py
@@ -1,0 +1,789 @@
+import os
+from argparse import ArgumentParser
+
+import numpy as np
+
+
+def get_args():
+    parser = ArgumentParser()
+    parser.add_argument("--text", type=str, help="Path to text", required=True)
+    parser.add_argument("--out-dir", type=str, required=True)
+    return parser.parse_args()
+
+
+# 56% train, 19% val, 25% test
+
+train_spks = [
+    # mild: 113
+    "kurland12a",
+    "scale05c",
+    "kurland05a",
+    "scale17a",
+    "fridriksson02a",
+    "kurland09b",
+    "tap04a",
+    "adler25a",
+    "elman15a",
+    "MMA12a",
+    "MSU02b",
+    "kurland04b",
+    "kurland07b",
+    "elman01a",
+    "kurland02c",
+    "kurland03b",
+    "tucson16a",
+    "wozniak05a",
+    "fridriksson09a",
+    "williamson17a",
+    "adler20a",
+    "fridriksson05a",
+    "ACWT04a",
+    "adler14a",
+    "williamson15a",
+    "wozniak04a",
+    "kurland04a",
+    "scale35a",
+    "scale21a",
+    "MSU02a",
+    "star03a",
+    "fridriksson07a",
+    "whiteside18a",
+    "tucson06a",
+    "kansas19a",
+    "adler01a",
+    "kurland08a",
+    "kurland14a",
+    "scale06b",
+    "kurland17b",
+    "MMA05a",
+    "wozniak01a",
+    "UNH03a",
+    "kansas15a",
+    "williamson09a",
+    "tcu10b",
+    "ACWT07a",
+    "BU01a",
+    "adler24a",
+    "adler07a",
+    "williamson08a",
+    "MMA09a",
+    "scale12b",
+    "kansas11a",
+    "tucson06b",
+    "kurland10b",
+    "kurland02d",
+    "tap01a",
+    "adler08a",
+    "UNH01a",
+    "kurland14b",
+    "kurland25a",
+    "kurland23c",
+    "BU05a",
+    "whiteside06a",
+    "scale06a",
+    "fridriksson01a",
+    "elman05a",
+    "MMA02a",
+    "UNH11a",
+    "elman04a",
+    "MMA03a",
+    "elman13a",
+    "williamson05a",
+    "whiteside09a",
+    "kansas21a",
+    "UNH01c",
+    "scale06c",
+    "UNH05a",
+    "kempler02a",
+    "scale22a",
+    "kansas04a",
+    "tucson18a",
+    "elman10a",
+    "whiteside20a",
+    "BU04a",
+    "kurland12b",
+    "tucson10a",
+    "williamson14a",
+    "whiteside17a",
+    "wright202a",
+    "tcu01a",
+    "whiteside19a",
+    "cmu03a",
+    "williamson18a",
+    "scale34a",
+    "kurland08b",
+    "kurland17c",
+    "thompson02a",
+    "scale18d",
+    "fridriksson09b",
+    "fridriksson04a",
+    "tucson20a",
+    "kurland26b",
+    "williamson10a",
+    "wozniak06a",
+    "williamson07a",
+    "kurland28c",
+    "whiteside13a",
+    "MSU01a",
+    "thompson14a",
+    "adler12a",
+    "thompson01a",
+    # moderate: 83
+    "MSU08b",
+    "ACWT05a",
+    "whiteside02a",
+    "UNH08a",
+    "kempler04a",
+    "fridriksson10b",
+    "thompson03a",
+    "tcu03a",
+    "tucson15b",
+    "UNH10a",
+    "williamson03a",
+    "kansas22a",
+    "scale26a",
+    "cmu02b",
+    "fridriksson06b",
+    "BU09a",
+    "garrett01a",
+    "scale15b",
[…]
