### 2022-10-06 a623908
add zero pad for convolution (ref: icefall)
