### 2022-10-06 c4ef9cf
fix aux. lm loss, reporter variable and doc
