+++ espnet2/gan_svs/vits/loss.py
@@ -1,1 +1,0 @@
-../../gan_tts/vits/loss.py
