+++ espnet2/gan_svs/vits/posterior_encoder.py
@@ -1,1 +1,0 @@
-../../gan_tts/vits/posterior_encoder.py
