+++ espnet2/gan_svs/vits/transform.py
@@ -1,1 +1,0 @@
-../../gan_tts/vits/transform.py
