+++ espnet2/gan_svs/vits/residual_coupling.py
@@ -1,1 +1,0 @@
-../../gan_tts/vits/residual_coupling.py
