### 2022-10-07 0cdbe42
add whisper frontend, testing on libris_100
