### 2022-10-07 8a4c5ea
add streaming transducer recipe
