### 2022-10-08 e7c88dd
fix run.sh and change embedding dim
