### 2022-10-08 f39eddc
remove duplicate scripts
