### 2022-10-08 bdc09d3
fix training bug with syllable_score_feats
