+++ espnet2/fileio/xml_scp.py
@@ -1,5 +1,7 @@
 import collections.abc
 import logging
+from pathlib import Path
+from typing import Union
 
 import numpy as np
 from typeguard import check_argument_types
@@ -12,9 +14,16 @@ except ImportError or ModuleNotFoundError:
     m21 = None
 
 
-class XMLScpReader(collections.abc.Mapping):
-    # TODO(Yuning): It is designed for sub xmls after segmentation.
+class NOTE(object):
+    def __init__(self, lyric, pitch, st, et, dur):
+        self.lyric = lyric
+        self.pitch = pitch
+        self.st = st
+        self.et = et
+        self.dur = dur
 
+
+class XMLScpReader(collections.abc.Mapping):
     """Reader class for 'xml.scp'.
 
     Examples:
@@ -45,39 +54,55 @@ class XMLScpReader(collections.abc.Mapping):
 
     def __getitem__(self, key):
         score = m21.converter.parse(self.data[key])
+        m = score.metronomeMarkBoundaries()
+        tempo = m[0][2].number
         part = score.parts[0].flat
         notes_list = []
         prepitch = None
+        st = 0
         for note in part.notesAndRests:
             if note.isNote:
                 if note.lyric is not None:
-                    notes_list.append(note)
+                    notes_list.append(
+                        NOTE(
+                            note.lyric,
+                            note.pitch.midi,
+                            st,
+                            st + note.seconds,
+                            note.seconds,
+                        )
+                    )
                 else:
                     if note.pitch == prepitch:
-                        notes_list[-1].seconds += note.seconds
+                        notes_list[-1].dur += note.seconds
+                        notes_list[-1].et += note.seconds
                     else:
-                        # TODO(Yuning): If there is more than one pitch in a syllable.
-                        # Doesn't happen in Ofuton
-                        logging.error("More than one pitch in one syllable")
+                        notes_list.append(
+                            NOTE(
+                                "", note.pitch.midi, st, st + note.seconds, note.seconds
+                            )
+                        )
                 prepitch = note.pitch
             else:
-                break
-        m = score.metronomeMarkBoundaries()
-        tempo = m[0][2].number
+                if prepitch == 0:
+                    notes_list[-1].dur += note.seconds
+                    notes_list[-1].et += note.seconds
+                else:
+                    notes_list.append(NOTE("P", 0, st, st + note.seconds, note.seconds))
+                prepitch = 0
+            st += note.seconds
         seq_len = len(notes_list)
+        # NOTE: xmls segments should not ends with rest
+        if notes_list[-1].pitch == 0:
+            seq_len -= 1
         lyrics = []
         notes = np.zeros(seq_len, dtype=self.dtype)
         segs = np.zeros((seq_len, 2))
-        t = 0
         for i in range(seq_len):
-            st = t
-            et = st + notes_list[i].seconds
-            segs[i, 0] = np.float32(st)
-            segs[i, 1] = np.float32(et)
-            t += notes_list[i].seconds
-            notes[i] = notes_list[i].pitch.midi
+            segs[i, 0] = np.float32(notes_list[i].st)
+            segs[i, 1] = np.float32(notes_list[i].et)
+            notes[i] = notes_list[i].pitch
             lyrics.append(notes_list[i].lyric)
-
         return lyrics, notes, segs, tempo
 
     def get_path(self, key):
@@ -96,7 +121,71 @@ class XMLScpReader(collections.abc.Mapping):
         return self.data.keys()
[…]
