+++ egs2/kiritan/svs1/local/extract_from_midi.py
@@ -1,4 +1,3 @@
-
 """
 midi_obj = miditoolkit.midi.parser.MidiFile(midi_path)
 
