+++ espnet2/svs/espnet_model.py
@@ -425,7 +425,8 @@ class ESPnetSVSModel(AbsESPnetModel):
                                     or _findMidi != midi_lab_after[i][index]
                                 )
                             ):
-                                # If duration is too short, add missing phone into label_lab sequence
+                                # If duration is too short,
+                                # add missing phone into label_lab sequence
                                 label_lab_lengths_after[i] += 1
                                 midi_lab_lengths_after[i] += 1
                                 tempo_lab_lengths_after[i] += 1
@@ -1036,7 +1037,8 @@ class ESPnetSVSModel(AbsESPnetModel):
                                 or _findMidi != midi_lab_after[i][index]
                             )
                         ):
-                            # If duration is too short, add missing phone into label_lab sequence
+                            # If duration is too short, 
+                            # add missing phone into label_lab sequence
                             label_lab_lengths_after[i] += 1
                             midi_lab_lengths_after[i] += 1
                             tempo_lab_lengths_after[i] += 1
