+++ espnet2/fileio/xml_scp.py
@@ -1,5 +1,4 @@
 import collections.abc
-import logging
 from pathlib import Path
 from typing import Union
 
