+++ egs2/kiritan/svs1/local/extract_from_midi.py
@@ -12,11 +12,12 @@ tick_to_time = midi_obj.get_tick_to_time_mapping()
 notes = midi_obj.instruments[0].notes
 notes.sort(key=lambda x: (x.start, x.pitch))
 for note in notes:
-	st = tick_to_time[note["start"]]
-	ed = tick_to_time[note["end"]]
+    st = tick_to_time[note["start"]]
+    ed = tick_to_time[note["end"]]
     seg = [st, et]
-	pitch = note.pitch
+    pitch = note.pitch
 
 # rest note need to be added, pitch = 0
-# rest note's seg can be calculated accordind to the breaks between the above notes sequence
+# rest note's seg can be calculated
+# accordind to the breaks between the above notes sequence
 """
