+++ egs2/m_ailabs/tts1/README.md
@@ -1,0 +1,16 @@
+# M-AILABS RECIPE
+
+This is the recipe of an en_US single-speaker TTS model with [M-AILABS](https://www.caito.de/2019/01/03/the-m-ailabs-speech-dataset/) dataset.  
+
+See the following pages for the usage:
+- [How to run the recipe](../../TEMPLATE/tts1/README.md#how-to-run)
+- [How to train FastSpeech](../../TEMPLATE/tts1/README.md#fastspeech-training)
+- [How to train FastSpeech2](../../TEMPLATE/tts1/README.md#fastspeech2-training)
+- [How to train with X-vector](../../TEMPLATE/tts1/README.md#multi-speaker-model-with-x-vector-training)
+- [How to train with speaker ID](../../TEMPLATE/tts1/README.md#multi-speaker-model-with-speaker-id-embedding-training)
+- [How to train VITS](../../TEMPLATE/tts1/README.md#vits-training)
+- [How to train joint text2wav](../../TEMPLATE/tts1/README.md#joint-text2wav-training)
+
+See the following pages before asking the question:
+- [ESPnet2 Tutorial](https://espnet.github.io/espnet/espnet2_tutorial.html)
+- [ESPnet2 TTS FAQ](../../TEMPLATE/tts1/README.md#faq)
