+++ espnet2/tasks/svs.py
@@ -263,7 +263,8 @@ class SVSTask(AbsTask):
                 text_cleaner=args.cleaner,
                 g2p_type=args.g2p,
                 fs=args.fs,
-                time_shift=args.feats_extract_conf["hop_length"] / args.feats_extract_conf["fs"]
+                time_shift=args.feats_extract_conf["hop_length"]
+                / args.feats_extract_conf["fs"],
             )
         else:
             retval = None
