+++ egs2/m_ailabs/tts1/local/data_multilingual.sh
@@ -18,21 +18,34 @@ if [ $# -ne 0 ]; then
     exit 2
 fi
 
+stage=-1
+stop_stage=2
+
 . ./path.sh || exit 1
 . ./cmd.sh || exit 1
 . ./db.sh || exit 1
 
+if [ -z "${M_AILABS}" ]; then
+    log "Fill the value of 'JSUT' of db.sh"
+    exit 1
+fi
+db_root=${M_AILABS}
+
 # silence part trimming related
-do_trimming=false
+do_trimming=true
+nj=32
 
-# token type ("char" or "phn")
 token_type=phn
-
-# dataset configuration
-db_root=/home/saeki/workspace/ssd1/mailab
 
 # locale all
 langs=("de_DE" "en_UK" "it_IT" "es_ES" "en_US" "fr_FR" "uk_UK" "ru_RU" "pl_PL")
+
+if [ ${stage} -le -1 ] && [ ${stop_stage} -ge -1 ]; then
+    log "stage -1: Downloading data"
+    for lang in ${langs[@]}; do
+        local/download.sh ${db_root} ${lang}
+    done
+fi
 
 whole_set_all=""
 train_set_all=""
@@ -44,112 +57,117 @@ if [ "${token_type}" = phn ]; then
     suffix="_phn"
 fi
 
-for lang in ${langs[@]}; do
+if [ ${stage} -le 1 ] && [ ${stop_stage} -ge 1 ]; then
+    log "stage 0-1: Data preparation and division to subsets"
 
-    # Listing all the speakers for each locale
-    if [ ${lang} = "de_DE" ]; then
-        spkrs=("angela" "rebecca" "ramona" "eva" "karlsson")
-        g2p="espeak_ng_german"
-    elif [ ${lang} = "en_UK" ]; then
-        spkrs=("elizabeth")
-        g2p="g2p_en"
-    elif [ ${lang} = "it_IT" ]; then
-        spkrs=("lisa" "riccardo")
-        g2p="espeak_ng_italian"
-    elif [ ${lang} = "es_ES" ]; then
-        spkrs=("karen" "tux" "victor")
-        g2p="espeak_ng_spanish"
-    elif [ ${lang} = "en_US" ]; then
-        spkrs=("judy" "mary" "elliot")
-        g2p="g2p_en"
-    elif [ ${lang} = "fr_FR" ]; then
-        spkrs=("ezwa" "nadine" "bernard" "gilles" "zeckou")
-        g2p="espeak_ng_french"
-    elif [ ${lang} = "uk_UK" ]; then
-        spkrs=("sumska" "loboda" "miskun" "obruchov" "shepel")
-        # Using russian g2p because phonemizer does not support uk
-        g2p="espeak_ng_russian"
-    elif [ ${lang} = "ru_RU" ]; then
-        spkrs=("hajdurova" "minaev" "nikolaev")
-        g2p="espeak_ng_russian"
-    elif [ ${lang} = "pl_PL" ]; then
-        spkrs=("nina" "piotr")
-        g2p="espeak_ng_polish"
-    else
-        echo "${lang} is not supported."
-        exit 1
-    fi
+    for lang in ${langs[@]}; do
 
-    # Preparing the dataset for each locale and speaker
-    for spkr in ${spkrs[@]}; do
-
-        # org set, and train, dev, eval sets with suffix
-        org_set=${lang}_${spkr}
-        whole_set="${org_set}${suffix}"
-        train_set="${whole_set}_train"
-        dev_set="${whole_set}_dev"
-        eval_set="${whole_set}_eval"
-        # Appending each data dir to combine them later
-        whole_set_all="${whole_set_all} data/${whole_set}"
-        train_set_all="${train_set_all} data/${train_set}"
[…]
