+++ egs2/m_ailabs/tts1/run.sh
@@ -17,23 +17,54 @@ else
     opts="--audio_format flac "
 fi
 
-train_set=tr_no_dev
-valid_set=dev
-test_sets="dev eval"
+# Data prep related
+train_type=en_us # Use "en_us" or "multilingual".
+token_type=phn
 
-train_config=conf/train.yaml
+suffix=""
+# Training the model only on the en_us single-speaker data
+if [ "${train_type}" = en_us ]; then
+    lang=en
+    cleaner=tacotron
+    g2p=g2p_en
+    use_lid=false
+    use_sid=false
+    train_config=conf/train.yaml
+elif [ "${train_type}" = multilingual ]; then
+    # Training the model on the whole dataset
+    lang=noinfo
+    cleaner=none
+    g2p=none
+    use_lid=true
+    use_sid=true
+    if [ "${token_type}" = phn ]; then
+        suffix="_phn"
+    fi
+    train_config=conf/train_multilingual.yaml
+else
+    log train_type: "${train_type} is not supported."
+    exit 1
+fi
+
+local_data_opts+=" --train_type ${train_type}"
+
+train_set=tr_no_dev${suffix}
+valid_set=dev${suffix}
+test_sets="dev${suffix} eval${suffix}"
+
 inference_config=conf/decode.yaml
 
-g2p=g2p_en
-
 ./tts.sh \
-    --lang en \
+    --lang ${lang} \
+    --local_data_opts "${local_data_opts}" \
     --feats_type raw \
+    --use_lid ${use_lid} \
+    --use_sid ${use_sid} \
     --fs "${fs}" \
     --n_fft "${n_fft}" \
     --n_shift "${n_shift}" \
-    --token_type phn \
-    --cleaner tacotron \
+    --token_type "${token_type}" \
+    --cleaner "${cleaner}" \
     --g2p "${g2p}" \
     --train_config "${train_config}" \
     --inference_config "${inference_config}" \
