+++ espnet2/svs/feats_extract/score_feats_extract.py
@@ -102,7 +102,14 @@ class FrameScoreFeats(AbsFeatsExtract):
                 pad = self.win_length // 2
                 input_lengths = input_lengths + 2 * pad
 
-            olens = (input_lengths - self.win_length) // self.hop_length + 1
+            olens = (
+                torch.div(
+                    (input_lengths - self.win_length),
+                    self.hop_length,
+                    rounding_mode="floor",
+                )
+                + 1
+            )
             output.masked_fill_(make_pad_mask(olens, output, 1), 0.0)
         else:
             olens = None
