+++ espnet2/layers/stft.py
@@ -161,10 +161,7 @@ class Stft(torch.nn.Module, InversibleInterface):
                 pad = self.n_fft // 2
                 ilens = ilens + 2 * pad
 
-            olens = (
-                torch.div((ilens - self.n_fft), self.hop_length, rounding_mode="floor")
-                + 1
-            )
+            olens = torch.div((ilens - self.n_fft), self.hop_length) + 1
             output.masked_fill_(make_pad_mask(olens, output, 1), 0.0)
         else:
             olens = None
