### 2022-10-10 749f759
fix typo on config
