### 2022-10-10 1154f1b
fix typo in config
