### 2022-10-11 ba74263
fix rnn inference
