### 2022-10-11 0d0b9d6
fix rnn inference
