### 2022-10-11 93d11a7
add scripts for multi-iris
