### 2022-10-11 7de4cc8
add OCR recipe for IAM handwriting dataset
