+++ egs2/iam/ocr1/local/data_prep.py
@@ -1,4 +1,20 @@
-import string
+"""Prepare the IAM handwriting dataset for ESPnet ASR training
+
+Usage:
+    python local/data_prep.py [--feature_dim 100] [--downsampling_factor 0.5]
+
+Expects data to be in:
+    downloads/
+        lines.txt        # labels from IAM Handwriting dataset (from ascii.tgz)
+        lines/
+            */*/*.png    # "lines" images from IAM Handwriting dataset (from lines.tgz)
+        train.txt        # id's for train/valid/test splits
+        valid.txt
+        test.txt
+
+Required packages:
+    Pillow
+"""
 import os
 import argparse
 import numpy as np
@@ -7,7 +23,22 @@ from PIL import Image
 from espnet.utils.cli_writers import file_writer_helper
 
 def prepare_text(lines_file_path, output_dir, split_ids):
-    """Create text file (map of ids to transcriptions) in Kaldi format"""
+    """Create 'text' file (map of ids to transcriptions) in Kaldi format
+
+    Parameters
+    ----------
+    lines_file_path : str
+        The file path of the full "lines.txt" label file of the IAM dataset
+    output_dir : str
+        The folder path for output, e.g. data/
+    split_ids : list of str
+        a list of example ids to process, used to define train/valid/test splits
+
+    Returns
+    -------
+    skipped_ids : list of str
+        a list of ids that were skipped due to having an empty transcription
+    """
     output_lines = []
     skipped_ids = []
     with open(lines_file_path) as lines_file:
@@ -27,9 +58,6 @@ def prepare_text(lines_file_path, output_dir, split_ids):
                 # extract and format transcription into Kaldi style
                 transcription = " ".join(line_split[8:])
                 transcription = transcription.replace("|", " ")
-                # transcription = transcription.translate(str.maketrans('', '', string.punctuation))
-                # transcription = transcription.replace("  ", " ")
-                # transcription = transcription.strip().upper()
                 transcription = transcription.strip()
 
                 if transcription == "":
@@ -47,7 +75,18 @@ def prepare_text(lines_file_path, output_dir, split_ids):
     return skipped_ids
 
 def prepare_utt2spk_spk2utt(output_dir, split_ids, ids_to_skip=[]):
-    """Create (dummy) utt2spk and spk2utt files to satisfy Kaldi format"""
+    """Create (dummy) utt2spk and spk2utt files to satisfy Kaldi format
+
+    Parameters
+    ----------
+    output_dir : str
+        The folder path for output, e.g. data/
+    split_ids : list of str
+        a list of example ids to process, used to define train/valid/test splits
+    ids_to_skip : list of str
+        A list of ids to exclude from the output, e.g. used to ignore the ones
+        that were skipped due to an empty transcription
+    """
     output_lines = [f'{line_id} {line_id}\n' for line_id in split_ids if line_id not in ids_to_skip]
     output_lines.sort()
 
@@ -57,9 +96,25 @@ def prepare_utt2spk_spk2utt(output_dir, split_ids, ids_to_skip=[]):
     with open(os.path.join(output_dir, 'spk2utt'), 'w') as out_file:
         out_file.writelines(output_lines)
 
-def prepare_feats(img_dir, output_dir, split_ids, ids_to_skip=[], feature_dim=100, downsampling_factor=0.25):
-    """Create feats.scp file from OCR images"""
+def prepare_feats(img_dir, output_dir, split_ids, ids_to_skip=[], feature_dim=100, downsampling_factor=0.5):
+    """Create feats.scp file from OCR images
 
+    Parameters
+    ----------
+    output_dir : str
+        The folder path for output, e.g. data/
+    split_ids : list of str
+        a list of example ids to process, used to define train/valid/test splits
[…]
