### 2022-10-12 1e4e9ef
refactor data_prep.sh
