### 2022-10-13 4a3527c
Fix CI errors; add unit tests to check consistency between old and new beamformer implementations
