+++ espnet2/enh/layers/dnn_beamformer.py
@@ -7,45 +7,14 @@ from packaging.version import parse as V
 from torch.nn import functional as F
 from torch_complex.tensor import ComplexTensor
 
+import espnet2.enh.layers.beamformer as bf_v1
+import espnet2.enh.layers.beamformer_th as bf_v2
 from espnet2.enh.layers.complex_utils import stack, to_double, to_float
 from espnet2.enh.layers.mask_estimator import MaskEstimator
 
 is_torch_1_9_plus = V(torch.__version__) >= V("1.9.0")
 is_torch_1_12_1_plus = V(torch.__version__) >= V("1.12.1")
-if is_torch_1_12_1_plus:
-    from espnet2.enh.layers.beamformer_th import (
-        apply_beamforming_vector,
-        blind_analytic_normalization,
-        get_gev_vector,
-        get_lcmv_vector_with_rtf,
-        get_mvdr_vector,
-        get_mvdr_vector_with_rtf,
-        get_mwf_vector,
-        get_rank1_mwf_vector,
-        get_rtf_matrix,
-        get_sdw_mwf_vector,
-        get_WPD_filter_v2,
-        get_WPD_filter_with_rtf,
-        perform_WPD_filtering,
-        prepare_beamformer_stats,
-    )
-else:
-    from espnet2.enh.layers.beamformer import (
-        apply_beamforming_vector,
-        blind_analytic_normalization,
-        get_gev_vector,
-        get_lcmv_vector_with_rtf,
-        get_mvdr_vector,
-        get_mvdr_vector_with_rtf,
-        get_mwf_vector,
-        get_rank1_mwf_vector,
-        get_rtf_matrix,
-        get_sdw_mwf_vector,
-        get_WPD_filter_v2,
-        get_WPD_filter_with_rtf,
-        perform_WPD_filtering,
-        prepare_beamformer_stats,
-    )
+
 
 BEAMFORMER_TYPES = (
     # Minimum Variance Distortionless Response beamformer
@@ -113,6 +82,8 @@ class DNN_Beamformer(torch.nn.Module):
         mask_flooring: bool = False,
         flooring_thres: float = 1e-6,
         use_torch_solver: bool = True,
+        # False to use old APIs; True to use torchaudio-based new APIs
+        use_torchaudio_api: bool = False,
         # only for WPD beamformer
         btaps: int = 5,
         bdelay: int = 3,
@@ -186,6 +157,11 @@ class DNN_Beamformer(torch.nn.Module):
                 "Now it will always be true in DNN_Beamformer"
             )
 
+        if use_torchaudio_api and is_torch_1_12_1_plus:
+            self.bf_func = bf_v2
+        else:
+            self.bf_func = bf_v1
+
     def forward(
         self,
         data: Union[torch.Tensor, ComplexTensor],
@@ -237,7 +213,7 @@ class DNN_Beamformer(torch.nn.Module):
 
             if self.beamformer_type in ("lcmv", "lcmp", "wlcmp"):
                 raise NotImplementedError("Single source is not supported yet")
-            beamformer_stats = prepare_beamformer_stats(
+            beamformer_stats = self.bf_func.prepare_beamformer_stats(
                 data_d,
                 [mask_speech],
                 mask_noise,
@@ -287,7 +263,7 @@ class DNN_Beamformer(torch.nn.Module):
                 mask_speech = list(masks)
                 mask_noise = None
 
-            beamformer_stats = prepare_beamformer_stats(
+            beamformer_stats = self.bf_func.prepare_beamformer_stats(
                 data_d,
                 mask_speech,
                 mask_noise,
@@ -298,7 +274,7 @@ class DNN_Beamformer(torch.nn.Module):
[…]
