+++ test/espnet2/enh/separator/test_beamformer.py
@@ -1,3 +1,4 @@
+import numpy as np
 import pytest
 import torch
 from packaging.version import parse as V
@@ -157,6 +158,7 @@ def test_neural_beamformer_forward_backward(
         beamformer_type=beamformer_type,
         rtf_iterations=2,
         shared_power=True,
+        use_torchaudio_api=is_torch_1_12_1_plus,
     )
 
     model.train()
@@ -196,6 +198,7 @@ def test_neural_beamformer_wpe_output(
         taps=5,
         delay=3,
         use_beamformer=False,
+        use_torchaudio_api=is_torch_1_12_1_plus,
     )
     model.eval()
     input_spectrum, flens = stft(inputs, ilens)
@@ -263,6 +266,7 @@ def test_neural_beamformer_bf_output(
         diagonal_loading=diagonal_loading,
         mask_flooring=mask_flooring,
         use_torch_solver=use_torch_solver,
+        use_torchaudio_api=is_torch_1_12_1_plus,
     )
     model.eval()
     input_spectrum, flens = stft(inputs, ilens)
@@ -282,6 +286,61 @@ def test_neural_beamformer_bf_output(
             assert specs[n - 1].dtype == torch.complex64
         else:
             assert specs[n - 1].dtype == torch.float
+
+
+@pytest.mark.parametrize("num_spk", [1, 2])
+@pytest.mark.parametrize("use_noise_mask", [True, False])
+@pytest.mark.parametrize("beamformer_type", BEAMFORMER_TYPES)
+@pytest.mark.skipif(not is_torch_1_12_1_plus, reason="Only for torch>=1.12.1")
+def test_beamformer_net_consistency(num_spk, use_noise_mask, beamformer_type):
+    if beamformer_type in (
+        "lcmv",
+        "lcmp",
+        "wlcmp",
+        "mvdr_tfs",
+        "mvdr_tfs_souden",
+    ):
+        # skip these beamformers as they require real multi-speaker data to obtain
+        # consistent results
+        return
+
+    torch.random.manual_seed(0)
+    ch = 2
+    inputs = random_speech[..., :ch].float().repeat(1, 5, 1)
+    ilens = torch.LongTensor([80, 60])
+
+    stft1 = STFTEncoder(n_fft=8, hop_length=2, use_builtin_complex=True)
+    stft2 = STFTEncoder(n_fft=8, hop_length=2, use_builtin_complex=False)
+    args = dict(
+        num_spk=num_spk,
+        use_wpe=False,
+        taps=2,
+        delay=3,
+        use_beamformer=True,
+        blayers=2,
+        bunits=2,
+        bprojs=2,
+        badim=2,
+        use_noise_mask=use_noise_mask,
+        beamformer_type=beamformer_type,
+        diagonal_loading=True,
+        mask_flooring=True,
+        use_torch_solver=True,
+    )
+    model1 = NeuralBeamformer(stft1.output_dim, use_torchaudio_api=True, **args)
+    model2 = NeuralBeamformer(stft2.output_dim, use_torchaudio_api=False, **args)
+    model1.eval()
+    model2.eval()
+    models = [model1, model2]
+    input_spectrum1, flens = stft1(inputs, ilens)
+    input_spectrum2 = stft2(inputs, ilens)[0]
+    inputs = [input_spectrum1, input_spectrum2]
+
+    specs = [
+        model(input_spectrum, flens)[0] for model, input_spectrum in zip(models, inputs)
+    ]
+    for spk in range(num_spk):
+        np.testing.assert_allclose(*[s[spk].detach().numpy() for s in specs], atol=1e-1)
 
 
 def test_beamformer_net_invalid_bf_type():
