+++ utils/precompute-asr-feats.py
@@ -1,0 +1,510 @@
+#!/usr/bin/env python3
+
+# Copyright 2022 Nanyang Technological University (Kwok Chin Yuen)
+#  Apache 2.0  (http://www.apache.org/licenses/LICENSE-2.0)
+
+import os
+from pathlib import Path
+import argparse
+import logging
+from typing import Any, List, Optional, Sequence, Tuple, Union, Callable
+from typeguard import check_argument_types
+import re
+
+import kaldiio
+import numpy as np
+import resampy
+import torch
+
+from espnet2.torch_utils.set_all_random_seed import set_all_random_seed
+from espnet2.torch_utils.device_funcs import to_device
+from espnet2.tasks.asr import ASRTask
+from espnet2.utils.types import str2bool, str2triple_str, str_or_none
+from espnet.transform.spectrogram import logmelspectrogram
+from espnet.utils.cli_utils import get_commandline_args
+from espnet.utils.cli_writers import file_writer_helper
+
+
+def build_asr_model(
+        enh_s2t_task: bool = False,
+        quantize_asr_model: bool = False,
+        quantize_modules: List[str] = ["Linear"],
+        quantize_dtype: str = "qint8",
+        asr_train_config: Union[Path, str] = None,
+        asr_model_file: Union[Path, str] = None,
+        device: str = "cpu",
+        dtype: str = np.float32,
+    ):
+        task = ASRTask if not enh_s2t_task else EnhS2TTask
+
+        if quantize_asr_model:
+            if quantize_dtype == "float16" and torch.__version__ < LooseVersion(
+                "1.5.0"
+            ):
+                raise ValueError(
+                    "float16 dtype for dynamic quantization is not supported with "
+                    "torch version < 1.5.0. Switch to qint8 dtype instead."
+                )
+
+        quantize_modules = set([getattr(torch.nn, q) for q in quantize_modules])
+        quantize_dtype = getattr(torch, quantize_dtype)
+
+        # 1. Build ASR model
+        asr_model, asr_train_args = task.build_model_from_file(
+            asr_train_config, asr_model_file, device
+        )
+        if enh_s2t_task:
+            asr_model.inherite_attributes(
+                inherite_s2t_attrs=[
+                    "ctc",
+                    "decoder",
+                    "eos",
+                    "joint_network",
+                    "sos",
+                    "token_list",
+                    "use_transducer_decoder",
+                ]
+            )
+        asr_model.to(dtype=getattr(torch, dtype)).eval()
+
+        if quantize_asr_model:
+            logging.info("Use quantized asr model for decoding.")
+
+            asr_model = torch.quantization.quantize_dynamic(
+                asr_model, qconfig_spec=quantize_modules, dtype=quantize_dtype
+            )
+
+        decoder = asr_model.decoder
+
+        return asr_model, asr_train_args
+
+
+class Speech2Feat:
+
+    def __init__(
+        self,
+        data_path_and_name_and_type: str,
+        enh_s2t_task: bool = False,
+        quantize_asr_model: bool = False,
+        quantize_modules: List[str] = ["Linear"],
+        quantize_dtype: str = "qint8",
+        asr_train_config: Union[Path, str] = None,
+        asr_model_file: Union[Path, str] = None,
+        device: str = "cpu",
+        key_file: str = None,
+        batch_size: int = 1,
+        dtype: str = np.float32,
+        num_workers: int = 1,
+        allow_variable_data_keys: bool = False,
+        ngpu: int = 0,
+        inference: bool = False,
+        output_modules: List[str] = ["encoder.encoders.*"], 
+        skip_pad_modules: List[str] = [],
+    ):
+
+        # 1. Build ASR model
+        self.asr_model, self.asr_train_args = build_asr_model(
+            enh_s2t_task=enh_s2t_task,
[…]
