+++ egs2/README.md
@@ -97,7 +97,7 @@ See: https://espnet.github.io/espnet/espnet2_tutorial.html#recipes-using-espnet2
 | must_c_v2              | https://ict.fbk.eu/must-c/                                                                                                        | ASR/MT/ST               | ENG->DEU            | https://ict.fbk.eu/must-c/                                                                                    |              |
 | nsc                     | National Speech Corpus                                                                                                           | ASR                     | ENG-SG               | https://www.imda.gov.sg/programme-listing/digital-services-lab/national-speech-corpus                        |              |
 | ofuton_p_utagoe_db      | Ofuton_p_utagoe Singing voice synthesis corpus                                                                                   | SVS                     | JPN                  | https://sites.google.com/view/oftn-utagoedb/%E3%83%9B%E3%83%BC%E3%83%A0                                      |              |
-| oniku      | oniku_kurumi_utagoe_db corpus                                                                                   | SVS                     | JPN                  | https://github.com/oatsu-gh/ virtual_singer_oniku_kurumi                                      |              |
+| oniku      | oniku_kurumi_utagoe_db corpus                                                                                   | SVS                     | JPN                  | https://github.com/oatsu-gh/virtual_singer_oniku_kurumi                                      |              |
 | open_li110              | Corpus combination with 110 languages                                                                                            | Multilingual ASR        | 100+ languages       |                                                                                                              |              |
 | open_li52               | Corpus combination with 52 languages(Commonvocie + voxforge)                                                                     | Multilingual ASR        | 52 languages         |                                                                                                              |              |
 | polyphone_swiss_french  | Swiss French Polyphone corpus                                                                                                    | ASR                     | FRA                  | http://catalog.elra.info/en-us/repository/browse/ELRA-S0030_02                                               |              |
