### 2022-10-14 a339d8a
add configurations for pre-training
