+++ egs2/iam/ocr1/local/data_prep.py
@@ -22,6 +22,7 @@ from PIL import Image
 
 from espnet.utils.cli_writers import file_writer_helper
 
+
 def prepare_text(lines_file_path, output_dir, split_ids):
     """Create 'text' file (map of ids to transcriptions) in Kaldi format
 
@@ -45,7 +46,7 @@ def prepare_text(lines_file_path, output_dir, split_ids):
         for line in lines_file.readlines():
 
             # skip comment lines
-            if line[0] == '#':
+            if line[0] == "#":
                 continue
 
             line_split = line.strip().split()
@@ -61,7 +62,7 @@ def prepare_text(lines_file_path, output_dir, split_ids):
                 transcription = transcription.strip()
 
                 if transcription == "":
-                    print(f'Line {line_id} has an empty transcription, skipping it')
+                    print(f"Line {line_id} has an empty transcription, skipping it")
                     skipped_ids.append(line_id)
                     continue
 
@@ -69,10 +70,11 @@ def prepare_text(lines_file_path, output_dir, split_ids):
 
     output_lines.sort()
 
-    with open(os.path.join(output_dir, 'text'), 'w') as out_file:
+    with open(os.path.join(output_dir, "text"), "w") as out_file:
         out_file.writelines(output_lines)
 
     return skipped_ids
+
 
 def prepare_utt2spk_spk2utt(output_dir, split_ids, ids_to_skip=[]):
     """Create (dummy) utt2spk and spk2utt files to satisfy Kaldi format
@@ -87,16 +89,26 @@ def prepare_utt2spk_spk2utt(output_dir, split_ids, ids_to_skip=[]):
         A list of ids to exclude from the output, e.g. used to ignore the ones
         that were skipped due to an empty transcription
     """
-    output_lines = [f'{line_id} {line_id}\n' for line_id in split_ids if line_id not in ids_to_skip]
+    output_lines = [
+        f"{line_id} {line_id}\n" for line_id in split_ids if line_id not in ids_to_skip
+    ]
     output_lines.sort()
 
-    with open(os.path.join(output_dir, 'utt2spk'), 'w') as out_file:
+    with open(os.path.join(output_dir, "utt2spk"), "w") as out_file:
         out_file.writelines(output_lines)
 
-    with open(os.path.join(output_dir, 'spk2utt'), 'w') as out_file:
+    with open(os.path.join(output_dir, "spk2utt"), "w") as out_file:
         out_file.writelines(output_lines)
 
-def prepare_feats(img_dir, output_dir, split_ids, ids_to_skip=[], feature_dim=100, downsampling_factor=0.5):
+
+def prepare_feats(
+    img_dir,
+    output_dir,
+    split_ids,
+    ids_to_skip=[],
+    feature_dim=100,
+    downsampling_factor=0.5,
+):
     """Create feats.scp file from OCR images
 
     Parameters
@@ -117,9 +129,9 @@ def prepare_feats(img_dir, output_dir, split_ids, ids_to_skip=[], feature_dim=10
     """
     writer = file_writer_helper(
         wspecifier=f'ark,scp:{os.path.join(output_dir, "feats.ark")},{os.path.join(output_dir, "feats.scp")}',
-        filetype='mat',
-        write_num_frames=f'ark,t:{output_dir}/num_frames.txt',
-        compress=False
+        filetype="mat",
+        write_num_frames=f"ark,t:{output_dir}/num_frames.txt",
+        compress=False,
     )
 
     num_processed = 0
@@ -129,8 +141,10 @@ def prepare_feats(img_dir, output_dir, split_ids, ids_to_skip=[], feature_dim=10
         if img_id in ids_to_skip:
             continue
 
[…]
