### 2022-10-15 c479445
remove redundant combine_data.sh
