### 2022-10-15 e45c034
remove redundant asr combine_data.sh
