+++ egs2/puebla_nahuatl/asr1/local/data.sh
@@ -52,7 +52,8 @@ if [ ${stage} -le 2 ] && [ ${stop_stage} -ge 2 ]; then
     log "stage2: Preparing data for Puebla Nahuatl"
     ### Task dependent. You have to make data the following preparation part by yourself.
     for x in train dev "test"; do
-        python local/data_prep.py -w $wavdir -t data/${x} -m ${annotation_type} -i local/split/speaker_wav_mapping_nahuatl_${x}.csv -a ${annotation_dir}
+        python local/data_prep.py -w $wavdir -t data/${x} -m ${annotation_type} -i local/split/speaker_wav_mapping_nahuatl_${x}.csv -a ${annotation_dir} \
+            -d local/split/Puebla-Nahuat-and-Totonac-consultants_for-LDC-archive.xml
         utils/fix_data_dir.sh data/${x}
         chmod +x data/${x}/remix_script.sh
         mkdir -p remixed
