### 2022-10-18 a480525
Update conv_input.py
The subsampling_factor should be 6 instead of 4 as the two pooling strides along the time axis are 3 and 2.
