### 2022-10-18 #4746 @faroit
Hi @neillu23, thanks for your submission to JOSS! I will be serving as the editor for this. We'll use this issue to identify reviewers and resolve technical issues.

It would be great to have reviewers with domain knowledge who also have experience with speech frameworks... I have a few potential reviewers in mind:

- __@ jonashaag__
- __@ adefossez__
- __@ jordipons__

@neillu23, I am also interested in your input on potential reviewers, so any suggestions you have are also welcome.
For this, please check the JOSS [COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) and only mention them with a space between the @ and their handle (so you would refer to me as "@ faroit").
