+++ egs2/ofuton_p_utagoe_db/svs1/conf/tuning/train_vits.yaml
@@ -12,6 +12,7 @@
 ##########################################################
 svs: vits
 svs_conf:
+    use_visinger: true
     # generator related
     generator_type: vits_generator
     generator_params:

[…]
