+++ utils/precompute-asr-feats.py
@@ -5,6 +5,7 @@
 
 import argparse
 import logging
+import sys
 import os
 import re
 from pathlib import Path
@@ -14,8 +15,10 @@ import kaldiio
 import numpy as np
 import resampy
 import torch
+from packaging.version import parse as V
 from typeguard import check_argument_types
 
+from espnet2.utils import config_argparse
 from espnet2.tasks.asr import ASRTask
 from espnet2.torch_utils.device_funcs import to_device
 from espnet2.torch_utils.set_all_random_seed import set_all_random_seed
@@ -25,33 +28,36 @@ from espnet.utils.cli_utils import get_commandline_args
 from espnet.utils.cli_writers import file_writer_helper
 
 
-def build_asr_model(
-    enh_s2t_task: bool = False,
-    quantize_asr_model: bool = False,
-    quantize_modules: List[str] = ["Linear"],
-    quantize_dtype: str = "qint8",
-    asr_train_config: Union[Path, str] = None,
-    asr_model_file: Union[Path, str] = None,
-    device: str = "cpu",
-    dtype: str = np.float32,
-):
-    task = ASRTask if not enh_s2t_task else EnhS2TTask
+if V(torch.__version__) >= V("1.6.0"):
+    from torch.cuda.amp import autocast
+else:
+    # Nothing to do if torch<1.6.0
+    @contextmanager
+    def autocast(enabled=True):
+        yield
 
-    if quantize_asr_model:
-        if quantize_dtype == "float16" and torch.__version__ < LooseVersion("1.5.0"):
+
+# Build method from espnet2/bin/asr_inference.py
+def build_asr_model(
+    args: argparse.Namespace,
+):
+    task = ASRTask if not args.enh_s2t_task else EnhS2TTask
+
+    if args.quantize_asr_model:
+        if args.quantize_dtype == "float16" and torch.__version__ < LooseVersion("1.5.0"):
             raise ValueError(
                 "float16 dtype for dynamic quantization is not supported with "
                 "torch version < 1.5.0. Switch to qint8 dtype instead."
             )
 
-    quantize_modules = set([getattr(torch.nn, q) for q in quantize_modules])
-    quantize_dtype = getattr(torch, quantize_dtype)
+    quantize_modules = set([getattr(torch.nn, q) for q in args.quantize_modules])
+    quantize_dtype = getattr(torch, args.quantize_dtype)
 
     # 1. Build ASR model
     asr_model, asr_train_args = task.build_model_from_file(
-        asr_train_config, asr_model_file, device
+        args.asr_train_config, args.asr_model_file, args.device
     )
-    if enh_s2t_task:
+    if args.enh_s2t_task:
         asr_model.inherite_attributes(
             inherite_s2t_attrs=[
                 "ctc",
@@ -63,9 +69,9 @@ def build_asr_model(
                 "use_transducer_decoder",
             ]
         )
-    asr_model.to(dtype=getattr(torch, dtype)).eval()
+    asr_model.to(dtype=getattr(torch, args.dtype)).eval()
 
-    if quantize_asr_model:
+    if args.quantize_asr_model:
         logging.info("Use quantized asr model for decoding.")
 
         asr_model = torch.quantization.quantize_dynamic(
@@ -80,59 +86,49 @@ def build_asr_model(
 class Speech2Feat:
     def __init__(
         self,
-        data_path_and_name_and_type: str,
-        enh_s2t_task: bool = False,
-        quantize_asr_model: bool = False,
-        quantize_modules: List[str] = ["Linear"],
-        quantize_dtype: str = "qint8",
-        asr_train_config: Union[Path, str] = None,
-        asr_model_file: Union[Path, str] = None,
-        device: str = "cpu",
[…]
