+++ utils/precompute-asr-feats.py
@@ -5,9 +5,9 @@
 
 import argparse
 import logging
-import sys
 import os
 import re
+import sys
 from pathlib import Path
 from typing import Any, Callable, List, Optional, Sequence, Tuple, Union
 
@@ -18,15 +18,14 @@ import torch
 from packaging.version import parse as V
 from typeguard import check_argument_types
 
-from espnet2.utils import config_argparse
 from espnet2.tasks.asr import ASRTask
 from espnet2.torch_utils.device_funcs import to_device
 from espnet2.torch_utils.set_all_random_seed import set_all_random_seed
+from espnet2.utils import config_argparse
 from espnet2.utils.types import str2bool, str2triple_str, str_or_none
 from espnet.transform.spectrogram import logmelspectrogram
 from espnet.utils.cli_utils import get_commandline_args
 from espnet.utils.cli_writers import file_writer_helper
-
 
 if V(torch.__version__) >= V("1.6.0"):
     from torch.cuda.amp import autocast
@@ -44,7 +43,9 @@ def build_asr_model(
     task = ASRTask if not args.enh_s2t_task else EnhS2TTask
 
     if args.quantize_asr_model:
-        if args.quantize_dtype == "float16" and torch.__version__ < LooseVersion("1.5.0"):
+        if args.quantize_dtype == "float16" and torch.__version__ < LooseVersion(
+            "1.5.0"
+        ):
             raise ValueError(
                 "float16 dtype for dynamic quantization is not supported with "
                 "torch version < 1.5.0. Switch to qint8 dtype instead."
@@ -124,9 +125,9 @@ class Speech2Feat:
         self.dtype = args.dtype
         self.device = args.device
         self.batch_size = args.batch_size
-        self.skip_pad = [] # Name of features that should not be unpadded
-        self.pad_dict = {} # Dimensions to pad for each padded tensor
-        self.batch_dict = {} # Batch dimension for module outputs
+        self.skip_pad = []  # Name of features that should not be unpadded
+        self.pad_dict = {}  # Dimensions to pad for each padded tensor
+        self.batch_dict = {}  # Batch dimension for module outputs
         self.tuple_warn = []
         self.dim_search_done = False
         self.feats_dict = None
@@ -163,9 +164,9 @@ class Speech2Feat:
 
             # Stack multiple layers from from frontend.upstream
             if (
-                path == "frontend.upstream" and 
-                isinstance(output,    list) and 
-                isinstance(output[0], torch.Tensor)
+                path == "frontend.upstream"
+                and isinstance(output, list)
+                and isinstance(output[0], torch.Tensor)
             ):
                 output = torch.stack(output, dim=1)
 
@@ -235,7 +236,7 @@ class Speech2Feat:
         # If we use a batch size larger than one, try to unpad the tensors
         if self.batch_size > 1:
             frontend_max_len = frontend_out_lens.max().item()
-            encoder_max_len  = encoder_out_lens.max().item()
+            encoder_max_len = encoder_out_lens.max().item()
 
             for k, v in self.feats_dict.items():
                 if "frontend" in k and k not in self.skip_pad:
@@ -246,14 +247,17 @@ class Speech2Feat:
                     continue
 
                 try:
-                    pad_dims = set([dim for dim, len_ in enumerate(v.size()) 
-                        if len_ == max_len]) 
+                    pad_dims = set(
+                        [dim for dim, len_ in enumerate(v.size()) if len_ == max_len]
+                    )
                     if k in self.pad_dict.keys():
                         pad_dims = self.pad_dict[k].intersection(pad_dims)
                     if len(pad_dims) == 0:
-                        raise ValueError("No padding dimension found. " + 
-                            f"Tensor has shape {v.size()} but " + 
-                            f"the guessed padded length is {max_len}")
+                        raise ValueError(
[…]
