### 2022-10-19 962cc41
Change the number of times to compute features from 2 to 50 to find batch and padding dimension
