### 2022-10-19 1cfaefe
fix and simplify convinput
