### 2022-10-19 eabe4a5
Add checks to corner case if module output is not a torch Tensor. Add comment to suggest user to change batch size to 1 if unpadding fails
