### 2022-10-20 8eb292c
fix aux. lm loss, reporter variable and doc
