+++ _sources/_gen/espnet.transform.rst.txt
@@ -4,12 +4,12 @@ espnet.transform package
 Initialize main package.
 
 
-.. _espnet.transform.add_deltas:
+.. _espnet.transform.transformation:
 
-espnet.transform.add_deltas
----------------------------
+espnet.transform.transformation
+-------------------------------
 
-.. automodule:: espnet.transform.add_deltas
+.. automodule:: espnet.transform.transformation
     :members:
     :undoc-members:
     :show-inheritance:
@@ -26,56 +26,12 @@ espnet.transform.__init__
     :show-inheritance:
 
 
-.. _espnet.transform.transformation:
-
-espnet.transform.transformation
--------------------------------
-
-.. automodule:: espnet.transform.transformation
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-.. _espnet.transform.functional:
-
-espnet.transform.functional
----------------------------
-
-.. automodule:: espnet.transform.functional
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
 .. _espnet.transform.cmvn:
 
 espnet.transform.cmvn
 ---------------------
 
 .. automodule:: espnet.transform.cmvn
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-.. _espnet.transform.perturb:
-
-espnet.transform.perturb
-------------------------
-
-.. automodule:: espnet.transform.perturb
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-.. _espnet.transform.spectrogram:
-
-espnet.transform.spectrogram
-----------------------------
-
-.. automodule:: espnet.transform.spectrogram
     :members:
     :undoc-members:
     :show-inheritance:
@@ -103,6 +59,17 @@ espnet.transform.channel_selector
     :show-inheritance:
 
 
+.. _espnet.transform.add_deltas:
+
+espnet.transform.add_deltas
+---------------------------
+
+.. automodule:: espnet.transform.add_deltas
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
 .. _espnet.transform.wpe:
 
 espnet.transform.wpe
@@ -114,12 +81,45 @@ espnet.transform.wpe
     :show-inheritance:
 
 
+.. _espnet.transform.perturb:
+
+espnet.transform.perturb
+------------------------
+
+.. automodule:: espnet.transform.perturb
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+.. _espnet.transform.functional:
+
+espnet.transform.functional
+---------------------------
+
+.. automodule:: espnet.transform.functional
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
 .. _espnet.transform.spec_augment:
 
 espnet.transform.spec_augment
 -----------------------------
 
 .. automodule:: espnet.transform.spec_augment
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+.. _espnet.transform.spectrogram:
+
+espnet.transform.spectrogram
+----------------------------
+
+.. automodule:: espnet.transform.spectrogram
     :members:
     :undoc-members:
     :show-inheritance:
