+++ _sources/_gen/espnet2.tasks.rst.txt
@@ -4,23 +4,12 @@ espnet2.tasks package
 
 
 
-.. _espnet2.tasks.__init__:
+.. _espnet2.tasks.lm:
 
-espnet2.tasks.__init__
-----------------------
+espnet2.tasks.lm
+----------------
 
-.. automodule:: espnet2.tasks.__init__
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-.. _espnet2.tasks.enh:
-
-espnet2.tasks.enh
------------------
-
-.. automodule:: espnet2.tasks.enh
+.. automodule:: espnet2.tasks.lm
     :members:
     :undoc-members:
     :show-inheritance:
@@ -37,12 +26,12 @@ espnet2.tasks.diar
     :show-inheritance:
 
 
-.. _espnet2.tasks.st:
+.. _espnet2.tasks.__init__:
 
-espnet2.tasks.st
-----------------
+espnet2.tasks.__init__
+----------------------
 
-.. automodule:: espnet2.tasks.st
+.. automodule:: espnet2.tasks.__init__
     :members:
     :undoc-members:
     :show-inheritance:
@@ -59,89 +48,12 @@ espnet2.tasks.mt
     :show-inheritance:
 
 
-.. _espnet2.tasks.asr_transducer:
-
-espnet2.tasks.asr_transducer
-----------------------------
-
-.. automodule:: espnet2.tasks.asr_transducer
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-.. _espnet2.tasks.svs:
-
-espnet2.tasks.svs
------------------
-
-.. automodule:: espnet2.tasks.svs
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-.. _espnet2.tasks.enh_s2t:
-
-espnet2.tasks.enh_s2t
----------------------
-
-.. automodule:: espnet2.tasks.enh_s2t
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-.. _espnet2.tasks.lm:
-
-espnet2.tasks.lm
-----------------
-
-.. automodule:: espnet2.tasks.lm
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-.. _espnet2.tasks.abs_task:
-
-espnet2.tasks.abs_task
-----------------------
-
-.. automodule:: espnet2.tasks.abs_task
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-.. _espnet2.tasks.tts:
-
-espnet2.tasks.tts
------------------
-
-.. automodule:: espnet2.tasks.tts
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
 .. _espnet2.tasks.hubert:
 
 espnet2.tasks.hubert
 --------------------
 
 .. automodule:: espnet2.tasks.hubert
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-.. _espnet2.tasks.gan_tts:
-
-espnet2.tasks.gan_tts
----------------------
-
-.. automodule:: espnet2.tasks.gan_tts
     :members:
     :undoc-members:
     :show-inheritance:
@@ -164,6 +76,94 @@ espnet2.tasks.slu
 -----------------
 
 .. automodule:: espnet2.tasks.slu
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+.. _espnet2.tasks.enh:
+
+espnet2.tasks.enh
+-----------------
+
+.. automodule:: espnet2.tasks.enh
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+.. _espnet2.tasks.gan_tts:
+
+espnet2.tasks.gan_tts
[…]
