+++ _sources/_gen/espnet2.iterators.rst.txt
@@ -37,23 +37,23 @@ espnet2.iterators.chunk_iter_factory
     :show-inheritance:
 
 
-.. _espnet2.iterators.sequence_iter_factory:
-
-espnet2.iterators.sequence_iter_factory
----------------------------------------
-
-.. automodule:: espnet2.iterators.sequence_iter_factory
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
 .. _espnet2.iterators.abs_iter_factory:
 
 espnet2.iterators.abs_iter_factory
 ----------------------------------
 
 .. automodule:: espnet2.iterators.abs_iter_factory
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+.. _espnet2.iterators.sequence_iter_factory:
+
+espnet2.iterators.sequence_iter_factory
+---------------------------------------
+
+.. automodule:: espnet2.iterators.sequence_iter_factory
     :members:
     :undoc-members:
     :show-inheritance:
