+++ _sources/_gen/espnet2.main_funcs.rst.txt
@@ -4,23 +4,23 @@ espnet2.main_funcs package
 
 
 
-.. _espnet2.main_funcs.__init__:
-
-espnet2.main_funcs.__init__
----------------------------
-
-.. automodule:: espnet2.main_funcs.__init__
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
 .. _espnet2.main_funcs.calculate_all_attentions:
 
 espnet2.main_funcs.calculate_all_attentions
 -------------------------------------------
 
 .. automodule:: espnet2.main_funcs.calculate_all_attentions
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+.. _espnet2.main_funcs.__init__:
+
+espnet2.main_funcs.__init__
+---------------------------
+
+.. automodule:: espnet2.main_funcs.__init__
     :members:
     :undoc-members:
     :show-inheritance:
@@ -37,23 +37,23 @@ espnet2.main_funcs.average_nbest_models
     :show-inheritance:
 
 
-.. _espnet2.main_funcs.collect_stats:
-
-espnet2.main_funcs.collect_stats
---------------------------------
-
-.. automodule:: espnet2.main_funcs.collect_stats
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
 .. _espnet2.main_funcs.pack_funcs:
 
 espnet2.main_funcs.pack_funcs
 -----------------------------
 
 .. automodule:: espnet2.main_funcs.pack_funcs
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+.. _espnet2.main_funcs.collect_stats:
+
+espnet2.main_funcs.collect_stats
+--------------------------------
+
+.. automodule:: espnet2.main_funcs.collect_stats
     :members:
     :undoc-members:
     :show-inheritance:

[…]
