+++ espnet2/gan_svs/vits/generator.py
@@ -14,6 +14,7 @@ from typing import List, Optional, Tuple
 import numpy as np
 import torch
 import torch.nn.functional as F
+from espnet2.gan_svs.vits.phoneme_predictor import PhonemePredictor
 
 from espnet2.gan_tts.hifigan import HiFiGANGenerator
 from espnet2.gan_tts.utils import get_random_segments
@@ -102,7 +103,7 @@ class VITSGenerator(torch.nn.Module):
         stochastic_duration_predictor_dropout_rate: float = 0.5,
         stochastic_duration_predictor_flows: int = 4,
         stochastic_duration_predictor_dds_conv_layers: int = 3,
-        use_sdp: bool = False,
+        use_dp: bool = True,
         use_visinger: bool = True,
     ):
         """Initialize VITS generator module.
@@ -247,21 +248,13 @@ class VITSGenerator(torch.nn.Module):
             use_only_mean=use_only_mean_in_flow,
         )
         self.use_visinger = use_visinger
+        self.use_dp = use_dp
         if self.use_visinger:
             self.project = Projection(hidden_channels, hidden_channels)
 
         # TODO(kan-bayashi): Add deterministic version as an option
-        self.use_sdp = use_sdp
-        if use_sdp:
-            self.duration_predictor = StochasticDurationPredictor(
-                channels=hidden_channels,
-                kernel_size=stochastic_duration_predictor_kernel_size,
-                dropout_rate=stochastic_duration_predictor_dropout_rate,
-                flows=stochastic_duration_predictor_flows,
-                dds_conv_layers=stochastic_duration_predictor_dds_conv_layers,
-                global_channels=global_channels,
-            )
-        else:
+
+        if use_dp:
             self.duration_predictor = DurationPredictor(
                 channels=hidden_channels,
                 filter_channels=256,
@@ -280,6 +273,12 @@ class VITSGenerator(torch.nn.Module):
         self.frame_prior_net = FramePriorNet(
             attention_dim=hidden_channels,
             blocks=4,
+        )
+
+        self.phoneme_predictor = PhonemePredictor(
+            vocabs=vocabs,
+            attention_dim=hidden_channels,
+            blocks=2,
         )
 
         self.upsample_factor = int(np.prod(decoder_upsample_scales))
@@ -373,38 +372,6 @@ class VITSGenerator(torch.nn.Module):
                 - Tensor: Posterior encoder projected scale (B, H, T_feats).
 
         """
-        # forward text encoder
-        for i, length in enumerate(label_xml_lengths):
-            if length == label_xml.shape[1]:
-                label_xml_lengths[i] = feats.shape[2]
-        if label_xml.shape[1] < feats.shape[2]:
-            label_xml = F.pad(
-                input=label_xml,
-                pad=(0, feats.shape[2] - label_xml.shape[1], 0, 0),
-                mode="constant",
-                value=0,
-            )
-            midi_xml = F.pad(
-                input=midi_xml,
-                pad=(0, feats.shape[2] - midi_xml.shape[1], 0, 0),
-                mode="constant",
-                value=0,
-            )
-            beat_xml = F.pad(
-                input=beat_xml,
-                pad=(0, feats.shape[2] - beat_xml.shape[1], 0, 0),
-                mode="constant",
-                value=0,
-            )
-        else:
-            label_xml = label_xml[:, : feats.shape[2]]
-            midi_xml = midi_xml[:, : feats.shape[2]]
-            beat_xml = beat_xml[:, : feats.shape[2]]
-
[…]
