+++ egs2/iam/ocr1/local/data_prep.py
@@ -15,8 +15,9 @@ Expects data to be in:
 Required packages:
     Pillow
 """
-import os
 import argparse
+import os
+
 import numpy as np
 from PIL import Image
 
