### 2022-10-30 05267e2
Fix log to use the term precompute features instead of decoding
