### 2022-10-31 30e7bf2
finishh up whisper frontend
