### 2022-10-31 7a9aff9
Format document using black
