### 2022-11-01 65ee2cd
apply isort, black, flake8 to whisper frontend
