+++ espnet2/asr/frontend/whisper.py
@@ -1,7 +1,5 @@
-import copy
-import logging
 import contextlib
-from typing import Optional, Tuple, Union
+from typing import Tuple, Union
 
 import humanfriendly
 import torch
@@ -13,7 +11,7 @@ from espnet2.asr.frontend.abs_frontend import AbsFrontend
 
 class WhisperFrontend(AbsFrontend):
     """Speech Representation Using Encoder Outputs from OpenAI's Whisper Model:
-       URL: https://github.com/openai/whisper
+    URL: https://github.com/openai/whisper
     """
 
     def __init__(
@@ -23,7 +21,7 @@ class WhisperFrontend(AbsFrontend):
         win_length: int = None,
         hop_length: int = 160,
         n_mels: int = 80,
-        whisper_model: str = 'small',
+        whisper_model: str = "small",
         freeze_weights: bool = True,
         download_dir: str = None,
     ):
@@ -31,7 +29,10 @@ class WhisperFrontend(AbsFrontend):
             import whisper
         except Exception as e:
             print("Error: whisper is not properly installed.")
-            print("Please install whisper with: cd ${MAIN_ROOT}/tools && ./installers/install_whisper.sh")
+            print(
+                "Please install whisper with: cd ${MAIN_ROOT}/tools && "
+                "./installers/install_whisper.sh"
+            )
             raise e
 
         assert check_argument_types()
@@ -40,7 +41,11 @@ class WhisperFrontend(AbsFrontend):
         if isinstance(fs, str):
             fs = humanfriendly.parse_size(fs)
         if fs != 16000:
-            raise ValueError('whisper is trained on 16kHz audios, set fs=16000 instead of {}'.format(fs))
+            raise ValueError(
+                "whisper is trained on 16kHz audios, set fs=16000 instead of {}".format(
+                    fs
+                )
+            )
 
         self.n_fft = n_fft
         if win_length is None:
@@ -51,20 +56,19 @@ class WhisperFrontend(AbsFrontend):
         self.n_mels = n_mels
 
         if n_fft != 400 or self.win_length != 400 or hop_length != 160 or n_mels != 80:
-            raise ValueError('Please use STFT settings under which whisper is trained:\n' + \
-                             '  n_fft = 400, win_length = 400, hop_length = 160, n_mels = 80\n' + \
-                             '  you set n_fft = {}, win_length = {}, hop_length = {}, n_mels = {}'.format(
-                                self.n_fft, self.win_length, self.hop_length, self.n_mels
-                            ))
-        
+            raise ValueError(
+                "Please use STFT settings under which whisper is trained:\n"
+                + "  n_fft=400, win_length=400, hop_length=160, n_mels=80\n"
+                + "  you set n_fft={}, win_len={}, hop_length={}, n_mels={}".format(
+                    self.n_fft, self.win_length, self.hop_length, self.n_mels
+                )
+            )
+
         self.mel_filters = whisper.audio.mel_filters
         self.pad_or_trim = whisper.pad_or_trim
 
         assert whisper_model in whisper.available_models()
-        self.whisper = whisper.load_model(
-                                    whisper_model,
-                                    download_root=download_dir
-                                )
+        self.whisper = whisper.load_model(whisper_model, download_root=download_dir)
         self.whisper.eval()
 
         self.freeze_weights = freeze_weights
@@ -79,13 +83,9 @@ class WhisperFrontend(AbsFrontend):
     ) -> torch.Tensor:
         window = torch.hann_window(self.win_length).to(audio.device)
         stft = torch.stft(
[…]
