### 2022-11-01 8ee32f2
add tested whisper commit hash
