+++ egs2/fleurs/asr1/README.md
@@ -1,4 +1,38 @@
 <!-- Generated by scripts/utils/show_asr_result.sh -->
+# Multilingual ASR - SSL + Conformer + Self-condition [XLS-R, Conformer, utt_mvn, 6500 BPE](conf/train_asr_conformer_scctc.yaml)
+
+## Environments
+- date: `Sat Oct 22 14:55:21 EDT 2022`
+- python version: `3.8.6 (default, Dec 17 2020, 16:57:01)  [GCC 10.2.0]`
+- espnet version: `espnet 202207`
+- pytorch version: `pytorch 1.8.1+cu102`
+- Git hash: `e534106b837ff6cdd29977a52983c022ff1afb0f`
+  - Commit date: `Sun Sep 11 22:31:23 2022 -0400`
+
+## asr_train_asr_conformer_scctc_raw_all_bpe6500_sp
+### Language Identification
+|dataset|Accuracy|
+|---|---|
+|decode_asr_lm_lm_train_lm_all_bpe6500_valid.loss.ave_asr_model_valid.acc.ave_3best/test_all|0.9541 (74237/77809)|
+
+### WER
+
+|dataset|Snt|Wrd|Corr|Sub|Del|Ins|Err|S.Err|
+|---|---|---|---|---|---|---|---|---|
+|decode_asr_lm_lm_train_lm_all_bpe6500_valid.loss.ave_asr_model_valid.acc.ave_3best/test_all|77809|1592160|70.5|26.1|3.4|3.4|32.9|97.0|
+
+### CER
+
+|dataset|Snt|Wrd|Corr|Sub|Del|Ins|Err|S.Err|
+|---|---|---|---|---|---|---|---|---|
+|decode_asr_lm_lm_train_lm_all_bpe6500_valid.loss.ave_asr_model_valid.acc.ave_3best/test_all|77809|10235271|92.2|4.7|3.1|2.6|10.4|97.0|
+
+### TER
+
+|dataset|Snt|Wrd|Corr|Sub|Del|Ins|Err|S.Err|
+|---|---|---|---|---|---|---|---|---|
+|decode_asr_lm_lm_train_lm_all_bpe6500_valid.loss.ave_asr_model_valid.acc.ave_3best/test_all|77809|9622352|91.3|5.6|3.1|2.7|11.4|97.0|
+
 # Multilingual ASR - Self-supervised learning features [HuBERT_large_ll60k, Transformer, utt_mvn, 6500 BPE](conf/train_asr_hubert_large_ll60k_transformer.yaml)
 
 ## Environments
@@ -23,6 +57,7 @@
 |---|---|---|---|---|---|---|---|---|
 |decode_asr_asr_model_valid.acc.best/test_all|77809|1669969|54.2|38.1|7.8|4.0|49.9|99.7|
 |decode_asr_asr_model_valid.acc.best/dev_all|31622|610500|53.4|38.7|7.9|3.5|50.1|99.5|
+
 
 ### CER
 
