+++ egs2/README.md
@@ -99,6 +99,7 @@ See: https://espnet.github.io/espnet/espnet2_tutorial.html#recipes-using-espnet2
 | ofuton_p_utagoe_db      | Ofuton_p_utagoe Singing voice synthesis corpus                                                                                   | SVS                     | JPN                  | https://sites.google.com/view/oftn-utagoedb/%E3%83%9B%E3%83%BC%E3%83%A0                                      |              |
 | open_li110              | Corpus combination with 110 languages                                                                                            | Multilingual ASR        | 100+ languages       |                                                                                                              |              |
 | open_li52               | Corpus combination with 52 languages(Commonvocie + voxforge)                                                                     | Multilingual ASR        | 52 languages         |                                                                                                              |              |
+| opencpop                | Opencpop: Mandarin singing voice synthesis corpus                                                                                | SVS                     | CMN                  | https://wenet.org.cn/opencpop/                                                                               |              |
 | polyphone_swiss_french  | Swiss French Polyphone corpus                                                                                                    | ASR                     | FRA                  | http://catalog.elra.info/en-us/repository/browse/ELRA-S0030_02                                               |              |
 | primewords_chinese      | Primewords Chinese Corpus Set 1                                                                                                  | ASR                     | CMN                  | https://www.openslr.org/47/                                                                                  |              |
 | puebla_nahuatl          | Highland Puebla Nahuatl corpus (endangered language in central Mexico)                                                           | ASR/ST                  | HPN                  | https://www.openslr.org/92/                                                                                  |              |
