### 2022-11-03 1fd60c5
Refactored code as per the guidelines
