+++ egs2/TEMPLATE/asr1/db.sh
@@ -161,6 +161,7 @@ TEDX_SPANISH=downloads
 OFUTON=
 OPENCPOP=
 M_AILABS=downloads
+PARLAMENT=
 
 # For only CMU TIR environment
 if [[ "$(hostname)" == tir* ]]; then
