+++ egs2/parlament/asr1/README.md
@@ -1,37 +1,32 @@
+<!-- Generated by scripts/utils/show_asr_result.sh -->
 # RESULTS
 ## Environments
-- date: `Sat Dec 25 15:43:23 EST 2021`
-- python version: `3.9.7 (default, Sep 16 2021, 13:09:58)  [GCC 7.5.0]`
-- espnet version: `espnet 0.10.5a1`
-- pytorch version: `pytorch 1.9.0`
-- Git hash: `cdf0c002f2a64aa6a670cc7675192ac26f0d5add`
-  - Commit date: `Fri Dec 24 14:45:33 2021 -0500`
+- date: `Fri Sep 30 15:37:34 EDT 2022`
+- python version: `3.8.13 (default, Mar 28 2022, 11:38:47)  [GCC 7.5.0]`
+- espnet version: `espnet 202207`
+- pytorch version: `pytorch 1.12.1`
+- Git hash: `14fcb2d42b2609f766ffaa7a79e9c921cd8398d9`
+  - Commit date: `Tue Sep 27 20:02:22 2022 +0000`
 
-## asr_train_asr_transformer_raw_en_bpe30_sp
+## asr_train_asr_conformer_lr2e-3_warmup15k_amp_nondeterministic_raw_en_bpe47_sp
 ### WER
 
 |dataset|Snt|Wrd|Corr|Sub|Del|Ins|Err|S.Err|
 |---|---|---|---|---|---|---|---|---|
-|beam10_ctc0.3_lm0.1/test|130|773|92.9|5.2|1.9|0.3|7.4|31.5|
-|beam10_ctc0.3_lm0.1/train_dev|100|591|88.0|8.5|3.6|0.7|12.7|45.0|
-|beam10_ctc0.3_lm0/test|130|773|92.8|5.4|1.8|0.4|7.6|30.0|
-|beam10_ctc0.3_lm0/train_dev|100|591|88.0|8.0|4.1|0.8|12.9|47.0|
+|decode_asr_lm_lm_train_lm_en_bpe47_valid.loss.ave_asr_model_valid.acc.ave/test|2138|50994|95.5|3.7|0.9|0.8|5.3|57.8|
+|decode_asr_lm_lm_train_lm_en_bpe47_valid.loss.ave_asr_model_valid.acc.ave/train_dev|100|2569|96.5|3.0|0.6|0.8|4.4|56.0|
 
 ### CER
 
 |dataset|Snt|Wrd|Corr|Sub|Del|Ins|Err|S.Err|
 |---|---|---|---|---|---|---|---|---|
-|beam10_ctc0.3_lm0.1/test|130|2565|96.9|1.0|2.1|0.5|3.6|31.5|
-|beam10_ctc0.3_lm0.1/train_dev|100|1915|93.2|2.5|4.3|0.3|7.1|45.0|
-|beam10_ctc0.3_lm0/test|130|2565|97.1|1.2|1.8|0.7|3.6|30.0|
-|beam10_ctc0.3_lm0/train_dev|100|1915|93.2|2.2|4.6|0.4|7.3|47.0|
+|decode_asr_lm_lm_train_lm_en_bpe47_valid.loss.ave_asr_model_valid.acc.ave/test|2138|304602|98.7|0.6|0.7|0.6|1.8|57.8|
+|decode_asr_lm_lm_train_lm_en_bpe47_valid.loss.ave_asr_model_valid.acc.ave/train_dev|100|15444|99.1|0.4|0.5|0.6|1.5|56.0|
 
 ### TER
 
 |dataset|Snt|Wrd|Corr|Sub|Del|Ins|Err|S.Err|
[…]
