### 2022-11-05 9420874
add model, generator, discriminator, segmenter, and loss scripts
