+++ espnet2/uasr/espnet_model.py
@@ -1,0 +1,476 @@
+import logging
+import argparse
+
+from contextlib import contextmanager
+from typing import Dict, List, Optional, Tuple, Union
+
+import torch
+import kenlm
+import editdistance
+
+import torch.nn.functional as F
+from packaging.version import parse as V
+from typeguard import check_argument_types
+
+from espnet.nets.pytorch_backend.nets_utils import make_pad_mask
+from espnet2.text.token_id_converter import TokenIDConverter
+
+from espnet2.asr.frontend.abs_frontend import AbsFrontend
+from espnet2.uasr.discriminator.abs_discriminator import AbsDiscriminator
+from espnet2.uasr.generator.abs_generator import AbsGenerator
+from espnet2.uasr.segmenter.abs_segmenter import AbsSegmenter
+
+from espnet2.uasr.loss.discriminator_loss import UASRDiscriminatorLoss
+from espnet2.uasr.loss.gradient_penalty import UASRGradientPenalty
+from espnet2.uasr.loss.phoneme_diversity_loss import UASRPhonemeDiversityLoss
+from espnet2.uasr.loss.pseudo_label_loss import UASRPseudoLabelLoss
+from espnet2.uasr.loss.smoothness_penalty import UASRSmoothnessPenalty
+
+from espnet2.torch_utils.device_funcs import force_gatherable
+from espnet2.train.abs_espnet_model import AbsESPnetModel
+from espnet2.utils.types import str2bool
+
+if V(torch.__version__) >= V("1.6.0"):
+    from torch.cuda.amp import autocast
+else:
+    # Nothing to do if torch<1.6.0
+    @contextmanager
+    def autocast(enabled=True):
+        yield
+
+
+class ESPnetUASRModel(AbsESPnetModel):
+    """Unsupervised ASR model"""
+
+    def __init__(
+        self,
+        frontend: Optional[AbsFrontend],
+        segmenter: Optional[AbsSegmenter],
+        generator: AbsGenerator,
+        discriminator: AbsDiscriminator,
+        losses: Dict[str, torch.nn.Module],
+        kenlm_path: Optional[str],
+        token_list: Optional[list],
+        max_epoch: Optional[int],
+        ll_increasing_type: Optional[str],
+        vocab_size: int,
+        cfg: Optional[Dict] = None,
+        pad: int = 1,
+        sil_token: str = "<SIL>",
+        skip_softmax: str2bool = False,
+        use_gumbel: str2bool = False,
+        use_hard_gumbel: str2bool = True,
+        min_temperature: float = 0.1,
+        max_temperature: float = 2.0,
+        decay_temperature: float = 0.99995,
+        use_collected_training_feats: str2bool = False,
+    ):
+        assert check_argument_types()
+
+        super().__init__()
+        # note that eos is the same as sos (equivalent ID)
+        self.frontend = frontend
+        self.segmenter = segmenter
+        self.use_segmenter = True if segmenter is not None else False
+        self.generator = generator
+        self.discriminator = discriminator
+        self.pad = pad
+        if cfg is not None:
+            cfg = argparse.Namespace(**cfg)
+            self.skip_softmax = cfg.no_softmax
+            self.use_gumbel = cfg.gumbel
+            self.use_hard_gumbel = cfg.hard_gumbel
+        else:
+            self.skip_softmax = skip_softmax
+            self.use_gumbel = use_gumbel
+            self.use_hard_gumbel = use_hard_gumbel
+
+        self.use_collected_training_feats = use_collected_training_feats
+
+        self.min_temperature = min_temperature
+        self.max_temperature = max_temperature
+        self.decay_temperature = decay_temperature
+        self.current_temperature = max_temperature
[…]

[…]
