### 2022-11-05 0e144b5
add uasr trainer script
