+++ espnet2/tasks/uasr.py
@@ -1,0 +1,455 @@
+import argparse
+import logging
+import os
+import sys
+from dataclasses import dataclass
+from pathlib import Path
+from typing import Callable, Collection, Dict, List, Optional, Tuple
+
+import numpy as np
+import torch
+from typeguard import check_argument_types, check_return_type
+
+from espnet import __version__
+from espnet2.uasr.espnet_model import ESPnetUASRModel
+from espnet2.asr.frontend.abs_frontend import AbsFrontend
+from espnet2.asr.frontend.default import DefaultFrontend
+from espnet2.asr.frontend.windowing import SlidingWindow
+from espnet2.asr.frontend.s3prl import S3prlFrontend
+from espnet2.asr.frontend.fused import FusedFrontends
+from espnet2.uasr.segmenter.abs_segmenter import AbsSegmenter
+from espnet2.uasr.segmenter.join_segmenter import JoinSegmenter
+from espnet2.uasr.discriminator.abs_discriminator import AbsDiscriminator
+from espnet2.uasr.discriminator.conv_discriminator import ConvDiscriminator
+from espnet2.uasr.generator.abs_generator import AbsGenerator
+from espnet2.uasr.generator.conv_generator import ConvGenerator
+from espnet2.uasr.loss.abs_loss import AbsUASRLoss
+
+from espnet2.uasr.loss.discriminator_loss import UASRDiscriminatorLoss
+from espnet2.uasr.loss.gradient_penalty import UASRGradientPenalty
+from espnet2.uasr.loss.smoothness_penalty import UASRSmoothnessPenalty
+from espnet2.uasr.loss.phoneme_diversity_loss import UASRPhonemeDiversityLoss
+from espnet2.uasr.loss.pseudo_label_loss import UASRPseudoLabelLoss
+from espnet2.uasr.loss.lattice_loss import UASRLatticeLoss
+from espnet2.uasr.loss.autoencoder_loss import UASRAutoencoderlLoss
+
+from espnet2.tasks.abs_task import AbsTask, optim_classes
+from espnet2.torch_utils.initialize import initialize
+from espnet2.train.class_choices import ClassChoices
+from espnet2.train.collate_fn import CommonCollateFn
+from espnet2.train.preprocessor import CommonPreprocessor
+from espnet2.train.uasr_trainer import UASRTrainer
+from espnet2.utils.get_default_kwargs import get_default_kwargs
+from espnet2.utils.nested_dict_action import NestedDictAction
+from espnet2.utils.types import float_or_none, int_or_none, str2bool, str_or_none
+
+frontend_choices = ClassChoices(
+    name="frontend",
+    classes=dict(
+        default=DefaultFrontend,
+        sliding_window=SlidingWindow,
+        s3prl=S3prlFrontend,
+        fused=FusedFrontends,
+    ),
+    type_check=AbsFrontend,
+    default="default",
+)
+segmenter_choices = ClassChoices(
+    name="segmenter",
+    classes=dict(
+        join=JoinSegmenter,
+    ),
+    type_check=AbsSegmenter,
+    default=None,
+    optional=True,
+)
+discriminator_choices = ClassChoices(
+    name="discriminator",
+    classes=dict(
+        conv=ConvDiscriminator,
+    ),
+    type_check=AbsDiscriminator,
+    default="conv",
+)
+generator_choices = ClassChoices(
+    name="generator",
+    classes=dict(
+        conv=ConvGenerator,
+    ),
+    type_check=AbsGenerator,
+    default="conv",
+)
+loss_choices = ClassChoices(
+    name="loss",
+    classes=dict(
+        discriminator_loss=UASRDiscriminatorLoss,
[…]
