### 2022-11-05 7df7e11
add uasr trainer script
