+++ espnet2/train/uasr_trainer.py
@@ -1,0 +1,422 @@
+# Copyright 2022 Tomoki Hayashi
+#           2022 Dongji Gao
+#  Apache 2.0  (http://www.apache.org/licenses/LICENSE-2.0)
+
+"""Trainer module for GAN-based UASR training."""
+
+import argparse
+import dataclasses
+import logging
+import time
+import math
+from contextlib import contextmanager
+from typing import Dict, Iterable, List, Optional, Sequence, Tuple
+
+import editdistance
+import torch
+import kenlm
+from packaging.version import parse as V
+from typeguard import check_argument_types
+
+from espnet2.schedulers.abs_scheduler import AbsBatchStepScheduler, AbsScheduler
+from espnet2.torch_utils.device_funcs import to_device
+from espnet2.torch_utils.recursive_op import recursive_average
+from espnet2.train.distributed_utils import DistributedOption
+from espnet2.train.reporter import SubReporter
+from espnet2.train.trainer import Trainer, TrainerOptions
+from espnet2.utils.build_dataclass import build_dataclass
+from espnet2.utils.types import str2bool
+
+if torch.distributed.is_available():
+    from torch.distributed import ReduceOp
+
+if V(torch.__version__) >= V("1.6.0"):
+    from torch.cuda.amp import GradScaler, autocast
+else:
+    # Nothing to do if torch<1.6.0
+    @contextmanager
+    def autocast(enabled=True):  # NOQA
+        yield
+
+    GradScaler = None
+
+try:
+    import fairscale
+except ImportError:
+    fairscale = None
+
+
+@dataclasses.dataclass
+class UASRTrainerOptions(TrainerOptions):
+    """Trainer option dataclass for UASRTrainer."""
+
+    generator_first: bool
+    max_num_warning: int
+
+
+class UASRTrainer(Trainer):
+    """Trainer for GAN-based UASR training.
+
+    If you'd like to use this trainer, the model must inherit
+    espnet.train.abs_gan_espnet_model.AbsGANESPnetModel.
+
+    """
+
+    @classmethod
+    def build_options(cls, args: argparse.Namespace) -> TrainerOptions:
+        """Build options consumed by train(), eval(), and plot_attention()."""
+        assert check_argument_types()
+        return build_dataclass(UASRTrainerOptions, args)
+
+    @classmethod
+    def add_arguments(cls, parser: argparse.ArgumentParser):
+        """Add additional arguments for GAN-trainer."""
+        parser.add_argument(
+            "--generator_first",
+            type=str2bool,
+            default=False,
+            help="Whether to update generator first.",
+        )
+        parser.add_argument(
+            "--max_num_warning",
+            type=int,
+            default=10,
+            help="Maximum number of warning shown",
+        )
+
+    @classmethod
+    def train_one_epoch(
+        cls,
+        model: torch.nn.Module,
+        iterator: Iterable[Tuple[List[str], Dict[str, torch.Tensor]]],
+        optimizers: Sequence[torch.optim.Optimizer],
+        schedulers: Sequence[Optional[AbsScheduler]],
+        scaler: Optional[GradScaler],
+        reporter: SubReporter,
+        summary_writer,
+        options: UASRTrainerOptions,
+        distributed_option: DistributedOption,
+    ) -> bool:
+        """Train one epoch for UASR."""
+        assert check_argument_types()
+
+        grad_noise = options.grad_noise
+        accum_grad = options.accum_grad
+        grad_clip = options.grad_clip
+        grad_clip_type = options.grad_clip_type
+        log_interval = options.log_interval
+        no_forward_run = options.no_forward_run
+        ngpu = options.ngpu
+        use_wandb = options.use_wandb
+        distributed = distributed_option.distributed
+        max_num_warning = options.max_num_warning
[…]
