### 2022-11-06 f585dc0
Remove unused code and Fix minor errors
