+++ espnet2/asr/encoder/e_branchformer_encoder.py
@@ -3,9 +3,9 @@
 
 """E-Branchformer encoder definition.
 Reference:
-    Kwangyoun Kim, Felix Wu, Yifan Peng, Jing Pan, 
+    Kwangyoun Kim, Felix Wu, Yifan Peng, Jing Pan,
     Prashant Sridhar, Kyu J. Han, Shinji Watanabe,
-    "E-Branchformer: Branchformer with Enhanced merging 
+    "E-Branchformer: Branchformer with Enhanced merging
     for speech recognition," in SLT 2022.
 """
 
@@ -32,8 +32,6 @@ from espnet.nets.pytorch_backend.transformer.embedding import (  # noqa: H301
     ScaledPositionalEncoding,
 )
 from espnet.nets.pytorch_backend.transformer.layer_norm import LayerNorm
-from espnet.nets.pytorch_backend.transformer.multi_layer_conv import Conv1dLinear
-from espnet.nets.pytorch_backend.transformer.multi_layer_conv import MultiLayeredConv1d
 from espnet.nets.pytorch_backend.transformer.positionwise_feed_forward import (
     PositionwiseFeedForward,  # noqa: H301
 )
@@ -55,11 +53,8 @@ class EBranchformerEncoderLayer(torch.nn.Module):
         attn: standard self-attention or efficient attention, optional
         cgmlp: ConvolutionalGatingMLP, optional
         dropout_rate (float): dropout probability
-        merge_method (str): concat, learned_ave, fixed_ave
-        cgmlp_weight (float): weight of the cgmlp branch, between 0 and 1,
-            used if merge_method is fixed_ave
-        attn_branch_drop_rate (float): probability of dropping the attn branch,
-            used if merge_method is learned_ave
+        merge_conv_kernel (int): kernel size of the depth-wise convolution in the merge module
+        merge_stride (int): stride size of the depth-wise convolution in the merge module
     """
 
     def __init__(
@@ -70,9 +65,6 @@ class EBranchformerEncoderLayer(torch.nn.Module):
         feed_forward: Optional[torch.nn.Module],
         feed_forward_macaron: Optional[torch.nn.Module],
         dropout_rate: float,
-        merge_method: str,
-        cgmlp_weight: float = 0.5,
-        attn_branch_drop_rate: float = 0.0,
         merge_conv_kernel: int = 3,
         merge_stride: int = 1,
     ):
@@ -84,10 +76,6 @@ class EBranchformerEncoderLayer(torch.nn.Module):
         self.size = size
         self.attn = attn
         self.cgmlp = cgmlp
-        self.merge_method = merge_method
-        self.cgmlp_weight = cgmlp_weight
-        self.attn_branch_drop_rate = attn_branch_drop_rate
-        self.use_two_branches = (attn is not None) and (cgmlp is not None)
 
         self.feed_forward = feed_forward
         self.feed_forward_macaron = feed_forward_macaron
@@ -106,63 +94,16 @@ class EBranchformerEncoderLayer(torch.nn.Module):
 
         self.dropout = torch.nn.Dropout(dropout_rate)
 
-        if self.use_two_branches:
-            if merge_method == "concat":
-                self.merge_proj = torch.nn.Linear(size + size, size)
-
-            elif merge_method == "concat_conv":
-                self.depthwise_conv_fusion = torch.nn.Conv1d(
-                        size + size,
-                        size + size,
-                        kernel_size=merge_conv_kernel,
-                        stride=merge_stride,
-                        padding=(merge_conv_kernel - 1) // 2,
-                        groups=size + size,
-                        bias=True,
-                )
-                self.merge_proj = torch.nn.Linear(size + size, size)
-
-            elif merge_method == "learned_ave":
[…]
