### 2022-11-07 163e887
Apply changes from PR review Also replaced data.sh with data_multi_speaker.sh. - since that is used for this multi-speaker dataset
