+++ egs2/talromur2/tts1/local/data.sh
@@ -16,45 +16,39 @@ stop_stage=2
 
 . utils/parse_options.sh
 
-if [ $# -ne 1 ]; then
-    log "Error: Only speaker id is required. It should be one of the values [a, b, c, d, e, f, g, h]"
-    exit 2
-fi
-
-speaker_id=$1
-
-if [ $speaker_id = "all" ]; then
-    ./local/data_multi_speaker.sh
-    exit 0
-fi
-
 . ./path.sh || exit 1
 . ./cmd.sh || exit 1
 . ./db.sh || exit 1
 
-if [ -z "${TALROMUR}" ]; then
+if [ -z "${TALROMUR2}" ]; then
    log "Fill the value of 'TALROMUR' of db.sh"
    exit 1
 fi
-db_root=${TALROMUR}
+db_root=${TALROMUR2}
 
-full_set=full_${speaker_id}
-train_set=train_${speaker_id}
-train_dev=dev_${speaker_id}
-eval_set=eval1_${speaker_id}
+full_set=full
+train_set=train
+deveval_set=deveval
+dev_set=dev
+eval_set=eval1
 
 if [ ${stage} -le -1 ] && [ ${stop_stage} -ge -1 ]; then
     log "stage -1: Data Download"
-    local/data_download.sh "${db_root}"
+    if [ ${db_root}=="downloads" ]; then
+        log "Non-default db root provided, skipping download..."
+    else
+        ./local/data_download.sh "${db_root}"
+    fi
 fi
+
+# set filenames
+scp=data/${full_set}/wav.scp
+utt2spk=data/${full_set}/utt2spk
+spk2utt=data/${full_set}/spk2utt
+text=data/${full_set}/text
 
 if [ ${stage} -le 0 ] && [ ${stop_stage} -ge 0 ]; then
     log "stage 0: Data Preparation"
-    # set filenames
-    scp=data/${full_set}/wav.scp
-    utt2spk=data/${full_set}/utt2spk
-    spk2utt=data/${full_set}/spk2utt
-    text=data/${full_set}/text
 
     # check file existence
     [ ! -e data/${full_set} ] && mkdir -p data/${full_set}
@@ -62,52 +56,69 @@ if [ ${stage} -le 0 ] && [ ${stop_stage} -ge 0 ]; then
     [ -e ${utt2spk} ] && rm ${utt2spk}
     [ -e ${spk2utt} ] && rm ${spk2utt}
     [ -e ${text} ] && rm ${text}
+    [ -e data/alignments ] && rm -r data/alignments
 
     # make scp, utt2spk, and spk2utt
-    find ${db_root}/${speaker_id} -follow -name "*.wav" | sort | while read -r filename;do
+    find ${db_root}/s* -follow -name "*.wav" | sort | while read -r filename;do
+        spk_id=$(basename $(dirname $(dirname ${filename})))
         id=$(basename ${filename} | sed -e "s/\.[^\.]*$//g")
         echo "${id} ${filename}" >> ${scp}
-        echo "${id} ${speaker_id}" >> ${utt2spk}
+        echo "${id} ${spk_id}" >> ${utt2spk}
     done
+    sort ${utt2spk} > ${utt2spk}_sorted
+    mv ${utt2spk}_sorted ${utt2spk}
+    sort ${scp} > ${scp}_sorted
+    mv ${scp}_sorted ${scp}
     utils/utt2spk_to_spk2utt.pl ${utt2spk} > ${spk2utt}
 
-    # Trim leading and trailing silences from audio using sox
-    python local/data_utils.py ${db_root}/alignments/${speaker_id}/audio ${scp}
+    if [ -e ${db_root}/alignments ]
+    then
+        # Collect all alignment files in one directory
+        echo "Gathering alignment data"
+        mkdir -p data/alignments
[…]
