### 2022-11-07 6a62e9b
Update model download URL to correct version
