### 2022-11-07 e52cfd8
Fix shellcheck errors and warnings
