+++ egs2/TEMPLATE/asr1/db.sh
@@ -156,6 +156,7 @@ VOXFORGE=downloads
 VOXPOPULI=downloads
 HARPERVALLEY=downloads
 TALROMUR=downloads
+TALROMUR2=downloads
 DCASE=
 TEDX_SPANISH=downloads
 IAM=downloads
