+++ egs2/README.md
@@ -124,7 +124,8 @@ See: https://espnet.github.io/espnet/espnet2_tutorial.html#recipes-using-espnet2
 | swbd                    | Switchboard Corpus for 2-channel Conversational Telephone Speech (300h)                                                          | ASR                     | ENG                  | https://catalog.ldc.upenn.edu/LDC97S62                                                                       |              |
 | swbd_da                 | NXT Switchboard Annotations                                                                                                      | SLU                     | ENG                  | https://catalog.ldc.upenn.edu/LDC2009T26                                                                     |              |
 | swbd_sentiment          | Speech Sentiment Annotations                                                                                                     | SLU                     | ENG                  | https://catalog.ldc.upenn.edu/LDC2020T14                                                                     |              |
-| talromur                | Talromur: A large Icelandic TTS corpus                                                                                           | TTS                     | ISL                  | https://repository.clarin.is/repository/xmlui/handle/20.500.12537/104, https://aclanthology.org/2021.nodalida-main.50.pdf  |              |
+| talromur                | Talromur: A large Icelandic TTS corpus                                                                                           | TTS                     | ISL                  | https://repository.clarin.is/repository/xmlui/handle/20.500.12537/104, https://aclanthology.org/2021.nodalida-main.50.pdf                |              |
+| talromur2               | Talromur 2: Icelandic multi-speaker TTS corpus                                                                                   | TTS                     | ISL                  | https://repository.clarin.is/repository/xmlui/handle/20.500.12537/167                                        |              |
 | tedlium2                | TED-LIUM corpus release 2                                                                                                        | ASR                     | ENG                  | https://www.openslr.org/19/, http://www.lrec-conf.org/proceedings/lrec2014/pdf/1104_Paper.pdf                |              |
 | tedx_spanish_openslr67  | TEDx Spanish Corpus                                                                                                              | ASR                     | SPA                 | https://www.openslr.org/67/                                                                                   |              |
 | thchs30                 | A Free Chinese Speech Corpus Released by CSLT@Tsinghua University                                                                | ASR/TTS                 | CMN                  | https://www.openslr.org/18/                                                                                  |              |
