### 2022-11-08 cf1f704
add lid conditioning
