+++ egs2/fleurs/asr1/README.md
@@ -1,4 +1,39 @@
 <!-- Generated by scripts/utils/show_asr_result.sh -->
+# Multilingual ASR - SSL + Conformer + Hierarchical LID Self-condition [XLS-R, Conformer, utt_mvn, 6500 BPE](conf/train_asr_conformer_hier_lid_utt.yaml)
+
+## Environments
+- date: `Sat Oct 22 17:36:51 CDT 2022`
+- python version: `3.8.13 (default, Mar 28 2022, 11:38:47)  [GCC 7.5.0]`
+- espnet version: `espnet 202207`
+- pytorch version: `pytorch 1.12.1+cu116`
+- Git hash: `14fcb2d42b2609f766ffaa7a79e9c921cd8398d9`
+  - Commit date: `Tue Sep 27 20:02:22 2022 +0000`
+
+## asr_train_asr_conformer_lid_scctc_raw_all_bpe6500_train_data_path_and_name_and_typedumprawtrain_all_splid,lid,text_sp
+
+### Language Identification
+|dataset|Accuracy|
+|---|---|
+|decode_lid_lm_lm_train_lm_all_bpe6500_valid.loss.ave_asr_model_valid.acc.ave_3best/test_all|0.9492 (73860/77809)|
+
+### WER
+
+|dataset|Snt|Wrd|Corr|Sub|Del|Ins|Err|S.Err|
+|---|---|---|---|---|---|---|---|---|
+|decode_lid_lm_lm_train_lm_all_bpe6500_valid.loss.ave_asr_model_valid.acc.ave/test_all|77809|1592160|72.2|25.0|2.9|3.6|31.5|96.6|
+
+### CER
+
+|dataset|Snt|Wrd|Corr|Sub|Del|Ins|Err|S.Err|
+|---|---|---|---|---|---|---|---|---|
+|decode_lid_lm_lm_train_lm_all_bpe6500_valid.loss.ave_asr_model_valid.acc.ave/test_all|77809|10235271|92.5|4.7|2.8|2.6|10.1|96.7|
+
+### TER
+
+|dataset|Snt|Wrd|Corr|Sub|Del|Ins|Err|S.Err|
+|---|---|---|---|---|---|---|---|---|
+|decode_lid_lm_lm_train_lm_all_bpe6500_valid.loss.ave_asr_model_valid.acc.ave/test_all|77809|9622352|91.6|5.6|2.8|2.8|11.2|96.6|
+
 # Multilingual ASR - SSL + Conformer + Self-condition [XLS-R, Conformer, utt_mvn, 6500 BPE](conf/train_asr_conformer_scctc.yaml)
 
 ## Environments
