### 2022-11-08 12c2dc4
add pre-trained fleurs model
