+++ egs2/TEMPLATE/uasr1/scripts/utils/create_README_file.py
@@ -1,0 +1,1 @@
+../../../asr1/scripts/utils/create_README_file.py
