+++ egs2/mini_an4/asr1/multispkr_asr.sh
@@ -1,0 +1,1 @@
+../../TEMPLATE/asr1/multispkr_asr.sh
