+++ egs2/TEMPLATE/asr1/multispkr_asr.sh
@@ -28,6 +28,7 @@ stop_stage=10000     # Processes is stopped at the specified stage.
 skip_data_prep=false # Skip data preparation stages.
 skip_train=false     # Skip training stages.
 skip_eval=false      # Skip decoding and evaluation stages.
+skip_upload=true     # Skip packing and uploading stages.
 skip_upload_hf=true  # Skip uploading to hugging face stages.
 ngpu=1               # The number of gpus ("0" uses cpu, otherwise use gpu).
 num_nodes=1          # The number of nodes.
@@ -59,7 +60,7 @@ oov="<unk>"         # Out of vocabulary symbol.
 blank="<blank>"     # CTC blank symbol
 sos_eos="<sos/eos>" # sos and eos symbole
 bpe_input_sentence_size=100000000 # Size of input sentence for BPE.
-bpe_nlsyms=         # non-linguistic symbols list, separated by a comma, for BPE
+bpe_nlsyms=         # non-linguistic symbols list, separated by a comma or a file containing 1 symbol per line, for BPE
 bpe_char_cover=1.0  # character coverage when modeling BPE
 
 # Ngram model related
@@ -82,53 +83,55 @@ num_splits_lm=1   # Number of splitting for lm corpus.
 word_vocab_size=10000 # Size of word vocabulary.
 
 # ASR model related
-enh_asr_tag=       # Suffix to the result dir for asr model training.
-enh_asr_exp=       # Specify the directory path for ASR experiment.
-                   # If this option is specified, enh_asr_tag is ignored.
-enh_asr_stats_dir= # Specify the directory path for ASR statistics.
-enh_asr_config=    # Config for asr model training.
-enh_asr_args=      # Arguments for asr model training, e.g., "--max_epoch 10".
-                   # Note that it will overwrite args in asr config.
-pretrained_model=          # Pretrained model to load
-ignore_init_mismatch=false # Ignore initial mismatch
+asr_task=asr   # ASR task mode. Either 'asr' or 'asr_transducer'.
+asr_tag=       # Suffix to the result dir for asr model training.
+asr_exp=       # Specify the directory path for ASR experiment.
+               # If this option is specified, asr_tag is ignored.
+asr_stats_dir= # Specify the directory path for ASR statistics.
+asr_config=    # Config for asr model training.
+asr_args=      # Arguments for asr model training, e.g., "--max_epoch 10".
+               # Note that it will overwrite args in asr config.
+pretrained_model=              # Pretrained model to load
+ignore_init_mismatch=false      # Ignore initial mismatch
 feats_normalize=global_mvn # Normalizaton layer type.
 num_splits_asr=1           # Number of splitting for lm corpus.
+ref_num=1   # Number of references for training.
+            # In supervised learning based speech enhancement / separation, it is equivalent to number of speakers.
+inf_num=    # Number of inferences output by the model
+            # Note that if it is not specified, it will be the same as ref_num. Otherwise, it will be overwritten.
+            # In MixIT, number of outputs is larger than that of references.
 
 # Upload model related
 hf_repo=
 
 # Decoding related
 use_k2=false      # Whether to use k2 based decoder
+k2_ctc_decoding=true
+use_nbest_rescoring=true # use transformer-decoder
+                         # and transformer language model for nbest rescoring
+num_paths=1000 # The 3rd argument of k2.random_paths.
+nll_batch_size=100 # Affect GPU memory usage when computing nll
+                   # during nbest rescoring
+k2_config=./conf/decode_asr_transformer_with_k2.yaml
+
+use_streaming=false # Whether to use streaming decoding
+
+use_maskctc=false # Whether to use maskctc decoding
+
[…]
