+++ egs2/mini_an4/asr1/run_multispkr.sh
@@ -1,0 +1,14 @@
+#!/usr/bin/env bash
+# Set bash to 'debug' mode, it will exit on :
+# -e 'error', -u 'undefined variable', -o ... 'error in pipeline', -x 'print commands',
+set -e
+set -u
+set -o pipefail
+
+./multispkr_asr.sh \
+    --lang en \
+    --train_set train_nodev \
+    --valid_set train_dev \
+    --test_sets "train_dev test test_seg" \
+    --ref_num 2 \
+    --lm_train_text "data/train_nodev/text" "$@"

[…]
