### 2022-11-11 419b26a
add scripts for entire multi-iris experiment
