+++ egs2/chime4/enh_asr1/local/run_enh_pretraining.sh
@@ -1,0 +1,34 @@
+#!/usr/bin/env bash
+set -e
+set -u
+set -o pipefail
+
+train_set=
+valid_set=
+test_sets=
+enh_config=
+extra_annotations=
+ref_channel=
+
+. ../utils/parse_options.sh
+
+dir=$PWD
+cd ../../enh1
+
+./enh.sh \
+    --stage 1 \
+    --stop_stage 6 \
+    --train_set "${train_set}" \
+    --valid_set "${valid_set}" \
+    --test_sets "${test_sets}" \
+    --fs 16k \
+    --ngpu 1 \
+    --spk_num 1 \
+    --ref_channel "${ref_channel}" \
+    --local_data_opts "--extra-annotations ${extra_annotations} --stage 1 --stop-stage 2" \
+    --enh_config "${enh_config}" \
+    --use_dereverb_ref false \
+    --use_noise_ref false \
+    --inference_model "valid.loss.best.pth"
+
+cd $dir
