### 2022-11-11 16c276d
fix lm config and remove redundant lines in asr1/data.sh that are added in my previous PR
