### 2022-11-14 77e5cc6
Final version of SNIPER after submission
