+++ espnet2/tasks/abs_task.py
@@ -7,7 +7,7 @@ import sys
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
 from pathlib import Path
-from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple, Union
+from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple, Union, Iterable
 
 import humanfriendly
 import numpy as np
@@ -566,6 +566,18 @@ class AbsTask(ABC):
             help="Enable tensorboard logging",
         )
         group.add_argument(
+            "--val_teacher_forcing",
+            type=str2bool,
+            default=True,
+            help="Whether to use teacher forcing during validation",
+        )
+        group.add_argument(
+            "--save_epochs",
+            type=list,
+            default=[],
+            help="Which epochs to save during training",
+        )
+        group.add_argument(
             "--create_graph_in_tensorboard",
             type=str2bool,
             default=False,
@@ -838,7 +850,7 @@ class AbsTask(ABC):
             "--sniper_conf",
             action=NestedDictAction,
             default=dict(),
-            help="Config which should contain exclude_params and schedule",
+            help="Config which should contain schedule and exclude_params",
         )
 
         cls.trainer.add_arguments(parser)
@@ -849,7 +861,7 @@ class AbsTask(ABC):
 
     @classmethod
     def build_optimizers(
-        cls, args: argparse.Namespace, model: torch.nn.Module,
+        cls, args: argparse.Namespace, model: torch.nn.Module, param_groups: List[Dict] = None
     ) -> List[torch.optim.Optimizer]:
         if cls.num_optimizers != 1:
             raise RuntimeError(
@@ -859,14 +871,16 @@ class AbsTask(ABC):
         optim_class = optim_classes.get(args.optim)
         if optim_class is None:
             raise ValueError(f"must be one of {list(optim_classes)}: {args.optim}")
+
+        params = model.parameters() if param_groups is None else param_groups
         if args.sharded_ddp:
             if fairscale is None:
                 raise RuntimeError("Requiring fairscale. Do 'pip install fairscale'")
             optim = fairscale.optim.oss.OSS(
-                params=model.parameters(), optim=optim_class, **args.optim_conf
+                params=params, optim=optim_class, **args.optim_conf
             )
         else:
-            optim = optim_class(model.parameters(), **args.optim_conf)
+            optim = optim_class(params, **args.optim_conf)
 
         optimizers = [optim]
         return optimizers
@@ -1126,8 +1140,60 @@ class AbsTask(ABC):
                     logging.info(f"Setting {k}.requires_grad = False")
                     p.requires_grad = False
 
+
+        # 7. Build iterator factories
+        if args.multiple_iterator:
+            train_iter_factory = cls.build_multiple_iter_factory(
+                args=args, distributed_option=distributed_option, mode="train",
+            )
+        else:
+            train_iter_factory = cls.build_iter_factory(
+                args=args, distributed_option=distributed_option, mode="train",
+            )
+        valid_iter_factory = cls.build_iter_factory(
+            args=args, distributed_option=distributed_option, mode="valid",
+        )
+        if not args.use_matplotlib and args.num_att_plot != 0:
+            args.num_att_plot = 0
+            logging.info("--use_matplotlib false => Changing --num_att_plot to 0")
+
+        if args.num_att_plot != 0:
+            plot_attention_iter_factory = cls.build_iter_factory(
+                args=args, distributed_option=distributed_option, mode="plot_att",
+            )
+        else:
+            plot_attention_iter_factory = None
+
+        if args.use_sniper:
+            sniper = SniperTraining(sniper_dir=args.sniper_dir, device=device, logger=logging.root)
[…]
