+++ paper/test_results.py
@@ -1,0 +1,224 @@
+import time
+from ast import literal_eval
+import os
+import requests
+from glob import glob
+from pathlib import Path
+import traceback
+import pandas as pd
+from tacotron_cleaner.cleaners import *
+import base64
+import json
+import jiwer
+from vocoder_eval import eval_rmse_f0
+
+GROUND_TRUTH_DIR = '/home/perry/PycharmProjects/LJSpeech-1.1/wavs'
+
+SPEECHMATICS_POST_URL = 'https://asr.api.speechmatics.com/v2/jobs/'
+SPEECHMATICS_GET_URL = 'https://asr.api.speechmatics.com/v2/jobs/{job_id}/transcript'
+SPEECHMATICS_FILE = '/home/perry/PycharmProjects/speechmatics_key.txt'
+with open(SPEECHMATICS_FILE) as f:
+    SPEECHMATICS_KEY = f.read().strip()
+    SPEECHMATICS_HEADERS = {'Authorization': f'Bearer {SPEECHMATICS_KEY}'}
+
+
+SAMPLE_RATE = 22050
+
+def print_all_chars(split_csvs):
+    charset = set()
+    for split_csv in split_csvs:
+        for chars in pd.read_csv(split_csv).raw_text:
+            charset.update(set(chars))
+    print(''.sorted(charset))
+
+PUNCS = re.compile(r'[!"\'(),-.:;?\[\]’“”]')
+
+def normalize(text):
+    text = convert_to_ascii(text)
+    text = lowercase(text)
+    text = expand_numbers(text)
+    text = expand_abbreviations(text)
+    text = expand_symbols(text)
+    text = collapse_whitespace(text)
+    return PUNCS.sub('', text)
+
+
+def speechmatics_asr(speech_file):
+    with open(speech_file, 'rb') as f:
+        files = {
+            'data_file': f,
+            'config': (None, '{"type": "transcription",'
+                             '"transcription_config": {"operating_point": "standard", "language": "en" }}'),
+        }
+        response = requests.post(url=SPEECHMATICS_POST_URL, headers=SPEECHMATICS_HEADERS, files=files)
+    if response.ok:
+        job_id = literal_eval(response.text)['id']
+        kwargs = {'url': SPEECHMATICS_GET_URL.format(job_id=job_id),
+                  'params': {'format': 'txt'},
+                  'headers': SPEECHMATICS_HEADERS}
+        transcript = requests.get(**kwargs)
+        time.sleep(3)
+        while not transcript.ok:
+            if transcript.status_code == 404:
+                time.sleep(1)
+                transcript = requests.get(**kwargs)
+            else:
+                raise requests.HTTPError(transcript.text)
+        return transcript.text
+    else:
+        raise requests.HTTPError(response.text)
+
+
+def get_true_normalized_text(metadata_csv):
+    import csv
+    df = pd.read_csv(metadata_csv, sep='|', quoting=csv.QUOTE_NONE, names=['wav_name', 'raw_text', 'clean_text'])
+    normalized_text = df.clean_text.apply(normalize)
+    return dict(zip(df.wav_name, normalized_text))
+
+
+def compute_wer(transcripts_csv, wav_name_to_text, wav_suffix='_gen', verbose=False):
+    wer_csv = Path(transcripts_csv).parent / 'wer.csv'
+    trans_df = pd.read_csv(transcripts_csv)
+
+    results = []
+    for wav_name, transcript in zip(trans_df.wav_name, trans_df.transcript):
+        if wav_name.endswith(wav_suffix):
+            wav_name = wav_name[:-len(wav_suffix)]
+        true_text = wav_name_to_text[wav_name]
+        wer = jiwer.wer(truth=true_text, hypothesis=transcript)
+        words = len(true_text.split())
[…]
