### 2022-11-14 20bd312
Final version of SNIPER after submission
