+++ README.md
@@ -3,9 +3,9 @@
 # SNIPER Training: Single-Shot Initialization Pruning Evolving-Rate Training
 
 
-This repository contains the code for the paper "SNIPER TRAINING: VARIABLE SPARSITY RATE TRAINING FOR TEXT-TO-SPEECH" (will post the arXiv link when ready).
+This repository contains the code for the paper "SNIPER TRAINING: VARIABLE SPARSITY RATE TRAINING FOR TEXT-TO-SPEECH" ([arXiv link](https://arxiv.org/abs/2211.07283)), submitted to ICASSP 2023.
 
-To replicate my experiments, please install [Montreal Forced Aligner] (https://github.com/MontrealCorpusTools/Montreal-Forced-Aligner) and follow the ESPnet installation instructions but on this repo. Use [this guide](https://github.com/espnet/espnet/issues/4521) for stage 1-5 of training, and for stage 6:
+To replicate my experiments, please install [Montreal Forced Aligner](https://github.com/MontrealCorpusTools/Montreal-Forced-Aligner) and follow the ESPnet installation instructions but on this repo. Use [this guide](https://github.com/espnet/espnet/issues/4521) for stage 1-5 of training, and for stage 6:
 
 ```
 ./run_mfa.sh --stage 6 --stop_stage 6     --train_config conf/tuning/train_fastspeech2_sniper40to0.yaml     --teacher_dumpdir data     --tts_stats_dir data/stats      --tts_exp exp/fastspeech2_40to0
