### 2022-11-15 3104399
update core UASR recipes
