+++ egs2/TEMPLATE/uasr1/uasr.sh
@@ -67,6 +67,9 @@ sos_eos="<sos/eos>" # sos and eos symbole
 bpe_input_sentence_size=100000000 # Size of input sentence for BPE.
 bpe_nlsyms=         # non-linguistic symbols list, separated by a comma or a file containing 1 symbol per line, for BPE
 bpe_char_cover=1.0  # character coverage when modeling BPE
+postprocess_word_boundary='  '   # word boundary for post process
+postprocess_sil_token="\<SIL\>"  # silence injection tokens in post process
+postprocess_sil_prob=0.5         # silence injection probability in post process
 
 # Ngram language model related
 use_ngram=true      # Whether to use n-gram modeling (Current version must set to true)
@@ -195,6 +198,9 @@ Options:
     --bpe_input_sentence_size # Size of input sentence for BPE (default="${bpe_input_sentence_size}").
     --bpe_nlsyms              # Non-linguistic symbol list for sentencepiece, separated by a comma or a file containing 1 symbol per line . (default="${bpe_nlsyms}").
     --bpe_char_cover          # Character coverage when modeling BPE (default="${bpe_char_cover}").
+    --postprocess_word_boundary # word boundary for post process (default="${postprocess_word_boundary}").
+    --postprocess_sil_token     # silence injection tokens in post process (default="${postprocess_sil_token}").
+    --postprocess_sil_prob      # silence injection probability in post process (default="${postprocess_sil_prob}").
 
     # Ngram language model related
     --use_ngram       # Whether to use n-gram modeling (Current version must set to true)
@@ -671,112 +677,27 @@ if ! "${skip_data_prep}"; then
             utils/fix_data_dir.sh "${data_feats}/${dset}"
         done
 
-        # shellcheck disable=SC2002
-        cat ${lm_train_text} | awk ' { if( NF != 2 ) print $0; } ' > "${data_feats}/lm_train.txt"
     fi
 
 
     if [ ${stage} -le 6 ] && [ ${stop_stage} -ge 6 ]; then
+        _logdir="${tokenlist_dir}/logdir"
+
+        # shellcheck disable=SC2002
+        cat ${lm_train_text} | awk ' { if( NF != 2 ) print $0; } ' > "${data_feats}/lm_train.txt"
+
         if [ "${token_type}" = bpe ]; then
             log "Stage 6: Generate token_list from ${bpe_train_text} using BPE"
 
-            mkdir -p "${bpedir}"
-            # shellcheck disable=SC2002
-            cat ${bpe_train_text} | cut -f 2- -d" "  > "${bpedir}"/train.txt
-
-            if [ -n "${bpe_nlsyms}" ]; then
-                if test -f "${bpe_nlsyms}"; then
-                    bpe_nlsyms_list=$(awk '{print $1}' ${bpe_nlsyms} | paste -s -d, -)
-                    _opts_spm="--user_defined_symbols=${bpe_nlsyms_list}"
-                else
-                    _opts_spm="--user_defined_symbols=${bpe_nlsyms}"
-                fi
-            else
-                _opts_spm=""
-            fi
-
-            spm_train \
-                --input="${bpedir}"/train.txt \
-                --vocab_size="${nbpe}" \
-                --model_type="${bpemode}" \
-                --model_prefix="${bpeprefix}" \
-                --character_coverage=${bpe_char_cover} \
-                --input_sentence_size="${bpe_input_sentence_size}" \
-                ${_opts_spm}
-
-            {
-            echo "${blank}"
-            echo "${oov}"
-            # Remove <unk>, <s>, </s> from the vocabulary
-            <"${bpeprefix}".vocab awk '{ if( NR != 1 && NR != 2 && NR != 3 ){ print $1; } }'
[…]
