### 2022-11-17 eab00ce
change permissions and adopt to the latest enh.sh and enh_asr.sh
