### 2022-11-17 7b6655c
fix the epochs in lm config
