### 2022-11-17 84915c9
add entry script run.sh
