### 2022-11-17 bb693b2
refactor the way to specify stages
