### 2022-11-17 ee286e2
update egs2/README.md
