+++ egs2/README.md
@@ -84,6 +84,7 @@ See: https://espnet.github.io/espnet/espnet2_tutorial.html#recipes-using-espnet2
 | lt_slurp_spatialized    | Spatialized Libri-Trans and Spatialized SLURP (LT-S and SLURP-S), Enhancement for Translation and Understanding Dataset          | SE/ST/SLU               | ENG                  |                                                                                                              |              |
 | magicdata               | MAGICDATA Mandarin Chinese Read Speech Corpus                                                                                    | ASR                     | ENG                  | https://www.openslr.org/68/                                                                                  |              |
 | mediaspeech             | MediaSpeech: Multilanguage ASR Benchmark and Dataset                                                                             | ASR                     | FRA                  | https://www.openslr.org/108/                                                                                 |              |
+| meld                    | MELD: Multimodal EmotionLines Dataset                                                                                            | SLU                     | ENG                  | https://affective-meld.github.io/									     |              |
 | microsoft_speech        | Microsoft Speech Corpus (Indian languages)                                                                                       | ASR                     | 3 languages          | https://msropendata.com/datasets/7230b4b1-912d-400e-be58-f84e0512985e                                        |              |
 | mini_an4                | Mini version of CMU AN4 database for the integration test                                                                        | ASR/TTS/SE              | ENG                  | http://www.speech.cs.cmu.edu/databases/an4/                                                                  |              |
 | mini_librispeech        | Mini version of Librispeech corpus                                                                                               | DIAR                    | ENG                  | https://openslr.org/31/                                                                                      |              |
