+++ _sources/_gen/espnet2.text.rst.txt
@@ -4,6 +4,17 @@ espnet2.text package
 
 
 
+.. _espnet2.text.hugging_face_tokenizer:
+
+espnet2.text.hugging_face_tokenizer
+-----------------------------------
+
+.. automodule:: espnet2.text.hugging_face_tokenizer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
 .. _espnet2.text.__init__:
 
 espnet2.text.__init__
