+++ _sources/_gen/espnet2.bin.rst.txt
@@ -345,6 +345,17 @@ espnet2.bin.asr_train
     :show-inheritance:
 
 
+.. _espnet2.bin.hugging_face_export_vocabulary:
+
+espnet2.bin.hugging_face_export_vocabulary
+------------------------------------------
+
+.. automodule:: espnet2.bin.hugging_face_export_vocabulary
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
 .. _espnet2.bin.slu_train:
 
 espnet2.bin.slu_train
