+++ _sources/_gen/espnet2_bin.rst.txt
@@ -158,6 +158,8 @@ asr_inference.py
                             [--minlenratio MINLENRATIO] [--ctc_weight CTC_WEIGHT]
                             [--lm_weight LM_WEIGHT] [--ngram_weight NGRAM_WEIGHT]
                             [--streaming STREAMING]
+                            [--hugging_face_decoder HUGGING_FACE_DECODER]
+                            [--hugging_face_decoder_max_length HUGGING_FACE_DECODER_MAX_LENGTH]
                             [--transducer_conf TRANSDUCER_CONF]
                             [--token_type {char,bpe,None}] [--bpemodel BPEMODEL]
     
@@ -240,6 +242,8 @@ asr_inference.py
       --ngram_weight NGRAM_WEIGHT
                             ngram weight (default: 0.9)
       --streaming STREAMING
+      --hugging_face_decoder HUGGING_FACE_DECODER
+      --hugging_face_decoder_max_length HUGGING_FACE_DECODER_MAX_LENGTH
       --transducer_conf TRANSDUCER_CONF
                             The keyword arguments for transducer beam search.
                             (default: None)
@@ -432,7 +436,8 @@ asr_train.py
                         [--input_size INPUT_SIZE] [--ctc_conf CTC_CONF]
                         [--joint_net_conf JOINT_NET_CONF]
                         [--use_preprocessor USE_PREPROCESSOR]
-                        [--token_type {bpe,char,word,phn}] [--bpemodel BPEMODEL]
+                        [--token_type {bpe,char,word,phn,hugging_face}]
+                        [--bpemodel BPEMODEL]
                         [--non_linguistic_symbols NON_LINGUISTIC_SYMBOLS]
                         [--cleaner {None,tacotron,jaconv,vietnamese}]
                         [--g2p {None,g2p_en,g2p_en_no_space,pyopenjtalk,pyopenjtalk_kana,pyopenjtalk_accent,pyopenjtalk_accent_with_pause,pyopenjtalk_prosody,pypinyin_g2p,pypinyin_g2p_phone,pypinyin_g2p_phone_without_prosody,espeak_ng_arabic,espeak_ng_german,espeak_ng_french,espeak_ng_spanish,espeak_ng_russian,espeak_ng_greek,espeak_ng_finnish,espeak_ng_hungarian,espeak_ng_dutch,espeak_ng_english_us_vits,espeak_ng_hindi,espeak_ng_italian,espeak_ng_ukrainian,espeak_ng_polish,g2pk,g2pk_no_space,g2pk_explicit_space,korean_jaso,korean_jaso_no_space,g2p_is}]
@@ -454,7 +459,7 @@ asr_train.py
                         [--encoder_conf ENCODER_CONF]
                         [--postencoder {hugging_face_transformers,None}]
                         [--postencoder_conf POSTENCODER_CONF]
-                        [--decoder {transformer,lightweight_conv,lightweight_conv2d,dynamic_conv,dynamic_conv2d,rnn,transducer,mlm}]
+                        [--decoder {transformer,lightweight_conv,lightweight_conv2d,dynamic_conv,dynamic_conv2d,rnn,transducer,mlm,hugging_face_transformers}]
                         [--decoder_conf DECODER_CONF]
     
     base parser
@@ -804,7 +809,7 @@ asr_train.py
     
       --use_preprocessor USE_PREPROCESSOR
                             Apply preprocessing to data or not (default: True)
-      --token_type {bpe,char,word,phn}
+      --token_type {bpe,char,word,phn,hugging_face}
                             The text will be tokenized in the specified level token (default: bpe)
       --bpemodel BPEMODEL   The model file of sentencepiece (default: None)
       --cleaner {None,tacotron,jaconv,vietnamese}
@@ -852,7 +857,7 @@ asr_train.py
[…]
