+++ egs2/chime4/enh_asr1/README.md
@@ -1,4 +1,52 @@
-<!-- Generated by scripts/utils/show_asr_result.sh -->
+# MultiIRIS (WPD-beamformer-Enh, WavLM_Large + SpecAug + Conformer-ASR with Transformer-LM)
+
+## Notes
+- Joint finetuning requires pre-trained Enh and ASR models.
+- `local/run_multiiris.sh` performs (1) pre-training of the Enh model, (2) pre-training of the ASR model, and (3) joint finetuning of the entire system.
+- The language model is also trained on (2).
+- The final scores will be summarized in `enh_asr1/exp/*` similar to `run.sh`.
+
+## Environments
+- date: `Tue Oct 11 02:40:53 UTC 2022`
+- python version: `3.7.4 (default, Aug 13 2019, 20:35:49)  [GCC 7.3.0]`
+- espnet version: `espnet 202207`
+- pytorch version: `pytorch 1.10.1+cu111`
+- Git hash: `8ed83f45d5aa2ca6b3635e44b9c29afb9b5fb600`
+  - Commit date: `Tue Oct 11 18:59:57 2022 +0900`
+
+## enh_asr_train_enh_asr_wpd_init_noenhloss_wavlm_conformer_raw_en_char
+- Enh pre-training config: [../enh1/conf/tuning/train_enh_beamformer_wpd_ci_sdr_shorttap.yaml](../enh1/conf/tuning/train_enh_beamformer_wpd_ci_sdr_shorttap.yaml)
+- ASR pre-training config: [../asr1/conf/tuning/train_asr_conformer_wavlm2.yaml](../asr1/conf/tuning/train_asr_conformer_wavlm2.yaml)
+- Enh-ASR finetuning config: [./conf/tuning/train_enh_asr_wpd_init_noenhloss_wavlm_conformer.yaml](./conf/tuning/train_enh_asr_wpd_init_noenhloss_wavlm_conformer.yaml)
+- LM config: [../asr1/conf/train_lm_transformer.yaml](../asr1/conf/train_lm_transformer.yaml)
+- Pretrained model: [https://huggingface.co/Yoshiki/chime4_enh_asr1_wpd_wavlm_conformer](https://huggingface.co/Yoshiki/chime4_enh_asr1_wpd_wavlm_conformer)
+
+### WER
+
+|dataset|Snt|Wrd|Corr|Sub|Del|Ins|Err|S.Err|
+|---|---|---|---|---|---|---|---|---|
+|decode_asr_transformer_largelm_normalize_output_wavtrue_lm_lm_train_lm_transformer_en_char_valid.loss.ave_enh_asr_model_valid.acc.ave_10best/dt05_real_isolated_6ch_track|1640|27119|98.8|0.9|0.2|0.2|1.3|16.2|
+|decode_asr_transformer_largelm_normalize_output_wavtrue_lm_lm_train_lm_transformer_en_char_valid.loss.ave_enh_asr_model_valid.acc.ave_10best/dt05_simu_isolated_6ch_track|1640|27120|98.9|0.9|0.2|0.1|1.3|15.2|
+|decode_asr_transformer_largelm_normalize_output_wavtrue_lm_lm_train_lm_transformer_en_char_valid.loss.ave_enh_asr_model_valid.acc.ave_10best/et05_real_isolated_6ch_track|1320|21409|98.4|1.4|0.2|0.2|1.8|20.6|
[…]
