### 2022-11-18 65aa439
softlinked train config
