### 2022-11-18 f06215c
resolve run.sh requests
