### 2022-11-18 830c6d7
remove unused transcript file generation
