### 2022-11-19 37e54d4
fix random seed for stable test
