+++ egs2/TEMPLATE/uasr1/uasr.sh
@@ -75,6 +75,7 @@ postprocess_sil_prob=0.5         # silence injection probability in post process
 use_ngram=true      # Whether to use n-gram modeling (Current version must set to true)
 ngram_exp=          # Specify the diretory path for ngram LM experiments
 ngram_num=4         # Specify the n in ngram LM
+kenlm_path=         # Pre-trained ngram LM
 
 # Language model related
 use_lm=true       # Use language model for uasr decoding.
@@ -206,6 +207,7 @@ Options:
     --use_ngram       # Whether to use n-gram modeling (Current version must set to true)
     --ngram_exp       # Specify the diretory path for ngram LM experiments
     --ngram_num       # Specify the n in ngram LM
+    --kenlm_path      # Specify the pre-trained LM path
 
     # Language model related
     --lm_tag          # Suffix to the result dir for language model training (default="${lm_tag}").
@@ -531,14 +533,16 @@ if ! "${skip_data_prep}"; then
 
     if [ ${stage} -le 3 ] && [ ${stop_stage} -ge 3 ]; then
         log "Stage 3: VAD computing for all data"
-
-        # FIXME(jiatong): we now support limited wav format (i.e., only wav or flac for now) to process it
-        for dset in "${train_set}" "${valid_set}" ${test_sets}; do
-            # FIXME(jiatong): only fs=16000 is supported now
-            scripts/audio/compute_vad.sh  --cmd "${train_cmd}" --nj "${nj}" "${VAD_HOME}" "data/${dset}"
-
-            utils/fix_data_dir.sh data/"${dset}"
-        done
+        if "${silence_trim}"; then
+            # FIXME(jiatong): we now support limited wav format (i.e., only wav or flac for now) to process it
+            for dset in "${train_set}" "${valid_set}" ${test_sets}; do
+                # FIXME(jiatong): only fs=16000 is supported now
+                scripts/audio/compute_vad.sh  --cmd "${train_cmd}" --nj "${nj}" "${VAD_HOME}" "data/${dset}"
+                utils/fix_data_dir.sh data/"${dset}"
+            done
+        else
+            log "Skip stage 3: VAD computing for data"
+        fi
     fi
 
     if [ ${stage} -le 4 ] && [ ${stop_stage} -ge 4 ]; then
@@ -676,7 +680,7 @@ if ! "${skip_data_prep}"; then
 
 
     if [ ${stage} -le 6 ] && [ ${stop_stage} -ge 6 ]; then
-        _logdir="${tokenlist_dir}/logdir"
+        _logdir="${tokendir}/logdir"
 
         # shellcheck disable=SC2002
         cat ${lm_train_text} | awk ' { if( NF != 2 ) print $0; } ' > "${data_feats}/lm_train.txt"
@@ -703,46 +707,39 @@ if ! "${skip_data_prep}"; then
 
             utils/split_scp.pl "${data_feats}/lm_train.txt" ${split_text}
 
-            ${train_cmd} JOB=1:${nj} ${_logdir}/lm/tokenize_text.log.JOB \
+            ${train_cmd} JOB=1:${nj} ${_logdir}/lm/tokenize_text.JOB.log \
             ${python} -m espnet2.bin.tokenize_text \
                 --token_type ${token_type} ${_opts} \
                 --cleaner "${cleaner}" \
                 --input ${split_dir}/JOB/text \
                 --output ${split_dir}/JOB/tokens.txt \
-                --text_output ${split_dir}/JOB/unpaired_text_nosil \
+                --text_output ${split_dir}/JOB/unpaired_text \
                 --write_vocabulary true \
                 --write_text true \
                 --cutoff 2
-
-             ${train_cmd} JOB=1:${nj} ${_logdir}/lm/post_processing.log.JOB \
[…]
