+++ egs2/TEMPLATE/asr1/db.sh
@@ -75,7 +75,7 @@ SPGISPEECH=
 SWBD=
 SWBD_NXT=
 THCHS30=downloads
-TIMIT=/usr0/home/jiatongs/project/wav2vec-u-2-reproduce/espnet/egs2/timit/TIMIT
+TIMIT=
 TSUKUYOMI=downloads
 VOXFORGE=downloads
 AMI=
