+++ egs2/TEMPLATE/asr1/db.sh
@@ -163,6 +163,7 @@ IAM=downloads
 OFUTON=
 OPENCPOP=
 M_AILABS=downloads
+CATSLU=downloads
 
 # For only CMU TIR environment
 if [[ "$(hostname)" == tir* ]]; then
