+++ egs2/opencpop/svs1/local/data_prep.py
@@ -1,13 +1,16 @@
 import argparse
 import os
+import shutil
 
 import librosa
 import miditoolkit
 import numpy as np
 
-from espnet2.fileio.xml_scp import XMLScpWriter
-from espnet2.text.phoneme_tokenizer import pypinyin_g2p_phone_without_prosody
 
+def makedir(data_url):
+    if os.path.exists(data_url):
+        shutil.rmtree(data_url)
+    os.makedirs(data_url)
 
 def load_midi_note_scp(midi_note_scp):
     # Note(jiatong): midi format is 0-based
@@ -34,112 +37,39 @@ def load_midi(args):
         midis[i] = tempo
     return midis
 
-
-def create_xml(lyrics, phns, notes, syb_dur, tempo, keep):
+def create_score(uid, phns, notes, syb_dur, keep):
     assert len(phns) == len(notes)
     assert len(notes) == len(syb_dur)
     assert len(syb_dur) == len(keep)
     lyrics_seq = []
     notes_seq = []
     segs_seq = []
-    index = 0
+    phns_seq = []
     st = 0
-    et = 0
-    while index < len(phns) and notes[index] == 0:
-        et += syb_dur[index]
-        index += 1
-    if st != et:
-        lyrics_seq.append("-1")
-        notes_seq.append(-1)
-        segs_seq.append([st, et])
-    for syllable in lyrics:
-        et = st
-        lyrics_phn = pypinyin_g2p_phone_without_prosody(syllable)
-        lyrics_seq.append(syllable)
-        notes_seq.append(notes[index])
-        for i in range(len(lyrics_phn)):
-            assert keep[index + i] == 0
-            assert notes[index + i] == notes[index]
-            assert syb_dur[index + i] == syb_dur[index]
-        et += syb_dur[index]
-        segs_seq.append([st, et])
-        segs_seq.append([st, et])
-        st = et
-        index += len(lyrics_phn)
-        while index < len(phns) and keep[index] == 1:
-            lyrics_seq.append("")
-            notes_seq.append(notes[index])
-            et += syb_dur[index]
-            segs_seq.append([st, et])
-            index += 1
-            st = et
-        while index < len(phns) and notes[index] == 0:
-            et += syb_dur[index]
-            index += 1
-        if st != et:
-            lyrics_seq.append("P")
-            notes_seq.append(-1)
-            segs_seq.append([st, et])
-    return lyrics_seq, notes_seq, segs_seq, tempo
-
-
-def preprocess(phns, notes, syb_dur, phn_dur, keep, min_beat):
-    new_phns, new_notes, new_syb_dur, new_phn_dur, new_keep = [], [], [], [], []
-    start_index = 0
-    st = 0
-    while start_index < len(phns) and phns[start_index] in ["SP", "AP"]:
-        st += phn_dur[start_index]
-        start_index += 1
-    end_index = len(phns) - 1
-    while end_index >= 0 and phns[end_index] in ["SP", "AP"]:
-        end_index -= 1
-    index = start_index
-    while index <= end_index:
-        duration = 0
-        merge = 0
-        while phns[index] in ["SP", "AP"]:
-            duration += phn_dur[index]
-            index += 1
-        if duration == 0:  # not pause
-            if keep[index] == 1 and notes[index] == new_notes[-1]:
-                duration = phn_dur[index]
-                merge = 1
-            else:
-                new_phns.append(phns[index])
-                new_notes.append(notes[index])
-                new_syb_dur.append(syb_dur[index])
-                new_phn_dur.append(phn_dur[index])
[…]
