+++ egs2/ofuton_p_utagoe_db/svs1/local/data.sh
@@ -17,6 +17,8 @@ SECONDS=0
 stage=1
 stop_stage=100
 fs=None
+g2p=pyopenjtalk
+
 
 log "$0 $*"
 
@@ -51,17 +53,26 @@ if [ ${stage} -le 2 ] && [ ${stop_stage} -ge 2 ]; then
     log "stage 2: Prepare segments"
     for x in ${train_set} ${train_dev} ${recog_set}; do
         src_data=data/${x}
-        local/prep_segments.py --silence pau --silence sil --silence br ${src_data} 10000 # in ms
+        local/prep_segments.py --silence pau --silence sil --silence br ${src_data}
         mv ${src_data}/segments.tmp ${src_data}/segments
         mv ${src_data}/label.tmp ${src_data}/label
-        local/prep_segments_from_xml.py ${src_data} 10000 # in ms
+        local/prep_segments_from_xml.py --silence P --silence B ${src_data}
         mv ${src_data}/text.tmp ${src_data}/text
         mv ${src_data}/segments_from_xml.tmp ${src_data}/segments_from_xml
-        mv ${src_data}/xmlnote.tmp ${src_data}/xmlnote
+        mv ${src_data}/score.tmp ${src_data}/score
         awk '{printf("%s ofuton\n", $1);}' < ${src_data}/segments > ${src_data}/utt2spk
         utils/utt2spk_to_spk2utt.pl < ${src_data}/utt2spk > ${src_data}/spk2utt
-        utils/fix_data_dir.sh --utt_extra_files label ${src_data}
+        utils/fix_data_dir.sh --utt_extra_files "label score" ${src_data}
     done
 fi
 
+if [ ${stage} -le 3 ] && [ ${stop_stage} -ge 3 ]; then
+    log "stage 3: Check alignments"
+    for x in ${train_set} ${train_dev} ${recog_set}; do
+        src_data=data/${x}
+        local/check_align.py ${src_data} --g2p ${g2p}
+        mv ${src_data}/score.tmp ${src_data}/score
+        utils/fix_data_dir.sh --utt_extra_files "label score" ${src_data}
+    done
+fi
 log "Successfully finished. [elapsed=${SECONDS}s]"

[…]
