+++ egs2/TEMPLATE/svs1/svs.sh
@@ -231,7 +231,7 @@ else
 fi
 
 # Extra files for SVS
-utt_extra_files="label xml.scp"
+utt_extra_files="label score"
 
 # Check token list type
 token_listdir="data/token_list/${token_type}"
@@ -306,7 +306,7 @@ if ! "${skip_data_prep}"; then
     if [ ${stage} -le 1 ] && [ ${stop_stage} -ge 1 ]; then
         log "Stage 1: Data preparation for data/${train_set}, data/${valid_set}, etc."
         # [Task dependent] Need to create data.sh for new corpus
-        local/data.sh ${local_data_opts} --fs "${fs}"
+        local/data.sh ${local_data_opts} --fs "${fs}" --g2p "${g2p}"
     fi
     
 
@@ -351,9 +351,6 @@ if ! "${skip_data_prep}"; then
                 scripts/audio/format_wav_scp.sh --nj "${nj}" --cmd "${train_cmd}" \
                     --audio-format "${audio_format}" --fs "${fs}" ${_opts} \
                     "data/${dset}/wav.scp" "${data_feats}${_suf}/${dset}"
-                scripts/audio/format_xml_scp.sh --nj "${nj}" --cmd "${train_cmd}" \
-                    ${_opts} \
-                    "xml_dump" "${data_feats}${_suf}/${dset}"
                 echo "${feats_type}" > "${data_feats}${_suf}/${dset}/feats_type"
             done
         fi
@@ -471,6 +468,17 @@ if ! "${skip_data_prep}"; then
               --add_symbol "${blank}:0" \
               --add_symbol "${oov}:1" \
               --add_symbol "${sos_eos}:-1"
+        
+        #${python} -m espnet2.bin.tokenize_text \
+        #      --token_type "word" -f 2- \
+        #      --input "${data_feats}/srctexts" --output "${token_list}_syb" \
+        #      --non_linguistic_symbols "${nlsyms_txt}" \
+        #      --cleaner "${cleaner}" \
+        #      --g2p "${g2p}" \
+        #      --write_vocabulary true \
+        #      --add_symbol "${blank}:0" \
+        #      --add_symbol "${oov}:1" \
+        #      --add_symbol "${sos_eos}:-1"
     fi
 else
     log "Skip the stages for data preparation"
@@ -509,7 +517,6 @@ if ! "${skip_train}"; then
             _opts+="--feats_extract_conf n_mels=${n_mels} "
         fi
 
-
         # Add extra configs for additional inputs
         # NOTE(kan-bayashi): We always pass this options but not used in default
         _opts+="--score_feats_extract ${score_feats_extract} "
@@ -517,6 +524,7 @@ if ! "${skip_train}"; then
         _opts+="--score_feats_extract_conf n_fft=${n_fft} "
         _opts+="--score_feats_extract_conf win_length=${win_length} "
         _opts+="--score_feats_extract_conf hop_length=${n_shift} "
+        _opts+="--pitch_extract ${pitch_extract} "
         _opts+="--pitch_extract_conf fs=${fs} "
         _opts+="--pitch_extract_conf n_fft=${n_fft} "
         _opts+="--pitch_extract_conf hop_length=${n_shift} "
@@ -597,11 +605,11 @@ if ! "${skip_train}"; then
                 --energy_normalize none \
                 --train_data_path_and_name_and_type "${_train_dir}/text,text,text" \
                 --train_data_path_and_name_and_type "${_train_dir}/label,label,duration" \
-                --train_data_path_and_name_and_type "${_train_dir}/xml.scp,midi,midi" \
+                --train_data_path_and_name_and_type "${_train_dir}/score,score,midi" \
[…]
