### 2022-11-23 aa0f1a0
fix chunk mask
