### 2022-11-23 9bd640a
remove right context + minor fixes
