### 2022-11-23 028b2d5
monkey patch chunk-by-chunk decoding before rework
