+++ espnet2/bin/asr_transducer_inference.py
@@ -6,7 +6,6 @@ from __future__ import annotations
 
 import argparse
 import logging
-import math
 import sys
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Sequence, Tuple, Union
@@ -55,7 +54,6 @@ class Speech2Text:
         streaming: Whether to perform chunk-by-chunk inference.
         chunk_size: Number of frames in chunk AFTER subsampling.
         left_context: Number of frames in left context AFTER subsampling.
-        right_context: Number of frames in right context AFTER subsampling.
         display_partial_hypotheses: Whether to display partial hypotheses.
 
     """
@@ -80,7 +78,6 @@ class Speech2Text:
         streaming: bool = False,
         chunk_size: int = 16,
         left_context: int = 32,
-        right_context: int = 0,
         display_partial_hypotheses: bool = False,
     ) -> None:
         """Construct a Speech2Text object."""
@@ -171,30 +168,33 @@ class Speech2Text:
         self.streaming = streaming
         self.chunk_size = max(chunk_size, 0)
         self.left_context = max(left_context, 0)
-        self.right_context = max(right_context, 0)
 
-        if not streaming or chunk_size == 0:
-            self.streaming = False
-            self.asr_model.encoder.dynamic_chunk_training = False
+        self.streaming = streaming and chunk_size > 0
+        self.asr_model.encoder.dynamic_chunk_training = False
 
         self.n_fft = asr_train_args.frontend_conf.get("n_fft", 512)
         self.hop_length = asr_train_args.frontend_conf.get("hop_length", 128)
+        self.win_length = asr_train_args.frontend_conf.get("win_length", self.n_fft)
 
-        if asr_train_args.frontend_conf.get("win_length", None) is not None:
-            self.frontend_window_size = asr_train_args.frontend_conf["win_length"]
-        else:
-            self.frontend_window_size = self.n_fft
-
-        self.window_size = self.chunk_size + self.right_context
-        self._raw_ctx = self.asr_model.encoder.get_encoder_input_raw_size(
-            self.window_size, self.hop_length
-        )
-
-        self.last_chunk_length = (
-            self.asr_model.encoder.embed.min_frame_length + self.right_context + 1
-        ) * self.hop_length
+        self._raw_ctx = self.compute_raw_size(self.chunk_size + 1)
+        self._chunk_raw_ctx = self.compute_raw_size(self.chunk_size)
+        self._one_chunk_raw_ctx = self.compute_raw_size(1)
 
         self.reset_inference_cache()
+
+    def compute_raw_size(self, size: int):
+        """Return the corresponding number of samples for a given size (in frames).
+
+        Args:
+            size: Number of frames AFTER applying subsampling.
+
+        Returns:
+            : Number of raw samples
+
+        """
+        size_bs = self.asr_model.encoder.embed.get_size_before_subsampling(size)
+
+        return (size_bs - 1) * self.hop_length
 
     def reset_inference_cache(self) -> None:
         """Reset Speech2Text parameters."""
@@ -221,100 +221,41 @@ class Speech2Text:
             feats_lengths: Features sequence length. (1, T_in, F)
 
         """
-        if self.frontend_cache is not None:
-            speech = torch.cat([self.frontend_cache["waveform_buffer"], speech], dim=0)
+        if self.streaming:
+            speech_cache = speech[0 : self._chunk_raw_ctx][-self._one_chunk_raw_ctx :]
 
-        if is_final:
-            if self.streaming and speech.size(0) < self.last_chunk_length:
-                pad = torch.zeros(
-                    self.last_chunk_length - speech.size(0), dtype=speech.dtype
+            if self.frontend_cache is not None:
[…]
