+++ espnet2/svs/xiaoice/XiaoiceSing.py
@@ -1,0 +1,680 @@
+# Copyright 2020 Nagoya University (Tomoki Hayashi)
+# Copyright 2021 Renmin University of China (Shuai Guo)
+#  Apache 2.0  (http://www.apache.org/licenses/LICENSE-2.0)
+
+"""XiaoiceSing related modules."""
+
+import logging
+import random
+from typing import Dict, Optional, Sequence, Tuple
+
+import torch
+import torch.nn.functional as F
+from torch.distributions import Beta
+from typeguard import check_argument_types
+
+from espnet2.svs.abs_svs import AbsSVS
+from espnet2.torch_utils.device_funcs import force_gatherable
+from espnet2.torch_utils.initialize import initialize
+from espnet2.tts.gst.style_encoder import StyleEncoder
+from espnet.nets.pytorch_backend.conformer.encoder import (  # noqa: H301
+    Encoder as ConformerEncoder,
+)
+from espnet.nets.pytorch_backend.e2e_tts_fastspeech import (
+    FeedForwardTransformerLoss as XiaoiceSingLoss,  # NOQA
+)
+from espnet.nets.pytorch_backend.fastspeech.duration_predictor import DurationPredictor
+from espnet.nets.pytorch_backend.fastspeech.length_regulator import LengthRegulator
+from espnet.nets.pytorch_backend.nets_utils import (
+    make_non_pad_mask,
+    make_pad_mask,
+    pad_list,
+)
+from espnet.nets.pytorch_backend.tacotron2.decoder import Postnet
+from espnet.nets.pytorch_backend.transformer.embedding import (
+    PositionalEncoding,
+    ScaledPositionalEncoding,
+)
+from espnet.nets.pytorch_backend.transformer.encoder import (  # noqa: H301
+    Encoder as TransformerEncoder,
+)
+
+
+class XiaoiceSing(AbsSVS):
+    """XiaoiceSing module for Singing Voice Synthesis.
+    This is a module of XiaoiceSing. A high-quality singing voice synthesis system which
+    employs an integrated network for spectrum, F0 and duration modeling. It follows the
+    main architecture of FastSpeech while proposing some singing-specific design:
+        1) Add features from musical score (e.g.note pitch and length)
+        2) Add a residual connection in F0 prediction to attenuate off-key issues
+        3) The duration of all the phonemes in a musical note is accumulated to calculate
+        the syllable duration loss for rhythm enhancement (syllable loss)
+    .. _`XiaoiceSing: A High-Quality and Integrated Singing Voice Synthesis System`:
+        https://arxiv.org/pdf/2006.06261.pdf
+    """
+
+    def __init__(
+        self,
+        # network structure related
+        idim: int,
+        midi_dim: int,
+        tempo_dim: int,
+        odim: int,
+        embed_dim: int,
+        adim: int = 384,
+        aheads: int = 4,
+        elayers: int = 6,
+        eunits: int = 1536,
+        dlayers: int = 6,
+        dunits: int = 1536,
+        postnet_layers: int = 5,
+        postnet_chans: int = 512,
+        postnet_filts: int = 5,
+        postnet_dropout_rate: float = 0.5,
+        positionwise_layer_type: str = "conv1d",
+        positionwise_conv_kernel_size: int = 1,
+        use_scaled_pos_enc: bool = True,
+        use_batch_norm: bool = True,
+        encoder_normalize_before: bool = True,
+        decoder_normalize_before: bool = True,
+        encoder_concat_after: bool = False,
+        decoder_concat_after: bool = False,
+        duration_predictor_layers: int = 2,
+        duration_predictor_chans: int = 384,
+        duration_predictor_kernel_size: int = 3,
[…]
