+++ espnet2/tasks/svs.py
@@ -20,6 +20,12 @@ from espnet2.svs.feats_extract.score_feats_extract import (
 )
 from espnet2.svs.naive_rnn.naive_rnn import NaiveRNN
 from espnet2.svs.naive_rnn.naive_rnn_dp import NaiveRNNDP
+
+# TODO(Yuning): Models to be added
+# from espnet2.svs.encoder_decoder.transformer.transformer import Transformer
+# from espnet2.svs.mlp_singer.mlp_singer import MLPSinger
+# from espnet2.svs.glu_transformer.glu_transformer import GLU_Transformer
+from espnet2.svs.xiaoice.XiaoiceSing import XiaoiceSing
 from espnet2.tasks.abs_task import AbsTask
 from espnet2.train.class_choices import ClassChoices
 from espnet2.train.collate_fn import CommonCollateFn
@@ -31,11 +37,6 @@ from espnet2.tts.feats_extract.energy import Energy
 from espnet2.tts.feats_extract.log_mel_fbank import LogMelFbank
 from espnet2.tts.feats_extract.log_spectrogram import LogSpectrogram
 
-# TODO(Yuning): Models to be added
-# from espnet2.svs.encoder_decoder.transformer.transformer import Transformer
-# from espnet2.svs.mlp_singer.mlp_singer import MLPSinger
-# from espnet2.svs.glu_transformer.glu_transformer import GLU_Transformer
-# from espnet2.svs.xiaoice.XiaoiceSing import XiaoiceSing
 # from espnet2.svs.xiaoice.XiaoiceSing import XiaoiceSing_noDP
 # from espnet2.svs.bytesing.bytesing import ByteSing
 from espnet2.tts.utils import ParallelWaveGANPretrainedVocoder
@@ -105,9 +106,9 @@ svs_choices = ClassChoices(
         # bytesing=ByteSing,
         naive_rnn=NaiveRNN,
         naive_rnn_dp=NaiveRNNDP,
-        # mlp=MLPSinger,
-        # xiaoice=XiaoiceSing,
+        xiaoice=XiaoiceSing,
         # xiaoice_noDP=XiaoiceSing_noDP,
+        # mlp=MLPSinger,
     ),
     type_check=AbsSVS,
     default="transformer",
