+++ espnet2/svs/naive_rnn/naive_rnn.py
@@ -291,22 +291,27 @@ class NaiveRNN(AbsSVS):
         feats: torch.Tensor,
         feats_lengths: torch.Tensor,
         ds: torch.Tensor,
+        ds_syb: Optional[torch.Tensor] = None,
         label_lab: Optional[torch.Tensor] = None,
         label_lab_lengths: Optional[torch.Tensor] = None,
-        label_xml: Optional[torch.Tensor] = None,
-        label_xml_lengths: Optional[torch.Tensor] = None,
+        label_score: Optional[torch.Tensor] = None,
+        label_score_lengths: Optional[torch.Tensor] = None,
         midi_lab: Optional[torch.Tensor] = None,
         midi_lab_lengths: Optional[torch.Tensor] = None,
-        midi_xml: Optional[torch.Tensor] = None,
-        midi_xml_lengths: Optional[torch.Tensor] = None,
+        midi_score: Optional[torch.Tensor] = None,
+        midi_score_lengths: Optional[torch.Tensor] = None,
         tempo_lab: Optional[torch.Tensor] = None,
         tempo_lab_lengths: Optional[torch.Tensor] = None,
-        tempo_xml: Optional[torch.Tensor] = None,
-        tempo_xml_lengths: Optional[torch.Tensor] = None,
+        tempo_score: Optional[torch.Tensor] = None,
+        tempo_score_lengths: Optional[torch.Tensor] = None,
         beat_lab: Optional[torch.Tensor] = None,
         beat_lab_lengths: Optional[torch.Tensor] = None,
-        beat_xml: Optional[torch.Tensor] = None,
-        beat_xml_lengths: Optional[torch.Tensor] = None,
+        beat_phn_score: Optional[torch.Tensor] = None,
+        beat_phn_score_lengths: Optional[torch.Tensor] = None,
+        beat_syb_score: Optional[torch.Tensor] = None,
+        beat_syb_score_lengths: Optional[torch.Tensor] = None,
+        pitch: Optional[torch.Tensor] = None,
+        pitch_lengths: Optional[torch.Tensor] = None,
         spembs: Optional[torch.Tensor] = None,
         sids: Optional[torch.Tensor] = None,
         lids: Optional[torch.Tensor] = None,
@@ -439,16 +444,19 @@ class NaiveRNN(AbsSVS):
     def inference(
         self,
         text: torch.Tensor,
-        ds: torch.Tensor = None,
-        feats: torch.Tensor = None,
+        ds: Optional[torch.Tensor] = None,
+        ds_syb: Optional[torch.Tensor] = None,
+        feats: Optional[torch.Tensor] = None,
         label_lab: Optional[torch.Tensor] = None,
-        label_xml: Optional[torch.Tensor] = None,
+        label_score: Optional[torch.Tensor] = None,
         midi_lab: Optional[torch.Tensor] = None,
-        midi_xml: Optional[torch.Tensor] = None,
+        midi_score: Optional[torch.Tensor] = None,
         tempo_lab: Optional[torch.Tensor] = None,
-        tempo_xml: Optional[torch.Tensor] = None,
+        tempo_score: Optional[torch.Tensor] = None,
         beat_lab: Optional[torch.Tensor] = None,
-        beat_xml: Optional[torch.Tensor] = None,
+        beat_phn_score: Optional[torch.Tensor] = None,
+        beat_syb_score: Optional[torch.Tensor] = None,
+        pitch: Optional[torch.Tensor] = None,
         spembs: Optional[torch.Tensor] = None,
         sids: Optional[torch.Tensor] = None,
         lids: Optional[torch.Tensor] = None,
