### 2022-11-27 66fe02a
temporarily remove autoencoder loss and lattice loss
