### 2022-11-28 1e505a6
time sync decoding for asr
