+++ espnet/nets/time_sync_beam_search.py
@@ -24,12 +24,15 @@ import logging
 from espnet.nets.beam_search import Hypothesis
 from espnet.nets.scorer_interface import ScorerInterface
 
+
 @dataclass
 class CacheItem:
     """For caching attentional decoder and LM states."""
+
     state: Any
     scores: Any
     log_sum: float
+
 
 class TimeSyncBeamSearch(torch.nn.Module):
     """Time synchronous beam search algorithm"""
@@ -39,16 +42,16 @@ class TimeSyncBeamSearch(torch.nn.Module):
         sos: int,
         beam_size: int,
         ctc: torch.nn.Module,
-        token_list = dict,
-        pre_beam_ratio: float=1.5,
-        decoder: ScorerInterface=None,
-        lm: ScorerInterface=None,
-        ctc_weight: float=1.0,
-        lm_weight: float=0.0,
-        penalty: float=1.0,
-        blank: int=0,
-        force_lid: bool=False,
-        temp: float=1.0,
+        token_list=dict,
+        pre_beam_ratio: float = 1.5,
+        decoder: ScorerInterface = None,
+        lm: ScorerInterface = None,
+        ctc_weight: float = 1.0,
+        lm_weight: float = 0.0,
+        penalty: float = 1.0,
+        blank: int = 0,
+        force_lid: bool = False,
+        temp: float = 1.0,
     ):
         """
         beam_size: num hyps
@@ -68,14 +71,14 @@ class TimeSyncBeamSearch(torch.nn.Module):
         self.pre_beam_size = int(pre_beam_ratio * beam_size)
         self.ctc_weight = ctc_weight
         self.lm_weight = lm_weight
-        self.decoder_weight = 1.0 - ctc_weight  
+        self.decoder_weight = 1.0 - ctc_weight
         self.penalty = penalty
         self.sos = sos
         self.sos_th = torch.tensor([self.sos])
         self.blank = blank
-        self.attn_cache = dict()     # cache for p_attn(Y|X)
-        self.lm_cache = dict()     # cache for p_lm(Y)
-        self.enc_output = None           # log p_ctc(Z|X)
+        self.attn_cache = dict()  # cache for p_attn(Y|X)
+        self.lm_cache = dict()  # cache for p_lm(Y)
+        self.enc_output = None  # log p_ctc(Z|X)
         self.force_lid = force_lid
         self.temp = temp
         self.token_list = token_list
@@ -87,66 +90,72 @@ class TimeSyncBeamSearch(torch.nn.Module):
 
         if self.decoder is not None:
             init_decoder_state = self.decoder.init_state(enc_output)
-            decoder_scores, decoder_state = self.decoder.score(self.sos_th, init_decoder_state, enc_output)
+            decoder_scores, decoder_state = self.decoder.score(
+                self.sos_th, init_decoder_state, enc_output
+            )
             self.attn_cache[(self.sos,)] = CacheItem(
-                state = decoder_state,
-                scores = decoder_scores,
-                log_sum = 0.0,
+                state=decoder_state,
+                scores=decoder_scores,
+                log_sum=0.0,
             )
         if self.lm is not None:
             init_lm_state = self.lm.init_state(enc_output)
             lm_scores, lm_state = self.lm.score(self.sos_th, init_lm_state, enc_output)
             self.lm_cache[(self.sos,)] = CacheItem(
-                state = lm_state,
-                scores = lm_scores,
-                log_sum = 0.0,
+                state=lm_state,
+                scores=lm_scores,
+                log_sum=0.0,
             )
 
-    def cached_score(
[…]
