+++ espnet2/tasks/asr.py
@@ -21,6 +21,7 @@ from espnet2.asr.decoder.transformer_decoder import (
     LightweightConvolutionTransformerDecoder,
     TransformerDecoder,
 )
+from espnet2.asr.decoder.whisper_decoder import OpenAIWhisperDecoder
 from espnet2.asr.encoder.abs_encoder import AbsEncoder
 from espnet2.asr.encoder.branchformer_encoder import BranchformerEncoder
 from espnet2.asr.encoder.conformer_encoder import ConformerEncoder
@@ -43,6 +44,7 @@ from espnet2.asr.encoder.transformer_encoder_multispkr import (
 )
 from espnet2.asr.encoder.vgg_rnn_encoder import VGGRNNEncoder
 from espnet2.asr.encoder.wav2vec2_encoder import FairSeqWav2Vec2Encoder
+from espnet2.asr.encoder.whisper_encoder import OpenAIWhisperEncoder
 from espnet2.asr.espnet_model import ESPnetASRModel
 from espnet2.asr.frontend.abs_frontend import AbsFrontend
 from espnet2.asr.frontend.default import DefaultFrontend
@@ -145,6 +147,7 @@ encoder_choices = ClassChoices(
         hubert_pretrain=FairseqHubertPretrainEncoder,
         longformer=LongformerEncoder,
         branchformer=BranchformerEncoder,
+        whisper=OpenAIWhisperEncoder,
         e_branchformer=EBranchformerEncoder,
     ),
     type_check=AbsEncoder,
@@ -170,6 +173,7 @@ decoder_choices = ClassChoices(
         rnn=RNNDecoder,
         transducer=TransducerDecoder,
         mlm=MLMDecoder,
+        whisper=OpenAIWhisperDecoder,
         hugging_face_transformers=HuggingFaceTransformersDecoder,
     ),
     type_check=AbsDecoder,
@@ -276,7 +280,10 @@ class ASRTask(AbsTask):
             "--token_type",
             type=str,
             default="bpe",
-            choices=["bpe", "char", "word", "phn", "hugging_face"],
+            choices=[
+                "bpe", "char", "word", "phn", "hugging_face",
+                "whisper_en", "whisper_multilingual"
+            ],
             help="The text will be tokenized " "in the specified level token",
         )
         group.add_argument(
@@ -293,7 +300,8 @@ class ASRTask(AbsTask):
         group.add_argument(
             "--cleaner",
             type=str_or_none,
-            choices=[None, "tacotron", "jaconv", "vietnamese"],
+            choices=[None, "tacotron", "jaconv", "vietnamese", \
+                     "whisper_en", "whisper_basic"],
             default=None,
             help="Apply text cleaning",
         )
@@ -438,6 +446,12 @@ class ASRTask(AbsTask):
     @classmethod
     def build_model(cls, args: argparse.Namespace) -> ESPnetASRModel:
         assert check_argument_types()
+        if "whisper" in args.token_list:
+            use_whisper_vocab = True
+            whisper_type = "en" if "whisper_en" in args.token_list \
+                           else "multilingual"
+        else:
+            use_whisper_vocab = False
         if isinstance(args.token_list, str):
             with open(args.token_list, encoding="utf-8") as f:
                 token_list = [line.rstrip() for line in f]
@@ -448,7 +462,14 @@ class ASRTask(AbsTask):
             token_list = list(args.token_list)
         else:
             raise RuntimeError("token_list must be str or list")
-        vocab_size = len(token_list)
+
+        if use_whisper_vocab:
+            vocab_size = 51865 if whisper_type == "multilingual" \
+                               else 51864
+            token_list.extend(['()' for _ in range(vocab_size - len(token_list))])
[…]
