+++ espnet2/bin/tokenize_text.py
@@ -220,7 +220,11 @@ def get_parser() -> argparse.ArgumentParser:
     parser.add_argument(
         "--cleaner",
         type=str_or_none,
-        choices=[None, "tacotron", "jaconv", "vietnamese", "korean_cleaner"],
+        choices=[
+            None, "tacotron", "jaconv", 
+            "vietnamese", "korean_cleaner",
+            "whisper_en", "whisper_basic",
+        ],
         default=None,
         help="Apply text cleaning",
     )

[…]
