### 2022-11-28 #4746 @faroit
> However, we haven't finished reducing the word count, so can you help us temporarily suspend the review process? I apologize for any inconvenience caused to you!

@neillu23 we could do that but I would like to highly encourage you to not suspend the review for this reason. I think reducing the word count can be done with small amount of effort. Let us know you decided soon.
