### 2022-11-29 f10f911
Changes requested by kan_bayashi for PR into espnet/main.
