+++ egs2/TEMPLATE/tts1/tts.sh
@@ -47,7 +47,7 @@ local_data_opts="" # Options to be passed to local/data.sh.
 
 # Feature extraction related
 feats_type=raw             # Input feature type.
-audio_format=wav          # Audio format: wav, flac, wav.ark, flac.ark  (only in feats_type=raw).
+audio_format=flac          # Audio format: wav, flac, wav.ark, flac.ark  (only in feats_type=raw).
 min_wav_duration=0.1       # Minimum duration in second.
 max_wav_duration=20        # Maximum duration in second.
 use_sid=false              # Whether to use speaker id as the inputs (Need utt2spk in data directory).
@@ -306,7 +306,6 @@ if ! "${skip_data_prep}"; then
     if [ ${stage} -le 1 ] && [ ${stop_stage} -ge 1 ]; then
         log "Stage 1: Data preparation for data/${train_set}, data/${valid_set}, etc."
         # [Task dependent] Need to create data.sh for new corpus
-        local_data_opts+="--token_type ${token_type} --g2p ${g2p}"
         local/data.sh ${local_data_opts}
     fi
 
@@ -970,7 +969,6 @@ if ! "${skip_eval}"; then
                 # Use groundtruth of durations
                 _teacher_dir="${teacher_dumpdir}/${dset}"
                 _ex_opts+="--data_path_and_name_and_type ${_teacher_dir}/durations,durations,text_int "
-                _ex_opts+="--use_teacher_forcing true "
                 # Overwrite speech arguments if use knowledge distillation
                 if [ -e "${teacher_dumpdir}/${train_set}/probs" ]; then
                     _speech_data="${_teacher_dir}/denorm"
