### 2022-11-29 f80c5db
Changes requested by kan_bayashi for PR into espnet/main.
(cherry picked from commit f10f911ef0f841bc3c0ffd8bf6258efebdf4411a)
