+++ egs2/TEMPLATE/asr1/scripts/utils/mfa_format.py
@@ -1,0 +1,215 @@
+#!/usr/bin/env python3
+
+"""Convert files to/from MFA format for use in ESPnet TTS.
+
+If you wish to add functions to create .lab files for MFA, add them like this:
+
+def make_labs_[dataset]:
+     ...
+"""
+
+import argparse
+import os
+import sys
+import traceback
+from pathlib import Path
+from tacotron_cleaner.cleaners import *
+from praatio import textgrid
+
+ROOT_DIR = os.getcwd()
+
+TEXTGRID_DIR = f'{ROOT_DIR}/textgrids'
+TRAIN_TEXT_PATH = f'{ROOT_DIR}/data/train/text'
+DURATIONS_PATH = f'{ROOT_DIR}/data/train/durations'
+
+punctuation = "!',.?" + '"'
+
+fs = 22050
+hop_length = 256
+
+
+def get_parser():
+    parser = argparse.ArgumentParser(
+        description='Utilities to format from MFA to ESPnet.\n'
+                    'Usage: python scripts/utils/mfa_format.py ACTION [dataset] [options]\n'
+                    'python scripts/utils/mfa_format.py labs ljspeech\n'
+                    'python scripts/utils/mfa_format.py validate\n'
+                    'python scripts/utils/mfa_format.py durations\n'
+    )
+
+    parser.add_argument('type', help='Must be "labs, "validate" or "durations')
+    parser.add_argument('--dataset', help='Currently only supports "ljspeech"')
+    parser.add_argument('--wavs_dir', help='Path to wavs dir, lab files will be added here')
+    parser.add_argument('--textgrid_dir', default=TEXTGRID_DIR, help='Path to output MFA .TextGrid files')
+    parser.add_argument('--train_text_path', default=TRAIN_TEXT_PATH,
+                        help='Path to output list of utterances to phonemes')
+    parser.add_argument('--durations_path', default=DURATIONS_PATH,
+                        help='Path to output durations file')
+
+    return parser
+
+
+def make_labs_ljspeech(args):
+    wavs_dir = Path(args.wavs_dir)
+    with open(wavs_dir / '../metadata.csv') as f:
+        for line in f:
+            items = line.rstrip().split('|')
+            filename = items[0]
+            text = items[-1]
+            with open(wavs_dir / f'{filename}.lab', 'w') as lab_f:
+                # same as custom_english_cleaners but without uppercasing.
+                text = convert_to_ascii(text)
+                text = lowercase(text)
+                text = expand_numbers(text)
+                text = expand_abbreviations(text)
+                text = expand_symbols(text)
+                text = remove_unnecessary_symbols(text)
+                text = collapse_whitespace(text)
+                # Convert single quotes into double quotes so that MFA doesn't confuse them with clitics
+                # Find ' not preceded by a letter to the last ' not followed by a letter
+                text = re.sub(r"(\W|^)'(\w[\w .,!?']*)'(\W|$)", r'\1"\2"\3', text)
+                lab_f.write(text)
+                print(filename)
+    print('Finished writing .lab files')
+
+
+def get_phoneme_durations(tg: textgrid.Textgrid,
+                          original_text: str):
+    orig_text = original_text.replace(' ', '').rstrip()
+    text_pos = 0
+
+    words = tg.tierDict['words'].entryList
+    phones = tg.tierDict['phones'].entryList
+
+    prev_end = 0.0
+    time2punc = {}
+    phrase_words = []
+    word = None
+    for word in words:
+        start = word.start
+        if start == prev_end:
+            phrase_words.append(word.label)
+        else:
+            # find punctuation at end of previous phrase
+            phrase = ''.join(phrase_words)
+            for letter in phrase:
+                char = orig_text[text_pos]
+                while char != letter:
+                    text_pos += 1
[…]
