+++ egs2/mini_an4/asr1/run_multispkr.sh
@@ -5,10 +5,10 @@ set -e
 set -u
 set -o pipefail
 
-./multispkr_asr.sh \
+./asr.sh \
     --lang en \
     --train_set train_nodev \
     --valid_set train_dev \
     --test_sets "train_dev test test_seg" \
-    --ref_num 2 \
+    --num_ref 2 \
     --lm_train_text "data/train_nodev/text" "$@"
