+++ egs2/TEMPLATE/asr1/asr.sh
@@ -96,6 +96,11 @@ pretrained_model=              # Pretrained model to load
 ignore_init_mismatch=false      # Ignore initial mismatch
 feats_normalize=global_mvn # Normalizaton layer type.
 num_splits_asr=1           # Number of splitting for lm corpus.
+num_ref=1   # Number of references for training.
+            # In supervised learning based speech enhancement / separation, it is equivalent to number of speakers.
+num_inf=    # Number of inferences output by the model
+            # Note that if it is not specified, it will be the same as num_ref. Otherwise, it will be overwritten.
+            # In MixIT, number of outputs is larger than that of references.
 
 # Upload model related
 hf_repo=
@@ -218,6 +223,10 @@ Options:
     --ignore_init_mismatch=      # Ignore mismatch parameter init with pretrained model (default="${ignore_init_mismatch}").
     --feats_normalize  # Normalizaton layer type (default="${feats_normalize}").
     --num_splits_asr   # Number of splitting for lm corpus  (default="${num_splits_asr}").
+    --num_ref    # Number of references for training (default="${num_ref}").
+                 # In supervised learning based speech recognition, it is equivalent to number of speakers. 
+    --num_inf    # Number of inference audio generated by the model (default="${num_inf}")
+                 # Note that if it is not specified, it will be the same as num_ref. Otherwise, it will be overwritten.
 
     # Decoding related
     --inference_tag       # Suffix to the result dir for decoding (default="${inference_tag}").
@@ -287,14 +296,34 @@ else
     exit 2
 fi
 
-# Use the same text as ASR for bpe training if not specified.
-[ -z "${bpe_train_text}" ] && bpe_train_text="${data_feats}/${train_set}/text"
+num_inf=${num_inf:=${num_ref}}
+# Preprocessor related
+if [ ${num_ref} -eq 1 ]; then
+    # For single speaker, text file path and name are text
+    ref_text_files_str="text "
+    ref_text_names_str="text "
+else
+    # For multiple speakers, text file path and name are text_spk[1-N] and [text, text_spk2, ...]
+    #TODO(simpleoier): later to support flexibly defined text prefix
+    ref_text_files_str="text_spk1 "
+    ref_text_names_str="text "
+    for n in $(seq 2 ${num_ref}); do
+        ref_text_files_str+="text_spk${n} "
+        ref_text_names_str+="text_spk${n} "
+    done
+fi
+# shellcheck disable=SC2206
+ref_text_files=(${ref_text_files_str// / })
+# shellcheck disable=SC2206
+ref_text_names=(${ref_text_names_str// / })
+
+[ -z "${bpe_train_text}" ] && bpe_train_text="${data_feats}/${train_set}/${ref_text_files[0]}"
 # Use the same text as ASR for lm training if not specified.
-[ -z "${lm_train_text}" ] && lm_train_text="${data_feats}/${train_set}/text"
+[ -z "${lm_train_text}" ] && lm_train_text="${data_feats}/${train_set}/${ref_text_files[0]}"
 # Use the same text as ASR for lm training if not specified.
-[ -z "${lm_dev_text}" ] && lm_dev_text="${data_feats}/${valid_set}/text"
+[ -z "${lm_dev_text}" ] && lm_dev_text="${data_feats}/${valid_set}/${ref_text_files[0]}"
 # Use the text of the 1st evaldir if lm_test is not specified
-[ -z "${lm_test_text}" ] && lm_test_text="${data_feats}/${test_sets%% *}/text"
[…]
