+++ egs2/mini_an4/asr1/multispkr_asr.sh
@@ -1,1 +1,0 @@
-../../TEMPLATE/asr1/multispkr_asr.sh

[…]
