### 2022-11-30 a472dfb
remove lattice loss and auto-encoder loss and unused code in uasr.py
