### 2022-11-30 dda79c3
Remove lattice loss and auto-encoder related code. Make some fix.
