### 2022-11-30 e62ed06
Stage 1 & 2: Data downloading and prepping
