+++ egs2/asvspoof2019LA/asr1/local/score.py
@@ -1,0 +1,1 @@
+../../../fsc/asr1/local/score.py
