+++ egs2/TEMPLATE/asr1/db.sh
@@ -10,6 +10,7 @@ AISHELL4=downloads
 ALFFA=downloads
 AMERICASNLP22=downloads
 AN4=downloads
+ASVSPOOF2019LA=LA
 AUDIOSET=
 DIRHA_ENGLISH_PHDEV=
 DIRHA_WSJ=

[…]
