### 2022-12-01 babef05
Minor bug fixes
