+++ egs2/asvspoof2019LA/asr1/run.sh
@@ -11,20 +11,20 @@ test_sets="test valid"
 
 asr_config=conf/train_asr.yaml
 
-./asr.sh \
-    --lang zh \
-    --ngpu 1 \
-    --stage 1 \
-    --use_lm false \
-    --nbpe 5000 \
-    --token_type word\
-    --audio_format wav\
-    --feats_type raw\
-    --feats_normalize null \
-    --max_wav_duration 30 \
-    --inference_asr_model valid.acc.ave_5best.pth\
-    --asr_config "${asr_config}" \
-    --expdir "exp" \
-    --train_set "${train_set}" \
-    --valid_set "${valid_set}" \
-    --test_sets "${test_sets}" "$@"
+# ./asr.sh \
+#     --lang zh \
+#     --ngpu 1 \
+#     --stage 1 \
+#     --use_lm false \
+#     --nbpe 5000 \
+#     --token_type word\
+#     --audio_format wav\
+#     --feats_type raw\
+#     --feats_normalize null \
+#     --max_wav_duration 30 \
+#     --inference_asr_model valid.acc.ave_5best.pth\
+#     --asr_config "${asr_config}" \
+#     --expdir "exp" \
+#     --train_set "${train_set}" \
+#     --valid_set "${valid_set}" \
+#     --test_sets "${test_sets}" "$@"
