+++ egs2/asvspoof2019LA/asr1/asv.sh
@@ -1,0 +1,162 @@
+#!/bin/bash
+
+# Set bash to 'debug' mode, it will exit on :
+# -e 'error', -u 'undefined variable', -o ... 'error in pipeline', -x 'print commands',
+
+# Yongyi Zang, November 2022, Modified from https://github.com/espnet/espnet/compare/master...2022fall_new_task_tutorial.
+
+set -e
+set -u
+set -o pipefail
+
+log() {
+    local fname=${BASH_SOURCE[1]##*/}
+    echo -e "$(date '+%Y-%m-%dT%H:%M:%S') (${fname}:${BASH_LINENO[0]}:${FUNCNAME[1]}) $*"
+}
+min() {
+  local a b
+  a=$1
+  for b in "$@"; do
+      if [ "${b}" -le "${a}" ]; then
+          a="${b}"
+      fi
+  done
+  echo "${a}"
+}
+SECONDS=0
+
+# General configuration
+stage=1              # Processes starts from the specified stage.
+stop_stage=10000     # Processes is stopped at the specified stage.
+skip_data_prep=false # Skip data preparation stages
+skip_train=false     # Skip training stages
+skip_eval=false      # Skip decoding and evaluation stages
+ngpu=1               # The number of gpus ("0" uses cpu, otherwise use gpu).
+num_nodes=1          # The number of nodes
+nj=32                # The number of parallel jobs.
+dumpdir=dump         # Directory to dump features.
+inference_nj=4      # The number of parallel jobs in decoding.
+gpu_inference=false  # Whether to perform gpu decoding.
+expdir=exp           # Directory to save experiments.
+python=python3       # Specify python to execute espnet commands
+
+# Data preparation related
+# TODO: What data_opts should be used?
+local_data_opts= # The options given to local/data.sh.
+
+# Feature extraction related
+audio_format=flac    # Audio format: wav, flac, wav.ark, flac.ark.
+fs=8k                # Sampling rate.
+
+# asvspoof model related
+asvspoof_tag=    # Suffix to the result dir for asvspoof model training.
+asvspoof_config= # Config for asvspoof model training.
+asvspoof_args=   # Arguments for asvspoof model training, e.g., "--max_epoch 10".
+             # Note that it will overwrite args in asvspoof config.
+feats_normalize=global_mvn # Normalizaton layer type.
+
+# asvspoof related
+inference_config= # Config for asvspoof model inference
+inference_model=valid.acc.best.pth
+inference_tag=    # Suffix to the inference dir for asvspoof model inference
+
+
+# [Task dependent] Set the datadir name created by local/data.sh
+train_set=       # Name of training set.
+valid_set=       # Name of development set.
+test_sets=       # Names of evaluation sets. Multiple items can be specified.
+
+
+help_message=$(cat << EOF
+Usage: $0 --train-set <train_set_name> --valid-set <valid_set_name> --test_sets <test_set_names>
+Options:
+    # General configuration
+    --stage          # Processes starts from the specified stage (default="${stage}").
+    --stop_stage     # Processes is stopped at the specified stage (default="${stop_stage}").
+    --skip_data_prep # Skip data preparation stages (default="${skip_data_prep}").
+    --skip_train     # Skip training stages (default="${skip_train}").
+    --skip_eval      # Skip decoding and evaluation stages (default="${skip_eval}").
+    --ngpu           # The number of gpus ("0" uses cpu, otherwise use gpu, default="${ngpu}").
+    --num_nodes      # The number of nodes
+    --nj             # The number of parallel jobs (default="${nj}").
+    --inference_nj   # The number of parallel jobs in inference (default="${inference_nj}").
+    --gpu_inference  # Whether to use gpu for inference (default="${gpu_inference}").
[…]
