+++ espnet2/svs/naive_rnn/naive_rnn.py
@@ -290,28 +290,18 @@ class NaiveRNN(AbsSVS):
         text_lengths: torch.Tensor,
         feats: torch.Tensor,
         feats_lengths: torch.Tensor,
-        ds: torch.Tensor,
-        ds_syb: Optional[torch.Tensor] = None,
-        label_lab: Optional[torch.Tensor] = None,
-        label_lab_lengths: Optional[torch.Tensor] = None,
-        label_score: Optional[torch.Tensor] = None,
-        label_score_lengths: Optional[torch.Tensor] = None,
-        midi_lab: Optional[torch.Tensor] = None,
-        midi_lab_lengths: Optional[torch.Tensor] = None,
-        midi_score: Optional[torch.Tensor] = None,
-        midi_score_lengths: Optional[torch.Tensor] = None,
-        tempo_lab: Optional[torch.Tensor] = None,
-        tempo_lab_lengths: Optional[torch.Tensor] = None,
-        tempo_score: Optional[torch.Tensor] = None,
-        tempo_score_lengths: Optional[torch.Tensor] = None,
-        beat_lab: Optional[torch.Tensor] = None,
-        beat_lab_lengths: Optional[torch.Tensor] = None,
-        beat_phn_score: Optional[torch.Tensor] = None,
-        beat_phn_score_lengths: Optional[torch.Tensor] = None,
-        beat_syb_score: Optional[torch.Tensor] = None,
-        beat_syb_score_lengths: Optional[torch.Tensor] = None,
+        label: Optional[Dict[str, torch.Tensor]] = None,
+        label_lengths: Optional[Dict[str, torch.Tensor]] = None,
+        melody: Optional[Dict[str, torch.Tensor]] = None,
+        melody_lengths: Optional[Dict[str, torch.Tensor]] = None,
+        tempo: Optional[Dict[str, torch.Tensor]] = None,
+        tempo_lengths: Optional[Dict[str, torch.Tensor]] = None,
+        beat: Optional[Dict[str, torch.Tensor]] = None,
+        beat_lengths: Optional[Dict[str, torch.Tensor]] = None,
         pitch: Optional[torch.Tensor] = None,
         pitch_lengths: Optional[torch.Tensor] = None,
+        ds: Optional[torch.Tensor] = None,
+        ds_syb: Optional[torch.Tensor] = None,
         spembs: Optional[torch.Tensor] = None,
         sids: Optional[torch.Tensor] = None,
         lids: Optional[torch.Tensor] = None,
@@ -324,28 +314,43 @@ class NaiveRNN(AbsSVS):
             text_lengths (LongTensor): Batch of lengths of each input batch (B,).
             feats (Tensor): Batch of padded target features (B, Lmax, odim).
             feats_lengths (LongTensor): Batch of the lengths of each target (B,).
-            label: (B, Tmax)
-            label_lengths: (B,)
-            midi: (B, Tmax)
-            midi_lengths: (B,)
+            label (Optional[Dict]): key is "lab" or "score";
+                value (LongTensor): Batch of padded label ids (B, Tmax).
+            label_lengths (Optional[Dict]): key is "lab" or "score";
+                value (LongTensor): Batch of the lengths of padded label ids (B, ).
+            melody (Optional[Dict]): key is "lab" or "score";
+                value (LongTensor): Batch of padded melody (B, Tmax).
+            melody_lengths (Optional[Dict]): key is "lab" or "score";
+                value (LongTensor): Batch of the lengths of padded melody (B, ).
+            tempo (Optional[Dict]): key is "lab" or "score";
+                value (LongTensor): Batch of padded tempo (B, Tmax).
+            tempo_lengths (Optional[Dict]):  key is "lab" or "score";
+                value (LongTensor): Batch of the lengths of padded tempo (B, ).
[…]
