+++ egs2/opencpop2/svs1/local/data_prep.py
@@ -1,384 +1,0 @@
-import argparse
-import os
-
-import librosa
-import miditoolkit
-import numpy as np
-
-# from espnet2.fileio.xml_scp import XMLScpWriter
-from espnet2.text.build_tokenizer import build_tokenizer
-
-# from espnet2.text.phoneme_tokenizer import pypinyin_g2p_phone_without_prosody
-
-
-customed_dic = {
-    "最长的相拥谁爱得太自由": [
-        "z_ui",
-        "ch_ang",
-        "d_e",
-        "x_iang",
-        "y_ong",
-        "sh_ui",
-        "ai",
-        "d_e",
-        "t_ai",
-        "z_i",
-        "y_ou",
-    ],
-    "如小雪落下海岸线": ["r_u", "x_iao", "x_ve", "l_uo", "x_ia", "h_ai", "an", "x_ian"],
-}
-
-
-def load_midi_note_scp(midi_note_scp):
-    # Note(jiatong): midi format is 0-based
-    midi_mapping = {}
-    with open(midi_note_scp, "r", encoding="utf-8") as f:
-        content = f.read().strip().split("\n")
-        for key in content:
-            key = key.split("\t")
-            midi_mapping[key[1]] = int(key[0])
-    return midi_mapping
-
-
-def load_midi(args):
-    midis = {}
-    for i in range(2001, 2101):
-        midi_path = os.path.join(
-            args.src_data, "raw_data", "midis", "{}.midi".format(i)
-        )
-        midi_obj = miditoolkit.midi.parser.MidiFile(midi_path)
-        tempos = midi_obj.tempo_changes
-        tempos.sort(key=lambda x: (x.time, x.tempo))
-        assert len(tempos) == 1
-        tempo = int(tempos[0].tempo + 0.5)
-        midis[i] = tempo
-    return midis
-
-
-"""
-def create_xml(lyrics, phns, notes, syb_dur, tempo, keep):
-    assert len(phns) == len(notes)
-    assert len(notes) == len(syb_dur)
-    assert len(syb_dur) == len(keep)
-    lyrics_seq = []
-    notes_seq = []
-    segs_seq = []
-    index = 0
-    st = 0
-    et = 0
-    while index < len(phns) and notes[index] == 0:
-        et += syb_dur[index]
-        index += 1
-    if st != et:
-        lyrics_seq.append("-1")
-        notes_seq.append(-1)
-        segs_seq.append([st, et])
-    for syllable in lyrics:
-        et = st
-        lyrics_phn = pypinyin_g2p_phone_without_prosody(syllable)
-        lyrics_seq.append(syllable)
-        notes_seq.append(notes[index])
-        for i in range(len(lyrics_phn)):
-            assert keep[index + i] == 0
-            assert notes[index + i] == notes[index]
-            assert syb_dur[index + i] == syb_dur[index]
-        et += syb_dur[index]
-        segs_seq.append([st, et])
-        segs_seq.append([st, et])
-        st = et
-        index += len(lyrics_phn)
-        while index < len(phns) and keep[index] == 1:
-            lyrics_seq.append("")
-            notes_seq.append(notes[index])
-            et += syb_dur[index]
-            segs_seq.append([st, et])
-            index += 1
-            st = et
-        while index < len(phns) and notes[index] == 0:
-            et += syb_dur[index]
-            index += 1
-        if st != et:
-            lyrics_seq.append("P")
-            notes_seq.append(-1)
-            segs_seq.append([st, et])
-    return lyrics_seq, notes_seq, segs_seq, tempo
-"""
-
-
-def create_score(uid, lyrics, phns, notes, syb_dur, keep):
[…]
