+++ egs2/ofuton_p_utagoe_db/svs1/local/data.sh
@@ -71,7 +71,7 @@ if [ ${stage} -le 3 ] && [ ${stop_stage} -ge 3 ]; then
     # We align music info at phone level if both annotation (label) and music score are used.
     for x in ${train_set} ${train_dev} ${recog_set}; do
         src_data=data/${x}
-        local/check_align.py ${src_data} --g2p ${g2p}
+        utils/check_align.py ${src_data} --g2p ${g2p}
         mv ${src_data}/score.scp.tmp ${src_data}/score.scp
         utils/fix_data_dir.sh --utt_extra_files "label score.scp" ${src_data}
     done
