+++ egs2/opencpop/svs1/local/data_prep.py
@@ -6,6 +6,8 @@ import librosa
 import miditoolkit
 import numpy as np
 
+from espnet2.fileio.score_scp import SingingScoreWriter
+
 """Generate segments according to structured annotation."""
 """Transfer music score into 'score' format."""
 
@@ -44,49 +46,51 @@ def load_midi(args):
     return midis
 
 
-def create_score(uid, phns, notes, syb_dur, keep):
+def create_score(uid, phns, midis, syb_dur, keep):
     # Transfer into 'score' format
-    assert len(phns) == len(notes)
-    assert len(notes) == len(syb_dur)
+    assert len(phns) == len(midis)
+    assert len(midis) == len(syb_dur)
     assert len(syb_dur) == len(keep)
     lyrics_seq = []
-    notes_seq = []
+    midis_seq = []
     segs_seq = []
     phns_seq = []
     st = 0
     index_phn = 0
+    note_list = []
     while index_phn < len(phns):
-        notes_seq.append(notes[index_phn])
-        segs_seq.append([st, st + syb_dur[index_phn]])
+        midi = midis[index_phn]
+        note_info = [st]
         st += syb_dur[index_phn]
         syb = [phns[index_phn]]
         index_phn += 1
         if (
             index_phn < len(phns)
             and syb_dur[index_phn] == syb_dur[index_phn - 1]
-            and notes[index_phn] == notes[index_phn - 1]
+            and midis[index_phn] == midis[index_phn - 1]
             and keep[index_phn] == 0
         ):
             syb.append(phns[index_phn])
             index_phn += 1
         syb = "_".join(syb)
-        lyrics_seq.append(syb)
-        phns_seq.append(syb)
+        note_info.extend([st, syb, midi, syb])
+        note_list.append(note_info)
+        # multi notes in one syllable
         while (
             index_phn < len(phns)
             and keep[index_phn] == 1
             and phns[index_phn] == phns[index_phn - 1]
         ):
-            lyrics_seq.append("—")
-            notes_seq.append(notes[index_phn])
-            segs_seq.append([st, st + syb_dur[index_phn]])
+            note_info = [st]
             st += syb_dur[index_phn]
-            phns_seq.append(phns[index_phn])
+            note_info.extend([st, "—", midis[index_phn], phns[index_phn]])
+            note_list.append(note_info)
             index_phn += 1
-    return lyrics_seq, notes_seq, segs_seq, phns_seq
+    return note_list
 
 
 def process_utterance(
+    writer,
     wavscp,
     text,
     utt2spk,
@@ -94,14 +98,13 @@ def process_utterance(
     audio_dir,
     wav_dumpdir,
     segment,
-    score,
     midi_mapping,
     tempos,
     tgt_sr=24000,
 ):
-    uid, lyrics, phns, notes, syb_dur, phn_dur, keep = segment.strip().split("|")
+    uid, lyrics, phns, midis, syb_dur, phn_dur, keep = segment.strip().split("|")
     phns = phns.split(" ")
-    notes = notes.split(" ")
+    midis = midis.split(" ")
     syb_dur = syb_dur.split(" ")
     phn_dur = phn_dur.split(" ")
     keep = keep.split(" ")
@@ -111,15 +114,13 @@ def process_utterance(
     tempo = tempos[id]
 
     # note to midi index
-    notes = [midi_mapping[note] if note != "rest" else 0 for note in notes]
+    midis = [midi_mapping[note] if note != "rest" else 0 for note in midis]
 
     # type convert
     phn_dur = [float(dur) for dur in phn_dur]
     syb_dur = [float(syb) for syb in syb_dur]
     keep = [int(k) for k in keep]
[…]
