+++ espnet2/tts/espnet_model_md.py
@@ -1,0 +1,510 @@
+from contextlib import contextmanager
+from distutils.version import LooseVersion
+import logging
+from typing import Dict
+from typing import List
+from typing import Optional
+from typing import Tuple
+from typing import Union
+
+import torch
+from typeguard import check_argument_types
+
+from espnet.nets.e2e_asr_common import ErrorCalculator as ASRErrorCalculator
+from espnet.nets.e2e_mt_common import ErrorCalculator as MTErrorCalculator
+from espnet.nets.pytorch_backend.nets_utils import th_accuracy
+from espnet.nets.pytorch_backend.transformer.add_sos_eos import add_sos_eos
+from espnet.nets.pytorch_backend.transformer.label_smoothing_loss import (
+    LabelSmoothingLoss,  # noqa: H301
+)
+from espnet.nets.pytorch_backend.nets_utils import pad_list
+from sklearn.metrics import f1_score
+from espnet2.asr.ctc import CTC
+from espnet2.asr.decoder.abs_decoder import AbsDecoder
+from espnet2.asr.encoder.abs_encoder import AbsEncoder
+from espnet2.asr.frontend.abs_frontend import AbsFrontend
+from espnet2.asr.postencoder.abs_postencoder import AbsPostEncoder
+from espnet2.asr.preencoder.abs_preencoder import AbsPreEncoder
+from espnet2.asr.specaug.abs_specaug import AbsSpecAug
+from espnet2.layers.abs_normalize import AbsNormalize
+from espnet2.torch_utils.device_funcs import force_gatherable
+from espnet2.train.abs_espnet_model import AbsESPnetModel
+from espnet2.layers.inversible_interface import InversibleInterface
+from espnet2.tts.abs_tts import AbsTTS
+from espnet2.tts.feats_extract.abs_feats_extract import AbsFeatsExtract
+
+if LooseVersion(torch.__version__) >= LooseVersion("1.6.0"):
+    from torch.cuda.amp import autocast
+else:
+    # Nothing to do if torch<1.6.0
+    @contextmanager
+    def autocast(enabled=True):
+        yield
+
+
+class ESPnetTTSMDModel(AbsESPnetModel):
+    """CTC-attention hybrid Encoder-Decoder model"""
+
+    def __init__(
+        self,
+        vocab_size: int,
+        token_list: Union[Tuple[str, ...], List[str]],
+        frontend: Optional[AbsFrontend],
+        asr_encoder: AbsEncoder,
+        asr_decoder: AbsDecoder,
+        ctc: CTC,
+        feats_extract: Optional[AbsFeatsExtract],
+        pitch_extract: Optional[AbsFeatsExtract],
+        energy_extract: Optional[AbsFeatsExtract],
+        normalize: Optional[AbsNormalize and InversibleInterface],
+        pitch_normalize: Optional[AbsNormalize and InversibleInterface],
+        energy_normalize: Optional[AbsNormalize and InversibleInterface],
+        tts: AbsTTS,
+        asr_weight: float = 0.5,
+        mt_weight: float = 0.0,
+        mtlalpha: float = 0.0,
+        ignore_id: int = -1,
+        lsm_weight: float = 0.0,
+        length_normalized_loss: bool = False,
+        report_cer: bool = True,
+        report_wer: bool = True,
+        report_bleu: bool = True,
+        sym_space: str = "<space>",
+        sym_blank: str = "<blank>",
+        extract_feats_in_collect_stats: bool = True,
+        speech_attn: bool = False,
+    ):
+        assert check_argument_types()
+        assert 0.0 < asr_weight < 1.0, "asr_weight should be (0.0, 1.0)"
[…]

[…]
