+++ egs2/librispeech_100/tts1/run_md.sh
@@ -1,0 +1,42 @@
+#!/usr/bin/env bash
+# Set bash to 'debug' mode, it will exit on :
+# -e 'error', -u 'undefined variable', -o ... 'error in pipeline', -x 'print commands',
+set -e
+set -u
+set -o pipefail
+
+fs=16000
+n_fft=1024
+n_shift=256
+
+
+train_set="train_clean_100"
+valid_set="dev_clean"
+test_sets="test_clean dev_clean"
+
+train_config=conf/tuning/train_transformer_md.yaml
+inference_config=conf/decode.yaml
+
+# g2p=g2p_en # Include word separator
+g2p=g2p_en_no_space # Include no word separator
+
+./tts.sh \
+    --ngpu 1 \
+    --stage 7 \
+    --inference_nj 64 \
+    --use_multidecoder true \
+    --lang en \
+    --feats_type raw \
+    --fs "${fs}" \
+    --n_fft "${n_fft}" \
+    --n_shift "${n_shift}" \
+    --token_type phn \
+    --cleaner tacotron \
+    --g2p "${g2p}" \
+    --train_config "${train_config}" \
+    --inference_config "${inference_config}" \
+    --train_set "${train_set}" \
+    --valid_set "${valid_set}" \
+    --test_sets "${test_sets}" \
+    --srctexts "data/${train_set}/text" \
+    --audio_format "flac" "$@"
