+++ egs2/librispeech_100/tts1/conf/tuning/diff.txt
@@ -1,0 +1,162 @@
+1,4c1,5
+< # This configuration reuqires 3 gpus in the case of each gpu memory = 12GB, and it takes 2~3 days.
+< # If you cannot use 3 gpus, it is better to make batch-bins small, increase accum_grad and epochs.
+< # This model uses speaker emebdding with add integration, leading fast attention learning compared
+< # to concat integration wihtout guided attention loss.
+---
+> # This configuration is for ESPnet2 to train Transformer-TTS. Compared to the
+> # original paper, this configuration additionally use the guided attention
+> # loss to accelerate the learning of the diagonal attention. It requires
+> # 4 GPUs with 32 GB memory and it will finish within 1 day to finish the
+> # training on Tesla V100.
+6,33c7,54
+< # network architecture related
+< model-module: espnet.nets.pytorch_backend.e2e_tts_transformer:Transformer
+< embed-dim: 0
+< eprenet-conv-layers: 0  # one more linear layer w/o non-linear will be added for 0-centor
+< eprenet-conv-filts: 0
+< eprenet-conv-chans: 0
+< dprenet-layers: 2  # one more linear layer w/o non-linear will be added for 0-centor
+< dprenet-units: 256
+< adim: 384
+< aheads: 4
+< elayers: 6
+< eunits: 1536
+< dlayers: 6
+< dunits: 1536
+< postnet-layers: 5
+< postnet-filts: 5
+< postnet-chans: 256
+< use-masking: True
+< bce-pos-weight: 5.0
+< use-batch-norm: True
+< use-scaled-pos-enc: True
+< encoder-normalize-before: False
+< decoder-normalize-before: False
+< encoder-concat-after: False
+< decoder-concat-after: False
+< reduction-factor: 1
+< use-speaker-embedding: true
+< spk-embed-integration-type: add
+---
+> ##########################################################
+> #                  TTS MODEL SETTING                     #
+> ##########################################################
+> tts: transformer           # model architecture
+> tts_conf:                  # keyword arguments for the selected model
+>     embed_dim: 0           # embedding dimension in encoder prenet
+>     eprenet_conv_layers: 0 # number of conv layers in encoder prenet
+>                            # if set to 0, no encoder prenet will be used
+>     eprenet_conv_filts: 0  # filter size of conv layers in encoder prenet
+>     eprenet_conv_chans: 0  # number of channels of conv layers in encoder prenet
+>     dprenet_layers: 2      # number of layers in decoder prenet
+>     dprenet_units: 128     # number of units in decoder prenet
+>     adim: 256              # attention dimension
+>     aheads: 4              # number of attention heads
+>     elayers: 6             # number of encoder layers
+>     eunits: 512           # number of encoder ff units
+>     dlayers: 6             # number of decoder layers
+>     dunits: 512           # number of decoder ff units
+>     positionwise_layer_type: conv1d  # type of position-wise layer
+>     positionwise_conv_kernel_size: 1 # kernel size of position wise conv layer
+>     postnet_layers: 5                # number of layers of postnset
+>     postnet_filts: 5                 # filter size of conv layers in postnet
+>     postnet_chans: 128               # number of channels of conv layers in postnet
+>     use_masking: True                # whether to apply masking for padded part in loss calculation
+>     bce_pos_weight: 5.0              # weight of positive sample in binary cross entropy calculation
+>     use_scaled_pos_enc: True         # whether to use scaled positional encoding
+>     encoder_normalize_before: True   # whether to perform layer normalization before the input
+>     decoder_normalize_before: True   # whether to perform layer normalization before the input
[…]
