### 2022-12-01 d0eaad0
raise exception when missing ctc task data
