### 2022-12-02 8fe2fcf
create aux task data in template
