+++ egs2/fleurs/asr1/local/data.sh
@@ -13,6 +13,8 @@ stop_stage=100
 nlsyms_txt=data/nlsyms.txt
 SECONDS=0
 lang=af_za # see https://huggingface.co/datasets/google/fleurs#dataset-structure for list of all langs
+data_tags=
+asr_data_dir=
 
  . utils/parse_options.sh || exit 1;
 
@@ -58,6 +60,12 @@ if [ ${stage} -le 1 ] && [ ${stop_stage} -ge 1 ]; then
     utils/filter_scp.pl --exclude data/${train_dev}/wav.scp data/${train_set}/wav.scp > data/${train_set}/temp_wav.scp
     utils/filter_scp.pl --exclude data/${test_set}/wav.scp data/${train_set}/temp_wav.scp > data/${train_set}/wav.scp
     utils/fix_data_dir.sh data/${train_set}
+fi
+
+if [ ${stage} -eq 3 ] && [ ${stop_stage} -ge 3 ]; then
+    log "stage3: Additional data processing - This should only be called after ASR stage 4"
+    # create file of lids for self-conditioning
+    python local/create_lids.py --data_dir ${asr_data_dir}
 fi
 
 log "Successfully finished. [elapsed=${SECONDS}s]"
