### 2022-12-02 5892317
add auxiliary data to training
