### 2022-12-02 92bb688
xvector and char training
