+++ egs2/librispeech_100/tts2/conf/tuning/sidd_conf.yaml
@@ -1,0 +1,68 @@
+# This configuration reuqires 3 gpus in the case of each gpu memory = 12GB, and it takes 2~3 days.
+# If you cannot use 3 gpus, it is better to make batch-bins small, increase accum_grad and epochs.
+# This model uses speaker emebdding with add integration, leading fast attention learning compared
+# to concat integration wihtout guided attention loss.
+
+# network architecture related
+model-module: espnet.nets.pytorch_backend.e2e_tts_transformer:Transformer
+embed-dim: 0
+eprenet-conv-layers: 0  # one more linear layer w/o non-linear will be added for 0-centor
+eprenet-conv-filts: 0
+eprenet-conv-chans: 0
+dprenet-layers: 2  # one more linear layer w/o non-linear will be added for 0-centor
+dprenet-units: 256
+adim: 384
+aheads: 4
+elayers: 6
+eunits: 1536
+dlayers: 6
+dunits: 1536
+postnet-layers: 5
+postnet-filts: 5
+postnet-chans: 256
+use-masking: True
+bce-pos-weight: 5.0
+use-batch-norm: True
+use-scaled-pos-enc: True
+encoder-normalize-before: False
+decoder-normalize-before: False
+encoder-concat-after: False
+decoder-concat-after: False
+reduction-factor: 1
+use-speaker-embedding: true
+spk-embed-integration-type: add
+
+# minibatch related
+batch-sort-key: output # shuffle or input or output
+batch-bins: 1500000   # batch-size * (max_out * dim_out + max_in * dim_in)
+                      # 5203 batches containing from 5 to 354 samples (avg 28 samples).
+
+# training related
+transformer-init: pytorch
+transformer-warmup-steps: 10000
+transformer-lr: 1.0
+initial-encoder-alpha: 1.0
+initial-decoder-alpha: 1.0
+eprenet-dropout-rate: 0.0
+dprenet-dropout-rate: 0.5
+postnet-dropout-rate: 0.5
+transformer-enc-dropout-rate: 0.1
+transformer-enc-positional-dropout-rate: 0.1
+transformer-enc-attn-dropout-rate: 0.1
+transformer-dec-dropout-rate: 0.1
+transformer-dec-positional-dropout-rate: 0.1
+transformer-dec-attn-dropout-rate: 0.1
+transformer-enc-dec-attn-dropout-rate: 0.1
+use-guided-attn-loss: false
+bce-pos-weight: 10.0
+
+# optimization related
+opt: noam
+accum-grad: 4
+grad-clip: 1.0
+weight-decay: 0.0
+patience: 0
+epochs: 100 # 100 * 5,203 / 4 = 130,075 iters
+
+# other
+save-interval-epoch: 1

[…]
