### 2022-12-02 caba0a4
fix bad if
