+++ espnet2/asr/espnet_model.py
@@ -249,7 +249,7 @@ class ESPnetASRModel(AbsESPnetModel):
                             raise Exception(
                                 "Auxillary CTC tasks were specified but no data was found"
                             )
-                elif loss_ic is None:
+                if loss_ic is None:
                     loss_ic, cer_ic = self._calc_ctc_loss(
                         intermediate_out, encoder_out_lens, text, text_lengths
                     )
