### 2022-12-02 a52159b
fix sms_wsj config
