### 2022-12-03 6f28b71
Add speech attention code
