+++ egs2/TEMPLATE/ssl1/scripts/mfa_aligments.sh
@@ -1,0 +1,191 @@
+#!/usr/bin/env bash
+
+set -e
+set -u
+set -o pipefail
+
+. ./path.sh || exit 1;
+. ./cmd.sh || exit 1;
+
+log() {
+    local fname=${BASH_SOURCE[1]##*/}
+    echo -e "$(date '+%Y-%m-%dT%H:%M:%S') (${fname}:${BASH_LINENO[0]}:${FUNCNAME[1]}) $*"
+}
+SECONDS=0
+
+# General configuration
+stage=0
+stop_stage=100
+nj=12
+corpus_dir=data/local/corpus  # TODO(fhrozen): remove maybe
+corpus_txt=data/train/text  # TODO(fhrozen): remove maybe
+workdir=data/local/mfa  
+clean_temp=false
+datasets="tr_no_dev dev eval1"  # TODO(fhrozen): use this as main
+
+# Feature extraction related
+fs=16000
+
+# Tokenization related
+lang="english_us_tacotron"
+g2p="espeak_ng_english_us_vits"
+train=true
+max_phonemes_word=7
+
+help_message=$(cat << EOF
+Usage: $0 --stage "<stage>" --stop-stage "<stop_stage>" --fs "<fs>"
+
+Options:
+    # General configuration
+    --stage                # Processes starts from the specified stage (default="${stage}").
+    --stop_stage            # Processes is stopped at the specified stage (default="${stop_stage}").
+    --nj                    # The number of parallel jobs (default="${nj}").
+    --workdir
+    --clean_temp
+    --datasets
+
+    # Feature extraction related
+    --fs                      # Sampling rate (default="${fs}").
+
+    # Tokenization related
+    --lang
+    --g2p
+    --train
+    --max_phonemes_word
+
+EOF
+)
+
+log "$0 $*"
+# Save command line args for logging (they will be lost after utils/parse_options.sh)
+run_args=$(pyscripts/utils/print_args.py $0 "$@")
+. utils/parse_options.sh
+echo $@
+if [ $# -ne 0 ]; then
+    log "${help_message}"
+    log "Error: No positional arguments are required."
+    exit 2
+fi
+
+mkdir -p ${corpus_dir}  # TODO(fhrozen): fix this hardcoded value
+mkdir -p ${workdir}
+tempdir=${workdir}/tmp
+
+if [ ${stage} -le 0 ] && [ ${stop_stage} -ge 0 ]; then
+    log "stage 0: Prepare Data set for MFA"
+    # TODO(fhrozen): replace this for a general data set, and use utt2spk to generate the folder structure
+    for fn in $(find ${db_root}/LJSpeech-1.1/wavs -name "*.wav");do
+        ln -sf $(readlink -e ${fn}) ${corpus_dir}
+    done
+
+    # Text cleaning and save it in lab files (independly)
+    python pyscripts/prepare_labs.py --cleaner tacotron \
+        ${corpus_txt} \
+        ${corpus_dir}
+    
+    cat ${corpus_dir}/*.lab | tr " " "\n" | sort | uniq | grep -v -e '^\s*$' > ${workdir}/list.txt.tmp
+    paste ${workdir}/list.txt.tmp ${workdir}/list.txt.tmp > ${workdir}/list2.txt.tmp
+
+    # Generate a text using espnet2-based g2p
+    python pyscripts/mfa/reformat_dict.py \
+            --g2p ${g2p} \
+            ${workdir}/list2.txt.tmp
+    
+    mv ${workdir}/modified_list2.txt.tmp ${workdir}/train_dict.txt
+    rm ${workdir}/*.tmp
+fi
+
+if [ ${stage} -le 1 ] && [ ${stop_stage} -ge 1 ]; then
+    if ${train}; then
+        log "stage 1: Training MFA"
+
+        # Complete training
[…]
