+++ egs2/TEMPLATE/ssl1/pyscripts/mfa/get_phones_alignments.py
@@ -1,0 +1,176 @@
+# Process audio frames for phonemes duration following a similar process as in:
+# https://github.com/NATSpeech/NATSpeech/blob/main/utils/audio/align.py
+
+import argparse
+import codecs
+import glob
+import json
+import os
+import tqdm
+
+from espnet2.text.phoneme_tokenizer import PhonemeTokenizer
+
+import numpy as np
+import torch
+
+# To Generate Phonemes from words:
+# from montreal_forced_aligner.g2p.generator import PyniniValidator
+# from montreal_forced_aligner.models import (
+#     G2PModel,
+#     ModelManager,
+# )
+# language = "english_us_mfa"
+# manager = ModelManager()
+# manager.download_model("g2p", language)
+# model_path = G2PModel.get_pretrained_path(language)
+# g2p = PyniniValidator(g2p_model_path=model_path, num_pronunciations=1, quiet=True)
+# g2p.word_list = "my word list".split(" ")
+# phones = g2p.generate_pronunciations()
+
+# stfs = torch.stft(
+#             values, n_fft=args.n_fft, win_length=args.n_fft, hop_length=args.n_shift,
+#             center=True, normalized=False, onesided = True, return_complex = False
+#         )
+
+
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser()
+    # dataset parameters
+    parser.add_argument("--g2p", type=str, required=True, help="G2P type.")
+    parser.add_argument("--silence_symbol", type=str, default="<sil>")
+    parser.add_argument("--samplerate", type=int, default=16000)
+    parser.add_argument("--n_shift", type=int, default=256)
+    parser.add_argument("--n_fft", type=int, default=1024)
+    parser.add_argument("--win_length", type=int, default=None)
+    parser.add_argument("--mfa_offset", type=float, default=0.02)
+    parser.add_argument("--max_phonemes_word", type=int, default=None)
+    parser.add_argument("input_dir", type=str, default="")
+    parser.add_argument("output_dir", type=str, default="")
+    args = parser.parse_args()
+
+    listfiles = sorted(glob.glob(os.path.join(args.input_dir, "*.json")))
+    phoneme_tokenizer = PhonemeTokenizer(args.g2p)
+
+    SAMPLERATE = args.samplerate
+    MFA_OFFSET = args.mfa_offset
+    HOP_SIZE = args.n_shift
+    MAX_PHONES = args.max_phonemes_word
+    
+    def getframe(timestamp, _maxtime=None, offset=True):
+        if offset:
+            timestamp += MFA_OFFSET
+        if _maxtime:
+            timestamp = min(timestamp, _maxtime) 
+        frame = timestamp * SAMPLERATE / HOP_SIZE + 0.5
+        return int(frame)
+
+    sil_tkn = args.silence_symbol   
+ 
+    txt_duration = list()
+    txt_phonemes = list()
+    txt_wrd_durs = list()
+    for fn in tqdm.tqdm(listfiles):
+        key = os.path.basename(fn).replace(".json", "")
+        with codecs.open(fn, "r", encoding="utf-8") as reader:
+            data = json.load(reader)
+        
+        max_time = data["end"]
+
+        # STFT frames for centered
+        max_frames = getframe(max_time, offset=False)
+
+        _words = [x for x in data["tiers"]["words"]["entries"]]
+        _phns = (x for x in data["tiers"]["phones"]["entries"])
+
+        previous_stop = 0
+
+        phonemes = list()
+        duration = list()
+        wrd_duration = list()
+        for wrd in _words:
+            wrd_bgn, wrd_end, wrd = wrd
+
+            if (previous_stop == 0) and (wrd_bgn > 0):
[…]

[…]
