### 2022-12-04 1d6f751
fix run command
