### 2022-12-04 18c33ab
add files for generating artifacs (mels)
