### 2022-12-05 c345c4d
Data prep and seq len fix for lf_selfattn
