+++ egs2/how2_2000h/sum1/local/data.sh
@@ -17,8 +17,9 @@ stop_stage=1
 . ./path.sh
 . ./cmd.sh
 
-url_how2_2000="https://drive.google.com/file/d/1SHg7La_hflMTIm6gaCus46sn4zYqWJvb/view?usp=sharing"
-data_how2=how2_feats
+data_url=
+data_how2=${HOW2_2kH}
+
 
 log "$0 $*"
 . utils/parse_options.sh
@@ -38,16 +39,24 @@ if [ ${stage} -le 0 ] && [ ${stop_stage} -ge 0 ]; then
     if [ -d ${data_how2} ]; then
         log "$0: HowTo directory or archive already exists in ${data_how2}. Skipping download."
     else
-        ../../../utils/download_from_google_drive.sh ${url_how2_2000} $PWD tar.gz
-        log "$0: Successfully downloaded and un-tarred how2_feats.tar.gz"
+        wget ${data_url} -o out.tar.bz2 
+        tar -xvf out.tar.bz2 -C ${data_how2}
+        log "$0: Successfully downloaded and un-tarred how2_feats"
     fi
 fi
 
 
 if [ ${stage} -le 1 ] && [ ${stop_stage} -ge 1 ]; then
     log "stage 1: Data preparation and verification"
-    mv how2_feats/data .
-    mv how2_feats/fbank .
+
+    mkdir -p data 
+    for dir in tr_2000h_sum cv05_sum dev5_test_sum; do  
+        [ -f data/${dir} ] || mv ${data_how2}/data/${dir} data/${dir}
+        [ -f data/${dir}/feats.scp ] || cat ${data_how2}/audio/fbank_pitch/${dir}.scp | awk -F ' ' -v x=$(realpath ${data_how2}) '{print $1,x"/audio/fbank_pitch/"$2}' > data/${dir}/feats.scp
+        [ -f data/${dir}/wav.scp ] || cut -d ' ' -f2 data/${dir}/segments | sort | uniq | awk -F ' ' '{print $1,"<DUMMY>"}' > data/${dir}/wav.scp
+        utils/fix_data_dir.sh data/${dir}
+    done 
+   
 fi 
 
 log "Successfully finished. [elapsed=${SECONDS}s]"
