### 2022-12-05 acf12b3
Fix bash errors
