### 2022-12-05 65b497e
Fix bash errors
