### 2022-12-30 #4746 @neillu23
@faroit sorry for the inconvenience, we have finished reducing the word count and made a PR (https://github.com/espnet/espnet/pull/4844). After the other authors check the draft again, we can merge the PR and proceed. Thank you very much for your help and I hope you have a wonderful New Year!
