### 2023-04-03 #4746 @faroit
@klukosiute @shruti-singh @vimalthilak @levimcclenny @Het-Shah @NMontanaBrown @pmeier @justusschock @yunusrf @JesseLivezey @ethanwharris @paragkulkarni11

The software under review is [espnet](https://github.com/espnet/espnet) is one if the most used frameworks in the field of audio machine learning based audio/speech processing – would any of you be willing to review this submission for JOSS? Domain knowledge is a plus but not required for this review.

The JOSS review process takes place on GitHub and focuses on the software and a short paper. We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
