### 2023-04-04 #4746 @faroit
@pmeier @NMontanaBrown @klukosiute thanks for your response and the willingness to review!

Given that we have a reviewer with audio background (thanks, @joimort for accepting) now, I would like to ask @justusschock to review this and bring in his experience with torchmetrics which might help with espnet as there is some overlap when it comes to speech separation metrics.
