### 2023-04-04 #4746 @faroit
@neillu23 before we can (finally) start the review process, would it make sense to update the version of espnet under review to [202301](https://github.com/espnet/espnet/releases/tag/v.202301)?

Also reading the paper again, i think it would be good to explicitly mention the specific PRs in the footnotes that were merged for espnet se++ so that its easier to see the code contributions.
