### 2023-04-22 #4746 @neillu23
Hi @faroit, we can upgrade to version [202301](https://github.com/espnet/espnet/releases/tag/v.202301), which includes all the features of ESPnet-SE++. However, it's worth noting that most of the updates in this new version aren't related to SE. Regarding code contributions, the ESPNet-SE++ paper provides instructions and software structure for both SSE and Joint-Task work. It might be challenging to pinpoint specific SE PRs since all [SE closed PRs](https://github.com/espnet/espnet/issues?page=3&q=label%3ASE+is%3Aclosed) are likely included. Instead, it would be easier to recognize Joint-Task PRs like the "enh_s2t joint model [#4226](https://github.com/espnet/espnet/pull/4226)" and the "LT-Spatialized and SLURP-Spatialized combined enhancement recipe [#4268](https://github.com/espnet/espnet/pull/4268)." We can update this information in the footnotes.
