### 2023-05-03 #5403 @justusschock
@faroit @neillu23 I completed my review and I recommend the paper for publication. I did not (yet) check the quality of writing box as I the layout of the paper (with the big figures on pages 3-7) doesn't look too good. I'd recommend to either change the formatting there or find another way to present the information. From the wording and language itself, the paper is fine though.
