### 2023-05-18 #5403 @joimort
Thanks for your submission, @neillu23! The software and documentation meets all the criteria of the JOSS. Also, the accompanying paper is generally well-written. It provides a comprehensive view of the technology.

However, I do have a few recommendations that could enhance readability and clarity. For instance, the section "ESPNet-SE++ Recipes for SSE and Joint-Task" commences with a screenshot of the folder structure, which can be somewhat confusing for readers. As suggested by @justusschock, it would be better to present this information in a different way. The textual content contained in the figures on pages 2-7 could be transformed into straightforward text, which would be easier to understand. For the exact structure of the folders, perhaps the authors could refer the reader to the respective README files for further clarification.

Figures throughout the manuscript are not numbered, which could potentially lead to confusion. It is unclear whether this is a JOSS requirement, so the authors might want to consult the JOSS guidelines regarding this matter. Specifically, the figures on pages 7 and 9 possess two titles, adding to the ambiguity.

Some other minor imperfections: In terms of the code examples provided on pages 11 and 12, it is unclear why the code is highlighted. A minor formatting issue was noticed in lines 154/155, where a link appears to be improperly formatted due to a line break. Lastly, there are a two points related to citations that need to be addressed. In line 198, the authors reference "Asteroid" but cite it with the arXiv version (published at Interspeech). Similarly, the conference name, Interspeech, is missing in line 217.

These recommendations are intended to further enhance the quality of this already solid manuscript. The ESPnet-SE++ is an exciting piece of software, and this paper does a good job of presenting it to the community.
