### 2023-06-01 #5403 @neillu23
Hi @faroit , at present, I'm in the process of transforming pictures into text and making a new README.md file to minimize the requirement for explaining folder paths inside the document. I expect to complete these activities and make a PR by the end of this weekend. Many thanks!
