### 2023-06-05 #5403 @neillu23
I have submitted a PR at https://github.com/espnet/espnet/pull/5212 to address the feedback from the reviewers. The modifications made are as follows:

1. Instead of including configuration figures, we have provided links to the configuration file along with brief explanations.
2. The figure illustrating the directory structure for the Joint-task has been relocated to a separate readme file "TEMPLATE/enh_asr/README.md."
3. Add captions for figures and remove redundant titles for the UML diagrams.
4. The conference details in the citations have been updated.
5. The formatting problem with the links has been resolved.

I will notify you once the PR is merged. Thank you very much!
