### 2023-06-25 #5403 @neillu23
Hello @faroit , I wanted to let you know that the PR(https://github.com/espnet/espnet/pull/5212) has been successfully merged! I apologize for the delay in completing it.
