### 2023-08-09 #5403 @faroit
hi @neillu23, sorry for my slow summer response time, I was moving to a new home. The paper is close if acceptance, can you please address the minor things mentioned in https://github.com/espnet/espnet/issues/5399 to proceed with the paper?
