### 2023-08-23 #5403 @neillu23
I have submitted a PR at  https://github.com/espnet/espnet/pull/5418 to address the comments, the change including

1. Add references to the actual model architecture papers as the paper mentions that espnet-se++ adds "several new models, loss functions, and training recipes".
2. Add DOIs for the references. This is required for a JOSS publication
3. Modify title={Onssen: an open-source ...} to title={{ONSSEN}: an open-source ...}

I will notify you when the PR is merged. Thanks again!
