### 2023-10-04 #5403 @faroit
> Hi @faroit, what's the subsequent action we should take?

@neillu23 please check the [task list above](https://github.com/openjournals/joss-reviews/issues/5403#issuecomment-1746898530). Basically you would need to tag a new release on github and notify us here and then archieve the repo on zenodo like described above.
