### 2023-10-27 #5403 @neillu23
Release: https://github.com/espnet/espnet/releases/tag/v.202310
(espnet-v.202310)
Release on Zenodo: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10048174.svg)](https://doi.org/10.5281/zenodo.10048174)

Hello @faroit, we've just released the newest version of ESPnet on Zenodo and completed the tasks on the checklist. Please let me know if there are any updates required. Thanks!
