### 2023-10-30 #5403 @neillu23
Hello @faroit , I appreciate your acceptance of the paper and the guidance you've provided! I've submitted a PR to address the issue and clean up unused references in the paper.bib. Would you be able to review it and see if it's right? Thanks again!
