### 2023-11-17 #5403 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
